% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepQuality.r
\name{accuracy.microaveraging}
\alias{accuracy.microaveraging}
\title{Multi-label classification accuracy}
\usage{
accuracy.microaveraging(
  actuals,
  preds,
  na.rm = FALSE,
  type = c("precision", "recall", "F1")
)
}
\arguments{
\item{actuals}{Numeric data (vector, array, matrix, data frame or list) of actual values.}

\item{preds}{Numeric data (vector, array, matrix, data frame or list) of prediction values.}

\item{na.rm}{A logical value indicating whether actual and prediction pairs with at least one NA value should be ignored.}

\item{type}{Denotes the type (\code{precision}, \code{recall}, \code{F1-score}) for calculating the micro-averaging accuracy.}
}
\value{
The type-specific accuracy of a multi-label classification problem.
}
\description{
Multi-label classification accuracy
}
\details{
\eqn{Precision = True Positives / (True Positives + False Positives)}. The denominator is the total predicted positives.
  \eqn{Recall = True Positives / (True Positives + False Negatives)}. The denominator is the total actual positives.
  \eqn{F1 = 2 * Precision * Recall / (Precision + Recall)}.
}
\seealso{
Other Quality: 
\code{\link{accuracy}()},
\code{\link{cross_entropy}()},
\code{\link{entropy}()},
\code{\link{gini_impurity}()},
\code{\link{huber_loss}()},
\code{\link{log_cosh_loss}()},
\code{\link{mae}()},
\code{\link{mape}()},
\code{\link{mse}()},
\code{\link{msle}()},
\code{\link{quantile_loss}()},
\code{\link{rmse}()},
\code{\link{rmsle}()},
\code{\link{vc}()}
}
\concept{Quality}
