% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepDummy.r
\name{append_rows}
\alias{append_rows}
\title{Append dummy rows}
\usage{
append_rows(dataset, columns = NULL, n = 1L, type = c("copy", "minmax"))
}
\arguments{
\item{dataset}{A data set, usually a data frame.}

\item{columns}{The names or indices of the columns to be included for creating dummy rows; if \code{NULL} (default), all columns are included.}

\item{n}{The number of repeating sample blocks or new samples.}

\item{type}{The type of creating dummy rows.
\code{copy} The mode copy repeats the entire dataset n times.
\code{minmax} The mode minmax creates n synthetic rows based upon the minimum and maximum values of each column.}
}
\value{
The dataset consisting of selected columns with dummy rows.
}
\description{
Append dummy rows
}
\seealso{
\code{\link{dummify}}.

Other Dummyfication: 
\code{\link{dummify}()},
\code{\link{effectcoding}()},
\code{\link{one_hot_decode}()},
\code{\link{one_hot_encode}()},
\code{\link{remove_columns}()},
\code{\link{resample.imbalanced}()}
}
\concept{Dummyfication}
