% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepCNN.r
\name{as.CNN.image.X}
\alias{as.CNN.image.X}
\title{Create a 4-dimensional image array}
\usage{
as.CNN.image.X(images, from_file = TRUE, width, height, channels = 3L)
}
\arguments{
\item{images}{Image data either a 3D array or a list of file names, e.g. returned by \code{list.files}.}

\item{from_file}{Boolean that indicates whether image data are stored in files (default) or already given.}

\item{width}{The width of an image, equal to the number of columns.}

\item{height}{The height of an image, equal to the number of rows.}

\item{channels}{The number of channels of an image. A color channel is a primary color (like red, green and blue), 
equal to a color valence (denotes how light effects the color sensation of an eye or in common of the brain).
Primary colors can be mixed to produce any color. 
A channel equal \code{1} indicates a grayscale image, \code{3} a color image.}
}
\value{
A 4D array with the dimensions samples (number of images), height, width and channel.
}
\description{
Create a 4-dimensional image array
}
\seealso{
\code{\link[base]{list.files}},
  \code{\link[keras]{image_load}}, \code{\link[keras]{image_to_array}}, \code{\link[reticulate]{array_reshape}}.

Other Convolutional Neural Network (CNN): 
\code{\link{as.CNN.image.Y}()}
}
\concept{Convolutional Neural Network (CNN)}
