% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepCNN.r
\name{as.CNN.image.Y}
\alias{as.CNN.image.Y}
\title{Create a one-hot vector for image labels (output)}
\usage{
as.CNN.image.Y(labels)
}
\arguments{
\item{labels}{The labels of the images}
}
\value{
A one-hot encoded vector for the image labels
}
\description{
Create a one-hot vector for image labels (output)
}
\seealso{
\code{\link{one_hot_encode}}, \code{\link{as.CNN.image.X}}

Other Convolutional Neural Network (CNN): 
\code{\link{as.CNN.image.X}()},
\code{\link{get.CNN.image.X.channels}()},
\code{\link{get.CNN.image.X.height}()},
\code{\link{get.CNN.image.X.samples}()},
\code{\link{get.CNN.image.X.width}()},
\code{\link{get.CNN.image.Y.samples}()},
\code{\link{get.CNN.image.Y.units}()}
}
\concept{Convolutional Neural Network (CNN)}
