% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepCNN.r
\name{as.CNN.temp.Y}
\alias{as.CNN.temp.Y}
\title{Outcomes (Y) data format for a temporal CNN}
\usage{
as.CNN.temp.Y(Y, timesteps = NULL, reverse = FALSE)
}
\arguments{
\item{Y}{An outcome data set, usually a vector, matrix or data frame, returned by \code{get.LSTM.XY}.}

\item{timesteps}{Number of timesteps; stands for the number of different periods within one sample (record) of the result, the resampled outcome matrix \code{Y}.}

\item{reverse}{A logical value indicating the order of the values in the resampled outcome matrix \code{Y}. The values can be in given order (forward in time) or in reverse order (backward in time).}
}
\value{
Dependent on the type of \code{Y} and timesteps. If \code{Y} is a factor, the result is a one-hot vector.
  If \code{timesteps = NULL} a 2D-array with the dimensions samples and number of output units, representing a scalar outcome;
  if \code{timesteps >= 2} a 3D-array with the dimensions samples, timesteps and number of output units, representing a sequence or multi-step outcome.
}
\description{
Outcomes (Y) data format for a temporal CNN
}
\seealso{
\code{\link{get.LSTM.XY}}, \code{\link{as.CNN.temp.X}}.

Other Convolutional Neural Network (CNN): 
\code{\link{as.CNN.image.X}()},
\code{\link{as.CNN.image.Y}()},
\code{\link{as.CNN.temp.X}()},
\code{\link{get.CNN.image.X.channels}()},
\code{\link{get.CNN.image.X.height}()},
\code{\link{get.CNN.image.X.samples}()},
\code{\link{get.CNN.image.X.width}()},
\code{\link{get.CNN.image.Y.samples}()},
\code{\link{get.CNN.image.Y.units}()},
\code{\link{get.CNN.temp.X.samples}()},
\code{\link{get.CNN.temp.X.subsequences}()},
\code{\link{get.CNN.temp.X.timesteps}()},
\code{\link{get.CNN.temp.X.units}()},
\code{\link{get.CNN.temp.Y.samples}()},
\code{\link{get.CNN.temp.Y.timesteps}()},
\code{\link{get.CNN.temp.Y.units}()}
}
\concept{Convolutional Neural Network (CNN)}
