% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepML.r
\name{as.degree}
\alias{as.degree}
\title{Radian to degree}
\usage{
as.degree(radian)
}
\arguments{
\item{radian}{Radian.}
}
\value{
The degree of \code{radian}.
}
\description{
Radian to degree
}
\examples{
  as.degree(pi) # 180
}
\seealso{
Other Machine Learning: 
\code{\link{as.radian}()},
\code{\link{cross_validation_split}()},
\code{\link{distance}()},
\code{\link{erfcinv}()},
\code{\link{erfc}()},
\code{\link{erfinv}()},
\code{\link{erf}()},
\code{\link{k_nearest_neighbors}()},
\code{\link{moving_average}()},
\code{\link{naive_forecast}()},
\code{\link{predict.kmeans}()},
\code{\link{sd.pop}()},
\code{\link{similarity}()},
\code{\link{var.pop}()}
}
\concept{Machine Learning}
