% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepRNN.r
\name{as.lag}
\alias{as.lag}
\title{Get ANN lag from ARIMA(X) lag}
\usage{
as.lag(arima_lag = 0, type = "univariate")
}
\arguments{
\item{arima_lag}{The lag or p-Value from an ARIMA(X) (p-d-q) model.}

\item{type}{The type of time series: \code{univariate} or \code{multivariate}.}
}
\value{
The lag number for a LSTM.
}
\description{
Get ANN lag from ARIMA(X) lag
}
\seealso{
Other Recurrent Neural Network (RNN), Long Short-Term Memory (LSTM): 
\code{\link{as.LSTM.X}()},
\code{\link{as.LSTM.Y}()},
\code{\link{as.LSTM.data.frame}()},
\code{\link{as.LSTM.period_outcome}()},
\code{\link{as.timesteps}()},
\code{\link{build.LSTM}()},
\code{\link{fit.LSTM}()},
\code{\link{get.LSTM.X.samples}()},
\code{\link{get.LSTM.X.timesteps}()},
\code{\link{get.LSTM.X.units}()},
\code{\link{get.LSTM.XY}()},
\code{\link{get.LSTM.Y.samples}()},
\code{\link{get.LSTM.Y.timesteps}()},
\code{\link{get.LSTM.Y.units}()},
\code{\link{get.LSTM.period_shift}()},
\code{\link{predict.LSTM}()},
\code{\link{start.LSTM.invert_differencing}()}
}
\concept{Recurrent Neural Network (RNN), Long Short-Term Memory (LSTM)}
