% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepRNN.r
\name{as.LSTM.data.frame}
\alias{as.LSTM.data.frame}
\title{Recreation of a data.frame based on a X- and Y-tensor}
\usage{
as.LSTM.data.frame(X.tensor, Y.tensor)
}
\arguments{
\item{X.tensor}{A three-dimensional array of the resampled feature matrix produced by \code{as.LSTM.X}.}

\item{Y.tensor}{A two-dimensional array of the outcome produced by \code{as.LSTM.Y}.}
}
\value{
A data.frame with outcome column(s) and a further resampled feature matrix.
  The feature matrix within this data.frame has the following form:
  x1(t1), x1(t2), x1(t3)...x2(t1), x2(t2), x2(t3)...x3(t1), x3(t2), x3(t3)...
}
\description{
Recreation of a data.frame based on a X- and Y-tensor
}
\seealso{
\code{\link{as.LSTM.X}}, \code{\link{as.LSTM.Y}}.

Other Recurrent Neural Network (RNN), Long Short-Term Memory (LSTM): 
\code{\link{as.LSTM.X}()},
\code{\link{as.LSTM.Y}()},
\code{\link{as.LSTM.period_outcome}()},
\code{\link{as.lag}()},
\code{\link{as.timesteps}()},
\code{\link{build.LSTM}()},
\code{\link{fit.LSTM}()},
\code{\link{get.LSTM.X.samples}()},
\code{\link{get.LSTM.X.timesteps}()},
\code{\link{get.LSTM.X.units}()},
\code{\link{get.LSTM.XY}()},
\code{\link{get.LSTM.Y.samples}()},
\code{\link{get.LSTM.Y.units}()},
\code{\link{get.LSTM.period_shift}()},
\code{\link{predict.LSTM}()},
\code{\link{start.LSTM.invert_differencing}()}
}
\concept{Recurrent Neural Network (RNN), Long Short-Term Memory (LSTM)}
