% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepRNN.r
\name{as.LSTM.period_outcome}
\alias{as.LSTM.period_outcome}
\title{Rebuild a data frame under consideration of time series type, timesteps and lags}
\usage{
as.LSTM.period_outcome(
  dataset,
  p,
  y,
  timesteps = 1L,
  lag = 0L,
  type = "univariate"
)
}
\arguments{
\item{dataset}{A matrix or data frame with training or test data.}

\item{p}{The column index for the period variable.}

\item{y}{The column indices for the actual outcomes.}

\item{timesteps}{The number of timesteps; stands for the number of different periods within one sample (record) of the resampled feature matrix, returned by \code{as.LSTM.X}.}

\item{lag}{The number of considered lags on feature side.}

\item{type}{The type of time series: \code{univariate} or \code{multivariate}.}
}
\value{
A data frame with period and actual outcome values that can be used for quality assurance or
  for graphical representation together with the predicted values produced by \code{predict.LSTM}.
}
\description{
Rebuild a data frame under consideration of time series type, timesteps and lags
}
\details{
Within time series analysis, \code{get.LSTM.XY} extracts X and Y of a matrix or data frame in a LSTM compatible preformat.
  Then, these values are resampled thru \code{as.LSTM.X} and \code{as.LSTM.Y}. During extraction, the period shift is calculated
  to determine the number of past periods which are used to extract X values for Y description/prediction. Because of this number
  of periods cannot be used to extract Y values, this number must be deleted from \code{dataset}.
}
\seealso{
\code{\link{get.LSTM.XY}}, \code{\link{get.period_shift}}, \code{\link{predict.ANN}}.

Other Recurrent Neural Network (RNN): 
\code{\link{as.LSTM.X}()},
\code{\link{as.LSTM.Y}()},
\code{\link{as.LSTM.data.frame}()},
\code{\link{as.lag}()},
\code{\link{as.timesteps}()},
\code{\link{build.LSTM}()},
\code{\link{fit.LSTM}()},
\code{\link{get.LSTM.X.samples}()},
\code{\link{get.LSTM.X.timesteps}()},
\code{\link{get.LSTM.X.units}()},
\code{\link{get.LSTM.XY}()},
\code{\link{get.LSTM.Y.samples}()},
\code{\link{get.LSTM.Y.timesteps}()},
\code{\link{get.LSTM.Y.units}()},
\code{\link{get.period_shift}()},
\code{\link{load_weights.ANN}()},
\code{\link{predict.ANN}()},
\code{\link{save_weights.ANN}()},
\code{\link{start.invert_differencing}()}
}
\concept{Recurrent Neural Network (RNN)}
