% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepRNN.r
\name{as.LSTM.X}
\alias{as.LSTM.X}
\title{Features (X) data format for LSTM}
\usage{
as.LSTM.X(X, timesteps = 1, forward = TRUE)
}
\arguments{
\item{X}{A feature data set, usually a matrix or data.frame, returned by \code{get.LSTM.XY}.}

\item{timesteps}{Number of timesteps; stands for the number of different periods within one sample (record) of the result, the resampled feature matrix \code{X}.}

\item{forward}{The resampled feature matrix \code{X} consists of its values forward in time/period (\code{TRUE}) or backward in time (\code{FALSE}).}
}
\value{
A three-dimensional array of the resampled feature matrix \code{X} needed within Tensorflow for recurrent neural networks, e.g. LSTM.
  1. dimension: Samples (s) = Number of records 
  2. dimension: Timesteps (t) = Number of different periods within one record
  3. dimension: Features (f) = Number of features within a sequence of a period
  Note: A 3D-array with dimensions (s x t x f) can be interpret as f (s x t)-matrices.
}
\description{
Features (X) data format for LSTM
}
\seealso{
\code{\link{get.LSTM.XY}}, \code{\link{as.LSTM.Y}}, \code{\link{as.ANN.matrix}}.

Other Recurrent Neural Network (RNN), Long Short-Term Memory (LSTM): 
\code{\link{as.LSTM.Y}()},
\code{\link{as.LSTM.data.frame}()},
\code{\link{as.LSTM.period_outcome}()},
\code{\link{as.lag}()},
\code{\link{as.timesteps}()},
\code{\link{build.LSTM}()},
\code{\link{fit.LSTM}()},
\code{\link{get.LSTM.X.samples}()},
\code{\link{get.LSTM.X.timesteps}()},
\code{\link{get.LSTM.X.units}()},
\code{\link{get.LSTM.XY}()},
\code{\link{get.LSTM.Y.samples}()},
\code{\link{get.LSTM.Y.units}()},
\code{\link{get.LSTM.period_shift}()},
\code{\link{predict.LSTM}()},
\code{\link{start.LSTM.invert_differencing}()}
}
\concept{Recurrent Neural Network (RNN), Long Short-Term Memory (LSTM)}
