% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepRNN.r
\name{as.LSTM.Y}
\alias{as.LSTM.Y}
\title{Outcomes (Y) data format}
\usage{
as.LSTM.Y(Y)
}
\arguments{
\item{Y}{An outcome data set, usually a vector, matrix or data.frame, returned by \code{get.LSTM.XY}.}
}
\value{
A two-dimensional array of the outcome \code{Y} needed within Tensorflow for recurrent neural networks, e.g. LSTM.
  1. dimension: Samples = Number of records 
  2. dimension: Outcomes = Number of output units
}
\description{
Outcomes (Y) data format
}
\seealso{
\code{\link{get.LSTM.XY}}, \code{\link{as.LSTM.X}}, \code{\link{as.ANN.matrix}}.

Other Recurrent Neural Network (RNN), Long Short-Term Memory (LSTM): 
\code{\link{as.LSTM.X}()},
\code{\link{as.LSTM.data.frame}()},
\code{\link{as.LSTM.period_outcome}()},
\code{\link{as.lag}()},
\code{\link{as.timesteps}()},
\code{\link{build.LSTM}()},
\code{\link{fit.LSTM}()},
\code{\link{get.LSTM.X.samples}()},
\code{\link{get.LSTM.X.timesteps}()},
\code{\link{get.LSTM.X.units}()},
\code{\link{get.LSTM.XY}()},
\code{\link{get.LSTM.Y.samples}()},
\code{\link{get.LSTM.Y.units}()},
\code{\link{get.LSTM.period_shift}()},
\code{\link{predict.LSTM}()},
\code{\link{start.LSTM.invert_differencing}()}
}
\concept{Recurrent Neural Network (RNN), Long Short-Term Memory (LSTM)}
