% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepMLP.r
\name{as.MLP.X}
\alias{as.MLP.X}
\title{Features data format}
\usage{
as.MLP.X(X)
}
\arguments{
\item{X}{A feature data set, usually a matrix or data.frame.}
}
\value{
A two-dimensional array of the feature matrix \code{X} needed within Tensorflow for feedforward SLP or MLP.
}
\description{
Features data format
}
\seealso{
\code{\link{as.MLP.Y}}, \code{\link{as.ANN.matrix}}.

Other Single & Multi Layer Perceptron (SLP, MLP): 
\code{\link{as.ANN.matrix}()},
\code{\link{as.MLP.Y}()},
\code{\link{as.tensor.1D}()},
\code{\link{as.tensor.2D}()},
\code{\link{as.tensor.3D}()},
\code{\link{as.tensor}()},
\code{\link{build.MLP}()},
\code{\link{fit.MLP}()},
\code{\link{get.MLP.X.units}()},
\code{\link{get.MLP.Y.units}()},
\code{\link{predict.MLP}()},
\code{\link{vector.as.ANN.matrix}()}
}
\concept{Single & Multi Layer Perceptron (SLP, MLP)}
