% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepMLP.r
\name{as.MLP.Y}
\alias{as.MLP.Y}
\title{Outcomes (Y) data format for SLP/MLP}
\usage{
as.MLP.Y(Y)
}
\arguments{
\item{Y}{An outcome data set, usually a vector, matrix or data frame.}
}
\value{
A two-dimensional array of the outcome \code{Y}. For a factor outcome, the result is a one-hot vector.
}
\description{
Outcomes (Y) data format for SLP/MLP
}
\seealso{
\code{\link{as.MLP.X}}, \code{\link{as.ANN.matrix}}, \code{\link{one_hot_encode}}.

Other Single & Multi Layer Perceptron (SLP, MLP): 
\code{\link{as.ANN.matrix}()},
\code{\link{as.MLP.X}()},
\code{\link{as.tensor.1D}()},
\code{\link{as.tensor.2D}()},
\code{\link{as.tensor.3D}()},
\code{\link{as.tensor}()},
\code{\link{build.MLP}()},
\code{\link{fit.MLP}()},
\code{\link{get.MLP.X.units}()},
\code{\link{get.MLP.Y.units}()},
\code{\link{predict.ANN}()},
\code{\link{vector.as.ANN.matrix}()},
\code{\link{vector.as.numeric}()}
}
\concept{Single & Multi Layer Perceptron (SLP, MLP)}
