% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepML.r
\name{as.radian}
\alias{as.radian}
\title{Degree to radian}
\usage{
as.radian(degree)
}
\arguments{
\item{degree}{Degree.}
}
\value{
The radian of \code{degree}.
}
\description{
Degree to radian
}
\examples{
  as.radian(180) # pi
}
\seealso{
Other Machine Learning: 
\code{\link{as.degree}()},
\code{\link{cross_validation_split}()},
\code{\link{distance}()},
\code{\link{erfcinv}()},
\code{\link{erfc}()},
\code{\link{erfinv}()},
\code{\link{erf}()},
\code{\link{k_nearest_neighbors}()},
\code{\link{moving_average}()},
\code{\link{naive_forecast}()},
\code{\link{predict.kmeans}()},
\code{\link{sd.pop}()},
\code{\link{similarity}()},
\code{\link{var.pop}()}
}
\concept{Machine Learning}
