% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepMLP.r
\name{as.tensor.1D}
\alias{as.tensor.1D}
\title{Transform data into a tensor with one rank or dimension.}
\usage{
as.tensor.1D(data, reverse = FALSE)
}
\arguments{
\item{data}{Usually a numeric vector.}

\item{reverse}{Controls the order of the values in the transformed \code{data}. By default they are used in the given order, but they can also be used in reverse order.}
}
\value{
A 1D-tensor (one-dimensional array equal to a vector).
}
\description{
Transform data into a tensor with one rank or dimension.
}
\seealso{
\code{\link{as.tensor.2D}}, \code{\link{as.tensor.3D}}.

Other Single & Multi Layer Perceptron (SLP, MLP): 
\code{\link{as.ANN.matrix}()},
\code{\link{as.MLP.X}()},
\code{\link{as.MLP.Y}()},
\code{\link{as.tensor.2D}()},
\code{\link{as.tensor.3D}()},
\code{\link{build.MLP}()},
\code{\link{fit.MLP}()},
\code{\link{get.MLP.X.units}()},
\code{\link{get.MLP.Y.units}()},
\code{\link{predict.MLP}()},
\code{\link{vector.as.ANN.matrix}()}
}
\concept{Single & Multi Layer Perceptron (SLP, MLP)}
