% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepMLP.r
\name{as.tensor.2D}
\alias{as.tensor.2D}
\title{Transform data into a tensor with two ranks or dimensions.}
\usage{
as.tensor.2D(data, reverse = FALSE)
}
\arguments{
\item{data}{A data set, usually a matrix or data frame.}

\item{reverse}{Controls the order of the values in the transformed \code{data}. By default they are used in the given order, but they can also be used in reverse order.}
}
\value{
A 2D-tensor (two-dimensional array equal to a matrix).
}
\description{
Transform data into a tensor with two ranks or dimensions.
}
\seealso{
\code{\link{as.tensor.1D}}, \code{\link{as.tensor.3D}}.

Other Single & Multi Layer Perceptron (SLP, MLP): 
\code{\link{as.ANN.matrix}()},
\code{\link{as.MLP.X}()},
\code{\link{as.MLP.Y}()},
\code{\link{as.tensor.1D}()},
\code{\link{as.tensor.3D}()},
\code{\link{build.MLP}()},
\code{\link{fit.MLP}()},
\code{\link{get.MLP.X.units}()},
\code{\link{get.MLP.Y.units}()},
\code{\link{predict.ANN}()},
\code{\link{vector.as.ANN.matrix}()}
}
\concept{Single & Multi Layer Perceptron (SLP, MLP)}
