% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepMLP.r
\name{as.tensor}
\alias{as.tensor}
\title{Create a (reshaped) tensor (n-dimensional array)}
\usage{
as.tensor(
  data,
  dim = NULL,
  byrow = FALSE,
  numeric = TRUE,
  reverse = FALSE,
  adjust = NULL
)
}
\arguments{
\item{data}{A data set, e.g. vector, array, matrix, data frame.}

\item{dim}{The new dimensions to be set on the tensor.}

\item{byrow}{The order in which elements of data should be read during rearrangement. 
\code{FALSE} (default) is equivalent to the \code{Fortran}-style ordering and means elements should be read in column-major order.
\code{TRUE} is equivalent to the \code{C}-style ordering and means elements should be read in row-major order.}

\item{numeric}{A boolean that indicates whether the elements should be coerced as numeric elements.}

\item{reverse}{Controls the order of the elements in the (reshaped) tensor. By default they are used in the given order, but they can also be used in reverse order.}

\item{adjust}{A number that is added to or subtracted from a factor level element, or even not (\code{NULL}).}
}
\value{
The (reshaped) data set.
}
\description{
Create a (reshaped) tensor (n-dimensional array)
}
\details{
The function \code{array_reshape} from reticulate package differs from the base function \code{dim}. 
  While \code{dim} will fill new dimensions in column-major (Fortran-style) ordering, \code{array_reshape} allows both row-major (C-style) ordering and column-major (Fortran-style) ordering.
}
\seealso{
\code{\link{dim}}, \code{\link[reticulate]{array_reshape}}.
}
