% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepUtils.r
\name{as_ANN_matrix}
\alias{as_ANN_matrix}
\title{Convert data into an ANN compatible matrix with only numbers}
\usage{
as_ANN_matrix(data)
}
\arguments{
\item{data}{A data set, usually a matrix or data frame.}
}
\value{
A matrix with only numbers.
}
\description{
Convert data into an ANN compatible matrix with only numbers
}
\seealso{
Other Utils: 
\code{\link{degree}()},
\code{\link{distance}()},
\code{\link{list_as_numeric}()},
\code{\link{probability}()},
\code{\link{radian}()},
\code{\link{random_seed}()},
\code{\link{re.factor}()},
\code{\link{sd_pop}()},
\code{\link{similarity}()},
\code{\link{var_pop}()},
\code{\link{vector_as_ANN_matrix}()},
\code{\link{vector_as_numeric}()}
}
\concept{Utils}
