% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepCNN.r
\name{as_images_array}
\alias{as_images_array}
\title{Convert (resized) images to 3D arrays}
\usage{
as_images_array(imagelist, FUN, ...)
}
\arguments{
\item{imagelist}{A list of (resized) images returned by either \code{images_load()} or \code{images_resize()}.}

\item{FUN}{The function to be applied for changing the representation of the images within \code{imagelist}. If no function is specified \code{image_to_array()} from keras is called.}

\item{...}{Optional arguments to \code{FUN}.}
}
\value{
A list of images represented in 3D arrays with dimensions height, width and channels.
}
\description{
Convert (resized) images to 3D arrays
}
\examples{
  For an example see \code{\link{as_images_tensor4D}}.
}
\seealso{
\code{\link[keras]{image_to_array}}.

Other Convolutional Neural Network (CNN): 
\code{\link{as.CNN.image.X}()},
\code{\link{as.CNN.image.Y}()},
\code{\link{as.CNN.temp.X}()},
\code{\link{as.CNN.temp.Y}()},
\code{\link{as_images_tensor4D}()},
\code{\link{get.CNN.image.X.channels}()},
\code{\link{get.CNN.image.X.height}()},
\code{\link{get.CNN.image.X.samples}()},
\code{\link{get.CNN.image.X.width}()},
\code{\link{get.CNN.image.Y.samples}()},
\code{\link{get.CNN.image.Y.units}()},
\code{\link{get.CNN.temp.X.samples}()},
\code{\link{get.CNN.temp.X.subsequences}()},
\code{\link{get.CNN.temp.X.timesteps}()},
\code{\link{get.CNN.temp.X.units}()},
\code{\link{get.CNN.temp.Y.samples}()},
\code{\link{get.CNN.temp.Y.timesteps}()},
\code{\link{get.CNN.temp.Y.units}()},
\code{\link{images_load}()},
\code{\link{images_resize}()}
}
\concept{Convolutional Neural Network (CNN)}
