% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepRNN.r
\name{as_lag}
\alias{as_lag}
\title{Get ANN lag from ARIMA(X) lag}
\usage{
as_lag(arima_lag = 0L, type = "univariate")
}
\arguments{
\item{arima_lag}{The lag or p-Value from an ARIMA(X) (p-d-q) model.}

\item{type}{The type of time series: \code{univariate} or \code{multivariate}.}
}
\value{
The lag number for a LSTM.
}
\description{
Get ANN lag from ARIMA(X) lag
}
\seealso{
Other Recurrent Neural Network (RNN): 
\code{\link{as_LSTM_X}()},
\code{\link{as_LSTM_Y}()},
\code{\link{as_LSTM_data_frame}()},
\code{\link{as_LSTM_period_outcome}()},
\code{\link{as_timesteps}()},
\code{\link{build_LSTM}()},
\code{\link{fit_LSTM}()},
\code{\link{get_LSTM_XY}()},
\code{\link{get_period_shift}()},
\code{\link{load_weights_ANN}()},
\code{\link{predict_ANN}()},
\code{\link{save_weights_ANN}()},
\code{\link{start_invert_differencing}()}
}
\concept{Recurrent Neural Network (RNN)}
