% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepRNN.r
\name{as_LSTM_data_frame}
\alias{as_LSTM_data_frame}
\title{Recreation of a data frame based on preformatted X and Y data sets}
\usage{
as_LSTM_data_frame(x, y, xnames, ynames, timesteps = 1L, suffix = "_t")
}
\arguments{
\item{x}{A feature data set, usually a matrix or data frame, returned by \code{get_LSTM_XY}.}

\item{y}{An outcome data set, usually a vector, matrix or data frame, returned by \code{get_LSTM_XY}.}

\item{xnames}{Names of the features.}

\item{ynames}{Names of the outcomes.}

\item{timesteps}{Number of timesteps; stands for the number of different periods within one sample (record) of the result, the resampled feature matrix \code{x}.}

\item{suffix}{The suffix for every feature per timestep or period.}
}
\value{
A data frame with outcome column(s) and a further resampled feature matrix.
The feature matrix within this data frame has the following forward oriented form:
x1(t1), x1(t2), x1(t3)...x2(t1), x2(t2), x2(t3)...x3(t1), x3(t2), x3(t3)...
}
\description{
Recreation of a data frame based on preformatted X and Y data sets
}
\seealso{
\code{\link{get_LSTM_XY}}.

Other Recurrent Neural Network (RNN): 
\code{\link{as_LSTM_X}()},
\code{\link{as_LSTM_Y}()},
\code{\link{as_LSTM_period_outcome}()},
\code{\link{as_lag}()},
\code{\link{as_timesteps}()},
\code{\link{build_LSTM}()},
\code{\link{fit_LSTM}()},
\code{\link{get_LSTM_XY}()},
\code{\link{get_period_shift}()},
\code{\link{load_weights_ANN}()},
\code{\link{predict_ANN}()},
\code{\link{save_weights_ANN}()},
\code{\link{start_invert_differencing}()}
}
\concept{Recurrent Neural Network (RNN)}
