% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepRNN.r
\name{as_LSTM_period_outcome}
\alias{as_LSTM_period_outcome}
\title{Rebuild data frame}
\usage{
as_LSTM_period_outcome(
  dataset,
  columns,
  timesteps = 1L,
  lag = 0L,
  type = "univariate"
)
}
\arguments{
\item{dataset}{A data set, usually a matrix or data frame with training or test data.}

\item{columns}{The names or indices of the columns which should be extracted.}

\item{timesteps}{The number of timesteps; stands for the number of different periods within one sample (record) of the resampled feature matrix, returned by \code{as_LSTM_X}.}

\item{lag}{The number of considered lags on feature side.}

\item{type}{The type of time series: \code{univariate} or \code{multivariate}.}
}
\value{
A data frame with \code{columns} that can be used for quality assurance or
for graphical representation together with the predicted values produced by \code{predict_LSTM}.
}
\description{
Extract columns from a data frame under consideration of time series type, timesteps and lags.
}
\details{
Within time series analysis, \code{get_LSTM_XY} extracts X and Y of a matrix or data frame in a LSTM compatible preformat.
Then, these values are resampled thru \code{as_LSTM_X} and \code{as_LSTM_Y}. During extraction, the period shift is calculated
to determine the number of past periods which are used to extract X values for Y description/prediction. Because of this number
of periods cannot be used to extract Y values, this number must be deleted from \code{dataset}.
}
\seealso{
\code{\link{get_LSTM_XY}}, \code{\link{get_period_shift}}, \code{\link{predict_ANN}}.

Other Recurrent Neural Network (RNN): 
\code{\link{as_LSTM_X}()},
\code{\link{as_LSTM_Y}()},
\code{\link{as_LSTM_data_frame}()},
\code{\link{as_lag}()},
\code{\link{as_timesteps}()},
\code{\link{build_LSTM}()},
\code{\link{fit_LSTM}()},
\code{\link{get_LSTM_XY}()},
\code{\link{get_period_shift}()},
\code{\link{load_weights_ANN}()},
\code{\link{predict_ANN}()},
\code{\link{save_weights_ANN}()},
\code{\link{start_invert_differencing}()}
}
\concept{Recurrent Neural Network (RNN)}
