% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepRNN.r
\name{as_LSTM_X}
\alias{as_LSTM_X}
\title{Features (X) data format for LSTM}
\usage{
as_LSTM_X(x, timesteps = 1L, reverse = FALSE)
}
\arguments{
\item{x}{A feature data set, usually a matrix or data frame, returned by \code{get_LSTM_XY}.}

\item{timesteps}{Number of timesteps; stands for the number of different periods within one sample (record) of the result, the resampled feature matrix \code{x}.}

\item{reverse}{Controls the order of the values in the resampled feature matrix \code{X}. By default they are used in the given order (forward in time), but they can also be used in reverse order (backward in time).}
}
\value{
A three-dimensional array of the resampled feature matrix \code{x} needed within Tensorflow for recurrent neural networks, e.g. LSTM.\cr
  1. dimension: Samples (s) = Number of records\cr
  2. dimension: Timesteps (t) = Number of different periods within one record\cr
  3. dimension: Features (f) = Number of features within a sequence of a period\cr
  Note: A 3D-array with dimensions (s x t x f) can be interpret as f (s x t)-matrices.
}
\description{

}
\seealso{
\code{\link{get_LSTM_XY}}, \code{\link{as_LSTM_Y}}, \code{\link{as_ANN_matrix}}, \code{\link{as_tensor_3D}}.

Other Recurrent Neural Network (RNN): 
\code{\link{as_LSTM_Y}()},
\code{\link{as_LSTM_data_frame}()},
\code{\link{as_LSTM_period_outcome}()},
\code{\link{as_lag}()},
\code{\link{as_timesteps}()},
\code{\link{build_LSTM}()},
\code{\link{fit_LSTM}()},
\code{\link{get_LSTM_XY}()},
\code{\link{get_period_shift}()},
\code{\link{load_weights_ANN}()},
\code{\link{predict_ANN}()},
\code{\link{save_weights_ANN}()},
\code{\link{start_invert_differencing}()}
}
\concept{Recurrent Neural Network (RNN)}
