% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepRNN.r
\name{as_LSTM_Y}
\alias{as_LSTM_Y}
\title{Outcomes (Y) data format for LSTM}
\usage{
as_LSTM_Y(y, timesteps = 1L, reverse = FALSE)
}
\arguments{
\item{y}{An outcome data set, usually a vector, matrix or data frame, returned by \code{get_LSTM_XY}.}

\item{timesteps}{Number of timesteps; stands for the number of different periods within one sample (record) of the result, the resampled outcome matrix \code{y}.}

\item{reverse}{Controls the order of the values in the resampled outcome matrix \code{Y}. By default they are used in the given order (forward in time), but they can also be used in reverse order (backward in time).}
}
\value{
Dependent on the type of \code{Y} and timesteps. If \code{y} is a factor, the result is a one-hot vector.
  If \code{timesteps = NULL|1} a 2D-array with the dimensions (1) samples as number of records and (2) number of output units, representing a scalar outcome \code{y},
  if \code{timesteps >= 2} a 3D-array with the dimensions (1) samples, (2) timesteps and (3) number of output units, representing a sequence or multi-step outcome \code{y}.
}
\description{

}
\seealso{
\code{\link{get_LSTM_XY}}, \code{\link{as_LSTM_X}}, \code{\link{as_ANN_matrix}}, \code{\link{one_hot_encode}},
  \code{\link{as_tensor_2d}}, \code{\link{as_tensor_3d}}.

Other Recurrent Neural Network (RNN): 
\code{\link{as_LSTM_X}()},
\code{\link{as_LSTM_data_frame}()},
\code{\link{as_LSTM_period_outcome}()},
\code{\link{as_lag}()},
\code{\link{as_timesteps}()},
\code{\link{build_LSTM}()},
\code{\link{fit_LSTM}()},
\code{\link{get_LSTM_XY}()},
\code{\link{get_period_shift}()},
\code{\link{load_weights_ANN}()},
\code{\link{predict_ANN}()},
\code{\link{save_weights_ANN}()},
\code{\link{start_invert_differencing}()}
}
\concept{Recurrent Neural Network (RNN)}
