% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepMLP.r
\name{as_MLP_X}
\alias{as_MLP_X}
\title{Features (X) data format for SLP/MLP}
\usage{
as_MLP_X(x)
}
\arguments{
\item{x}{A feature data set, usually a matrix or data frame.}
}
\value{
A two-dimensional array of the feature matrix \code{x}.
}
\description{
Features (X) data format for SLP/MLP
}
\seealso{
\code{\link{as_MLP_Y}}.

Other Single & Multi Layer Perceptron (SLP, MLP): 
\code{\link{as_MLP_Y}()},
\code{\link{as_tensor_1d}()},
\code{\link{as_tensor_2d}()},
\code{\link{as_tensor_3d}()},
\code{\link{build_MLP}()},
\code{\link{fit_MLP}()},
\code{\link{load_weights_ANN}()},
\code{\link{nsamples}()},
\code{\link{nsubsequences}()},
\code{\link{ntimesteps}()},
\code{\link{nunits}()},
\code{\link{predict_ANN}()},
\code{\link{save_weights_ANN}()}
}
\concept{Single & Multi Layer Perceptron (SLP, MLP)}
