% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepMLP.r
\name{as_MLP_Y}
\alias{as_MLP_Y}
\title{Outcomes (Y) data format for SLP/MLP}
\usage{
as_MLP_Y(y)
}
\arguments{
\item{y}{An outcome data set, usually a vector, matrix or data frame.}
}
\value{
A two-dimensional array of the outcome \code{Y}. For a factor outcome, the result is a one-hot vector.
}
\description{

}
\seealso{
\code{\link{as_MLP_X}}, \code{\link{one_hot_encode}}.

Other Single & Multi Layer Perceptron (SLP, MLP): 
\code{\link{as_MLP_X}()},
\code{\link{build_MLP}()},
\code{\link{fit_MLP}()},
\code{\link{load_weights_ANN}()},
\code{\link{predict_ANN}()},
\code{\link{save_weights_ANN}()}
}
\concept{Single & Multi Layer Perceptron (SLP, MLP)}
