% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepMLP.r
\name{as_tensor_1d}
\alias{as_tensor_1d}
\title{Transform data into a 1D tensor}
\usage{
as_tensor_1d(data, order = c("C", "F"))
}
\arguments{
\item{data}{The data to be reshaped to a one-dimensional tensor.}

\item{order}{The order in which elements of \code{data} should be read during flattening.}
}
\value{
A one-dimensional array.
}
\description{
Transform data into a 1D tensor
}
\seealso{
\code{\link{flatten}}.

Other Single & Multi Layer Perceptron (SLP, MLP): 
\code{\link{as_MLP_X}()},
\code{\link{as_MLP_Y}()},
\code{\link{as_tensor_2d}()},
\code{\link{as_tensor_3d}()},
\code{\link{build_MLP}()},
\code{\link{fit_MLP}()},
\code{\link{load_weights_ANN}()},
\code{\link{nsamples}()},
\code{\link{nsubsequences}()},
\code{\link{ntimesteps}()},
\code{\link{nunits}()},
\code{\link{predict_ANN}()},
\code{\link{save_weights_ANN}()}
}
\concept{Single & Multi Layer Perceptron (SLP, MLP)}
