% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepMLP.r
\name{as_tensor_2d}
\alias{as_tensor_2d}
\title{Transform data into a 2D tensor.}
\usage{
as_tensor_2d(data)
}
\arguments{
\item{data}{The data to be reshaped to a two-dimensional tensor.}
}
\value{
A two-dimensional array.
}
\description{
Transform data into a 2D tensor.
}
\seealso{
Other Single & Multi Layer Perceptron (SLP, MLP): 
\code{\link{as_MLP_X}()},
\code{\link{as_MLP_Y}()},
\code{\link{as_tensor_1d}()},
\code{\link{as_tensor_3d}()},
\code{\link{build_MLP}()},
\code{\link{fit_MLP}()},
\code{\link{load_weights_ANN}()},
\code{\link{nsamples}()},
\code{\link{nsubsequences}()},
\code{\link{ntimesteps}()},
\code{\link{nunits}()},
\code{\link{predict_ANN}()},
\code{\link{save_weights_ANN}()}
}
\concept{Single & Multi Layer Perceptron (SLP, MLP)}
