% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepMetrics.r
\docType{data}
\name{epsilon}
\alias{epsilon}
\title{Backend}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
epsilon
}
\description{
Backend
}
\details{
A fuzz factor used in numeric expressions.Default: \code{1e-7}
}
\keyword{datasets}
