% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepMLP.r
\name{build.MLP}
\alias{build.MLP}
\title{Build SLP/MLP architecture}
\usage{
build.MLP(
  features,
  hidden = NULL,
  dropout = NULL,
  output = c(1, "linear"),
  loss = "mean_squared_error",
  optimizer = "adam",
  metrics = c("mean_absolute_error")
)
}
\arguments{
\item{features}{Number of features, returned by \code{get.MLP.X.units}.}

\item{hidden}{A data frame with two columns whereby the first column contains the number of hidden units
and the second column the activation function. The number of rows determines the number of hidden layers.}

\item{dropout}{A numeric vector with dropout rates, the fractions of input units to drop or \code{NULL} if no dropout is desired.}

\item{output}{A vector with two elements whereby the first element determines the number of output units, returned by \code{get.MLP.Y.units},
and the second element the output activation function.}

\item{loss}{Name of objective function or objective function. If the model has multiple outputs,
different loss on each output can be used by passing a dictionary or a list of objectives.
The loss value that will be minimized by the model will then be the sum of all individual losses.}

\item{optimizer}{Name of optimizer or optimizer instance.}

\item{metrics}{Vector or list of metrics to be evaluated by the model during training and testing.}
}
\value{
A model object with stacked dense layers and dropout layers.
}
\description{
\code{build.MLP} creates a sequential feedforward model (SLP, MLP) with stacked dense layers and optional dropout layers.
}
\seealso{
\code{\link{get.MLP.X.units}}, \code{\link{get.MLP.Y.units}},
  \code{\link[keras]{keras_model_sequential}}, \code{\link[keras]{layer_dense}}, \code{\link[keras]{layer_dropout}},
  \code{\link[keras]{compile.keras.engine.training.Model}}.

Other Single & Multi Layer Perceptron (SLP, MLP): 
\code{\link{as.ANN.matrix}()},
\code{\link{as.MLP.X}()},
\code{\link{as.MLP.Y}()},
\code{\link{as.tensor.1D}()},
\code{\link{as.tensor.2D}()},
\code{\link{as.tensor.3D}()},
\code{\link{as.tensor}()},
\code{\link{fit.MLP}()},
\code{\link{get.MLP.X.samples}()},
\code{\link{get.MLP.X.units}()},
\code{\link{get.MLP.Y.samples}()},
\code{\link{get.MLP.Y.units}()},
\code{\link{predict.ANN}()},
\code{\link{vector.as.ANN.matrix}()},
\code{\link{vector.as.numeric}()}
}
\concept{Single & Multi Layer Perceptron (SLP, MLP)}
