% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepTimeSeries.r
\name{build.stationary}
\alias{build.stationary}
\title{Build a stationary data series thru differencing within a data set}
\usage{
build.stationary(
  dataset,
  y = 2,
  differences = 1,
  suffix = "_delta",
  adjust = TRUE
)
}
\arguments{
\item{dataset}{A data set, usually a data frame.}

\item{y}{The indices of the columns which are used to build stationary series.}

\item{differences}{The number of differences for building stationary series.}

\item{suffix}{The suffix for every newly created column of the stationary series.}

\item{adjust}{Controls whether NA values are included to fill up the entire data set in the newly
created columns for the stationary series (\code{FALSE}) or the entire data set is shorten to the length
of the stationary data series (\code{TRUE}).}
}
\value{
The data set with newly created columns for the stationary data series.
}
\description{
\code{build.stationary} builds a differenced data series.
}
\seealso{
\code{\link{invert_differencing}}.

Other TimeSeries: 
\code{\link{get.season}()},
\code{\link{invert_differencing}()},
\code{\link{lags}()},
\code{\link{period}()}
}
\concept{TimeSeries}
