% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepQuality.r
\name{coerce_dimension}
\alias{coerce_dimension}
\title{Coerce data to an array or matrix with no trailing dimension of 1
0}
\usage{
coerce_dimension(x)
}
\arguments{
\item{x}{A data structure like vector, matrix, array, data frame or list.}
}
\value{
The coerced data structure with no trailing dimension of 1.
}
\description{
Coerce data to an array or matrix with no trailing dimension of 1
0
}
