% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepArray.r
\name{column_stack}
\alias{column_stack}
\title{Array stack}
\usage{
column_stack(..., order = c("C", "F"))
}
\arguments{
\item{...}{Any numbers of objects they are coerced to 1D arrays. The objects can be packed into a \code{list}. The length of these objects must be equal.}

\item{order}{The order in which elements of the objects should be read during coercing to arrays.
By default, the order is equivalent to the \code{C}-style ordering and means elements should be read in row-major order.
In opposite, the \code{Fortran}-style ordering means elements should be read in column-major order.}
}
\value{
An 2D array formed by stacking all input arrays.
}
\description{
Stack 1D arrays as columns into a 2D array.
}
\details{
This function corresponds to \code{column_stack()} from NumPy.
}
\examples{
a <- 1:3
b <- 4:6
column_stack(a, b)
}
\seealso{
Other Array: 
\code{\link{DIM}()},
\code{\link{delete}()},
\code{\link{dimC<-}()},
\code{\link{embedseries}()},
\code{\link{empty}()},
\code{\link{expand_dims}()},
\code{\link{eye}()},
\code{\link{flatten}()},
\code{\link{flip}()},
\code{\link{full}()},
\code{\link{insert}()},
\code{\link{mabind}()},
\code{\link{mamatrix}()},
\code{\link{marray}()},
\code{\link{ndim}()},
\code{\link{nsize}()},
\code{\link{ones}()},
\code{\link{rearrange}()},
\code{\link{reshape.array}()},
\code{\link{rot90}()},
\code{\link{row_stack}()},
\code{\link{slice}()},
\code{\link{squeeze}()},
\code{\link{transpose}()},
\code{\link{vander}()},
\code{\link{zeros}()}
}
\concept{Array}
