% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepSampler.r
\name{concatenate.factor}
\alias{concatenate.factor}
\title{Concatenate two or more objects into a factor object}
\usage{
concatenate.factor(...)
}
\arguments{
\item{...}{Any number of objects that are combined into a factor.}
}
\value{
A factor.
}
\description{
Concatenate two or more objects into a factor object
}
