% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepQuality.r
\name{cross_entropy}
\alias{cross_entropy}
\title{Cross entropy}
\usage{
cross_entropy(p, q, base = NULL)
}
\arguments{
\item{p}{A vector of ground truth probabilities (true probability distribution).}

\item{q}{A vector of estimated probabilities (estimated probability distribution).}

\item{base}{A positive or complex number: the base with respect to which logarithms are computed.
Defaults to \code{NULL} is equal to e = \code{exp(1)}.}
}
\value{
Cross entropy.
}
\description{
Cross entropy
}
\details{
Cross entropy quantifies the difference between two probability distributions.
}
\examples{
  p <- c(0.10, 0.40, 0.50) # ground truth values
  q <- c(0.80, 0.15, 0.05) # estimated values, e.g. given by softmax function
  cross_entropy(p, q)
}
\seealso{
Other Quality: 
\code{\link{accuracy.microaveraging}()},
\code{\link{accuracy}()},
\code{\link{entropy}()},
\code{\link{gini_impurity}()},
\code{\link{huber_loss}()},
\code{\link{log_cosh_loss}()},
\code{\link{mae}()},
\code{\link{mape}()},
\code{\link{mse}()},
\code{\link{msle}()},
\code{\link{quantile_loss}()},
\code{\link{rmse}()},
\code{\link{rmsle}()},
\code{\link{vc}()}
}
\concept{Quality}
