% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepML.r
\name{cross_validation_split}
\alias{cross_validation_split}
\title{K-fold cross validation}
\usage{
cross_validation_split(dataset, folds = 3L, shuffle = FALSE)
}
\arguments{
\item{dataset}{A data set, usually a data frame.}

\item{folds}{Number of created folds.}

\item{shuffle}{Controls whether the samples of the data set should be randomly shuffled before fold creation.
For time series data, this argument must be set equal to \code{FALSE} because the order of the samples can't be changed.}
}
\value{
A named list with folds.
}
\description{
\code{cross_validation} splits a data set in partial sets, so-called folds, and creates a list of folds.
}
\seealso{
Other Machine Learning: 
\code{\link{distance}()},
\code{\link{erfcinv}()},
\code{\link{erfc}()},
\code{\link{erfinv}()},
\code{\link{erf}()},
\code{\link{k_nearest_neighbors}()},
\code{\link{moving_average}()},
\code{\link{naive_forecast}()},
\code{\link{predict.kmeans}()},
\code{\link{similarity}()}
}
\concept{Machine Learning}
