% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepSelection.r
\name{data_split}
\alias{data_split}
\alias{data_split.default}
\alias{data_split.array}
\alias{data_split.data.frame}
\alias{data_split.factor}
\title{Data split}
\usage{
data_split(.data, ...)

\method{data_split}{default}(
  .data,
  test_size = 0.25,
  train_size = 0.75,
  random_state = NULL,
  shuffle = TRUE,
  axis = 1
)

\method{data_split}{array}(
  .data,
  test_size = 0.25,
  train_size = 0.75,
  random_state = NULL,
  shuffle = TRUE,
  axis = 1
)

\method{data_split}{data.frame}(
  .data,
  test_size = 0.25,
  train_size = 0.75,
  random_state = NULL,
  shuffle = TRUE
)

\method{data_split}{factor}(
  .data,
  test_size = 0.25,
  train_size = 0.75,
  random_state = NULL,
  shuffle = TRUE
)
}
\arguments{
\item{.data}{Data to be split.}

\item{test_size}{An integer or float value between 0.0 and 1.0 representing the absolute number of test samples (integer) or the proportion of the data to include in the test split (float).
If \code{NULL} (default), the value is set to the complement of the \code{train size}. If \code{train_size} is also \code{NULL}, it will be set to 0.25.}

\item{train_size}{An integer or float value between 0.0 and 1.0 representing the absolute number of train samples (integer) or the proportion of the data to include in the train split (float).
If \code{NULL} (default), the value is set to the complement of the \code{test size}.}

\item{random_state}{Controls the shuffling applied to the data before applying the split. Pass an int for reproducible output across multiple function calls.}

\item{shuffle}{Whether or not to shuffle the data before splitting.}

\item{axis}{The axis along to split. Default is the first axis.}
}
\value{
List containing data pieces.
}
\description{
Split data into two pieces, usually train and test pieces.
}
