% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepMisc.r
\docType{data}
\name{day.name}
\alias{day.name}
\alias{day.abb}
\title{Built-in Constants}
\format{
An object of class \code{character} of length 7.

An object of class \code{character} of length 7.
}
\usage{
\code{day.name}

\code{day.abb}
}
\description{
The constants \code{day.name} and \code{day.abb} are additions to the built-in constants \code{month.name} and \code{month.abb} in \code{R}.
}
\details{
The following constants are available:
  \itemize{
  \item \code{day.name}: the English names for the day of the week: Monday, Tuesday, Wednesday, Thursday, Friday, Saturday, Sunday.
  \item \code{day.abb}: the three-letter abbreviations for the English day names: Mon, Tue, Wed, Thu, Fri, Sat, Sun.
  }
}
\seealso{
\code{\link{month.name}}, \code{\link{month.abb}}
}
\keyword{datasets}
