% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepArray.r
\name{delete}
\alias{delete}
\title{Array deletion}
\usage{
delete(a, axis = 1L, order = c("C", "F"))
}
\arguments{
\item{a}{An array.}

\item{axis}{The axis or axes to delete from \code{a}.}

\item{order}{The order in which elements of \code{x} should be read during recreation after deleting \code{axis}.
By default, the order is equivalent to the \code{C}-style ordering and means elements should be read in row-major order.
In opposite, the \code{Fortran}-style ordering means elements should be read in column-major order.}
}
\value{
The array \code{a} with deleted axes.
}
\description{
Delete axis of an array.
}
\details{
This function corresponds partially to \code{delete()} from NumPy.
  The number of elements is tailored for the reshaped array. The sign of \code{axis} determines the reading direction for the extraction
  of the elements from \code{a} for the newly created array. A positive sign indicates that the first n elements from \code{a} are used,
  a negative sign indicates that the last n elements will be used.
}
\examples{
# original array
a <- marray(1:24, dim = c(4, 3, 2), order = "F")
# delete the first dimension with reading elements start from first position
delete(a)
# delete the first dimension with reading elements start from last position
delete(a, axis = -1L)
# delete the axes one and two with reading elements start from first position
delete(a, axis = c(1L, 2L))

}
\seealso{
Other Array: 
\code{\link{DIM}()},
\code{\link{column_stack}()},
\code{\link{dimC<-}()},
\code{\link{embedseries}()},
\code{\link{empty}()},
\code{\link{expand_dims}()},
\code{\link{eye}()},
\code{\link{flatten}()},
\code{\link{flip}()},
\code{\link{full}()},
\code{\link{insert}()},
\code{\link{mabind}()},
\code{\link{mamatrix}()},
\code{\link{marray}()},
\code{\link{ndim}()},
\code{\link{nsize}()},
\code{\link{ones}()},
\code{\link{rearrange}()},
\code{\link{reshape.array}()},
\code{\link{rot90}()},
\code{\link{row_stack}()},
\code{\link{slice}()},
\code{\link{squeeze}()},
\code{\link{transpose}()},
\code{\link{vander}()},
\code{\link{zeros}()}
}
\concept{Array}
