% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepScaling.r
\name{denormalize}
\alias{denormalize}
\title{Invert Min-Max-Scaling}
\usage{
denormalize(x, minx, maxx)
}
\arguments{
\item{x}{A vector with min-max-scaled values.}

\item{minx}{A given minimum value.}

\item{maxx}{A given maximum value.}
}
\value{
A vector of same length as \code{x} with rescaled values.
}
\description{
Invert Min-Max-Scaling
}
\seealso{
\code{\link{normalize}}.

Other Scaling: 
\code{\link{normalize}()}
}
\concept{Scaling}
