% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepTimeSeries.r
\name{diff_log}
\alias{diff_log}
\title{Log-differencing of a numeric vector}
\usage{
diff_log(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
The log-differenced \code{x}.
}
\description{

}
\seealso{
\code{\link{stationary}}, \code{\link{diffinv_log}}.

Other Time Series: 
\code{\link{diff_percentage}()},
\code{\link{diffinv_log}()},
\code{\link{diffinv_percentage}()},
\code{\link{diffinv_simple}()},
\code{\link{get_season}()},
\code{\link{invert_differencing}()},
\code{\link{lags}()},
\code{\link{partition}()},
\code{\link{period}()},
\code{\link{stationary}()}
}
\concept{Time Series}
