% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepTimeSeries.r
\name{diffinv.percentage}
\alias{diffinv.percentage}
\title{Invert a percentage-differentiated vector}
\usage{
\method{diffinv}{percentage}(delta, origin)
}
\arguments{
\item{delta}{A percentage-differentiated numeric vector.}

\item{origin}{A scalar or numeric vector with original value(s) to invert the differentiated data series.}
}
\value{
The inverted percentage-differentiation \code{delta}.
}
\description{
Invert a percentage-differentiated vector
}
\seealso{
\code{\link{diff.percentage}}.

Other TimeSeries: 
\code{\link{build.stationary}()},
\code{\link{diff.log}()},
\code{\link{diff.percentage}()},
\code{\link{diffinv.log}()},
\code{\link{diffinv.simple}()},
\code{\link{get.season}()},
\code{\link{invert_differencing}()},
\code{\link{lags}()},
\code{\link{period}()}
}
\concept{TimeSeries}
