% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepTimeSeries.r
\name{diffinv_percentage}
\alias{diffinv_percentage}
\title{Invert a percentage-differenced vector}
\usage{
diffinv_percentage(delta, origin)
}
\arguments{
\item{delta}{A percentage-differenced numeric vector.}

\item{origin}{A scalar or numeric vector with original value(s) to invert \code{delta}.}
}
\value{
The inverted \code{delta}.
}
\description{

}
\seealso{
\code{\link{diff_percentage}}.

Other Time Series: 
\code{\link{diff_log}()},
\code{\link{diff_percentage}()},
\code{\link{diffinv_log}()},
\code{\link{diffinv_simple}()},
\code{\link{get_season}()},
\code{\link{invert_differencing}()},
\code{\link{lags}()},
\code{\link{partition}()},
\code{\link{period}()},
\code{\link{stationary}()}
}
\concept{Time Series}
