% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepArray.r
\name{DIM}
\alias{DIM}
\title{Retrieve dimensions of an object or its length}
\usage{
DIM(x)
}
\arguments{
\item{x}{An R object.}
}
\value{
The number of dimensions or in cases of an atomic object the length.
}
\description{
Retrieve dimensions of an object or its length
}
\references{
Implementation credits go to \url{https://github.com/t-kalinowski/listarrays}.
}
\seealso{
Other Array: 
\code{\link{column_stack}()},
\code{\link{delete}()},
\code{\link{dimC<-}()},
\code{\link{empty}()},
\code{\link{expand_dims}()},
\code{\link{eye}()},
\code{\link{flatten}()},
\code{\link{flip}()},
\code{\link{full}()},
\code{\link{insert}()},
\code{\link{mabind}()},
\code{\link{mamatrix}()},
\code{\link{marray}()},
\code{\link{ndim}()},
\code{\link{nsize}()},
\code{\link{ones}()},
\code{\link{reshape.array}()},
\code{\link{rot90}()},
\code{\link{row_stack}()},
\code{\link{slice}()},
\code{\link{squeeze}()},
\code{\link{transpose}()},
\code{\link{vander}()},
\code{\link{zeros}()}
}
\concept{Array}
