% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepDummy.r
\name{dummify_multilabel}
\alias{dummify_multilabel}
\title{Create dummy variables for multi-label columns}
\usage{
dummify_multilabel(
  dataset,
  columns = NULL,
  split = ",",
  effectcoding = FALSE,
  prefix = FALSE,
  remove_columns = FALSE
)
}
\arguments{
\item{dataset}{A data set with multi-label character variables.}

\item{columns}{The names or indices of the columns for which dummy variables are to be created; if \code{NULL} (default), all character columns are encoded.}

\item{split}{A character vector (or object which can be coerced to such) containing regular expression(s) to use for splitting.}

\item{effectcoding}{Instead of using default 0/1 value pairs for dummy variables, effectcoding allows to set -1/1 pairs.}

\item{prefix}{A logical value indicating whether the names of the dummy variables have the corresponding column names as a prefix.}

\item{remove_columns}{A logical value indicating whether the character variables should be removed from \code{dataset} after they have been encoded in dummy variables.}
}
\value{
The data set with encoded dummy variables.
}
\description{

}
\seealso{
\code{\link{effectcoding}}, \code{\link[base]{strsplit}}.

Other Dummifying: 
\code{\link{append_rows}()},
\code{\link{dummify}()},
\code{\link{effectcoding}()},
\code{\link{one_hot_decode}()},
\code{\link{one_hot_encode}()},
\code{\link{remove_columns}()},
\code{\link{resample_imbalanced}()}
}
\concept{Dummifying}
