% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepArray.r
\name{embedseries}
\alias{embedseries}
\alias{embedseries.default}
\alias{embedseries.matrix}
\alias{embedseries.data.frame}
\title{Series embedding}
\usage{
embedseries(data, length)

\method{embedseries}{default}(data, length = 1L)

\method{embedseries}{matrix}(data, length = 1L)

\method{embedseries}{data.frame}(data, length = 1L)
}
\arguments{
\item{data}{The data to be embedded into a shifted series.}

\item{length}{The length of a series.}
}
\value{
An ongoing shifted series of \code{data}.
}
\description{
Series embedding
}
\examples{
x <- seq_len(1e+6)
df <- data.frame(a = seq_len(1e+6), b = -seq_len(1e+6))
a <- embedseries(x, length = 11L)
head(a)
a <- embedseries(df, length = 23L)
head(a)
}
\seealso{
Other Array: 
\code{\link{DIM}()},
\code{\link{column_stack}()},
\code{\link{delete}()},
\code{\link{dimC<-}()},
\code{\link{empty}()},
\code{\link{expand_dims}()},
\code{\link{eye}()},
\code{\link{flatten}()},
\code{\link{flip}()},
\code{\link{full}()},
\code{\link{insert}()},
\code{\link{mabind}()},
\code{\link{mamatrix}()},
\code{\link{marray}()},
\code{\link{ndim}()},
\code{\link{nsize}()},
\code{\link{ones}()},
\code{\link{rearrange}()},
\code{\link{reshape.array}()},
\code{\link{rot90}()},
\code{\link{row_stack}()},
\code{\link{slice}()},
\code{\link{squeeze}()},
\code{\link{transpose}()},
\code{\link{vander}()},
\code{\link{zeros}()}
}
\concept{Array}
