% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepEncoder.r
\docType{class}
\name{Encoder}
\alias{Encoder}
\alias{is.Encoder}
\title{Class Encoder}
\usage{
is.Encoder(object)
}
\description{
This is the abstract base class for encoder objects.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Encoder-clone}{\code{Encoder$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Encoder-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Encoder-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Encoder$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
