% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepQuality.r
\name{entropy}
\alias{entropy}
\title{Shannon-Entropy}
\usage{
entropy(x)
}
\arguments{
\item{x}{A vector of values, usually character labels as raw instances or as class frequencies.}
}
\value{
Entropy.
}
\description{
Shannon-Entropy
}
\details{
Shannon-Entropy is a concept from information theory and represents a quantification of the level 
  of impurity or randomness that exists within a partition with class levels of a factor variable \code{x}.
}
\examples{
  entropy(c("no", "no", "yes", "yes", "yes", "no", "yes", "no", "yes", "yes", "yes", "yes", "yes", "no"))
  entropy(c("no" = 5, "yes" = 9))
}
\seealso{
Other Quality: 
\code{\link{accuracy}()},
\code{\link{gini_impurity}()},
\code{\link{huber_loss}()},
\code{\link{log_cosh_loss}()},
\code{\link{mae}()},
\code{\link{mape}()},
\code{\link{mse}()},
\code{\link{msle}()},
\code{\link{quantile_loss}()},
\code{\link{rmse}()},
\code{\link{rmsle}()},
\code{\link{vc}()}
}
\concept{Quality}
