% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepML.r
\name{erfc}
\alias{erfc}
\title{Complementary error function (from MATLAB)}
\usage{
erfc(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
Complementary error function, defined as 1 - \code{erf}.
}
\description{
Complementary error function (from MATLAB)
}
