% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepML.r
\name{erfcinv}
\alias{erfcinv}
\title{Inverse complementary error function (from MATLAB)}
\usage{
erfcinv(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
Inverse complementary error function.
}
\description{
Inverse complementary error function (from MATLAB)
}
