% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepML.r
\name{erfinv}
\alias{erfinv}
\title{Inverse error function (from MATLAB)}
\usage{
erfinv(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
Inverse error function.
}
\description{
Inverse error function (from MATLAB)
}
