% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepML.r
\name{euclidean_distance}
\alias{euclidean_distance}
\title{Euclidean distance}
\usage{
euclidean_distance(x1, x2)
}
\arguments{
\item{x1}{A numeric vector.}

\item{x2}{A numeric vector.}
}
\value{
The euclidean distance between these two vectors.
}
\description{
\code{euclidean_distance} calculates the euclidean distance between two vectors (or scalars).
}
\examples{
  x1 <- c(20, 1, 41, 13, 5, 69)
  x2 <- c(11, 2, 23, 4, 10, 67)
  euclidean_distance(x1, x2)
}
\seealso{
Other Machine Learning: 
\code{\link{cross_validation_split}()},
\code{\link{erfcinv}()},
\code{\link{erfc}()},
\code{\link{erfinv}()},
\code{\link{erf}()},
\code{\link{k_nearest_neighbors}()},
\code{\link{moving_average}()},
\code{\link{naive_forecast}()},
\code{\link{predict.kmeans}()}
}
\concept{Machine Learning}
