% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepArray.r
\name{eye}
\alias{eye}
\title{Multidimensional array creation: 2D identity matrix}
\usage{
eye(n, m = NULL)
}
\arguments{
\item{n}{The number of rows.}

\item{m}{The number of columns, default \code{NULL}.}
}
\value{
An identity matrix with n rows and n or rather m columns.
}
\description{
Multidimensional array creation: 2D identity matrix
}
\details{
The Python package NumPy offers an analog array creation function \code{np.eye(n, m)}.
}
\seealso{
Other Array: 
\code{\link{column_stack}()},
\code{\link{dimC<-}()},
\code{\link{flatten}()},
\code{\link{insert}()},
\code{\link{mabind}()},
\code{\link{mamatrix}()},
\code{\link{marray}()},
\code{\link{ndim}()},
\code{\link{nsize}()},
\code{\link{reshape.array}()},
\code{\link{row_stack}()},
\code{\link{slice}()},
\code{\link{squeeze}()},
\code{\link{transpose}()},
\code{\link{vander}()}
}
\concept{Array}
