% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepRNN.r
\name{fit_LSTM}
\alias{fit_LSTM}
\title{Fit LSTM model}
\usage{
fit_LSTM(
  model,
  x,
  y,
  timesteps = 1L,
  batch_size = 1,
  epochs = 10,
  verbose = 1,
  validation_split = 0,
  cross_validation = NULL
)
}
\arguments{
\item{model}{A model object to train, e.g. returned by \code{build_LSTM}.}

\item{x}{A feature data set, usually a matrix or data frame, returned by \code{get_LSTM_XY}.}

\item{y}{An outcome data set, usually a vector, matrix or data frame, returned by \code{get_LSTM_XY}.}

\item{timesteps}{A vector of timesteps for \code{x} and \code{y}.
A feature does always have at least one timestep, but an outcome is either a scalar with one implicit timestep or a sequence with at least two timesteps.
If only one value for \code{timesteps} is given, this value is used for the resampled feature tensor produced by \code{as_LSTM_X}. If two values are given,
the first value is used as before and the second value for the resampled sequence or multi-step outcome tensor produced by \code{as_LSTM_Y}.}

\item{batch_size}{Batch size, the number of samples per gradient update within training process.}

\item{epochs}{Number of epochs to train the model.}

\item{verbose}{Verbosity mode (0 = silent, 1 = progress bar, 2 = one line per epoch) determines how the training progress is visualized.}

\item{validation_split}{Float between 0 and 1. Fraction of the training data used as validation data.}

\item{cross_validation}{List or \code{NULL} (default). The list contains two elements whereby the first element stands for the number of folds (k)
and the second element indicates the type \code{min} or \code{max} for quality measuring.}
}
\value{
A trained model object.
}
\description{
\code{fit_LSTM} is a wrapper function for fitting a LSTM model.
}
\seealso{
\code{\link{build_LSTM}}, \code{\link{get_LSTM_XY}},
\code{\link[keras]{fit.keras.engine.training.Model}}, \code{\link[keras]{evaluate.keras.engine.training.Model}}.

Other Recurrent Neural Network (RNN): 
\code{\link{as_LSTM_X}()},
\code{\link{as_LSTM_Y}()},
\code{\link{as_LSTM_data_frame}()},
\code{\link{as_LSTM_period_outcome}()},
\code{\link{as_lag}()},
\code{\link{as_timesteps}()},
\code{\link{build_LSTM}()},
\code{\link{get_LSTM_XY}()},
\code{\link{get_period_shift}()},
\code{\link{load_weights_ANN}()},
\code{\link{predict_ANN}()},
\code{\link{save_weights_ANN}()},
\code{\link{start_invert_differencing}()}
}
\concept{Recurrent Neural Network (RNN)}
