% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepArray.r
\name{flip}
\alias{flip}
\title{Array flip}
\usage{
flip(a, axis = 1L)
}
\arguments{
\item{a}{An array.}

\item{axis}{Axis or axes along which to flip over.}
}
\value{
The reversed array \code{a} along axes.
}
\description{
Reverse the order of elements in an array along the given axes.
}
\details{
This function corresponds to \code{flip()} from NumPy.
Flipping along an axis can be exemplified with a matrix. If the order of the elements along the first dimension (row) is to be reversed,
it is helpful to imagine a horizontal axis (from left to right) in the middle of the matrix where flipping takes place. The first row
becomes the last, the second row the second last and so on until the last row becomes the first. The same applies for reversing the order
of the elements along the second dimension (column), with the distinction that the flipping axis is a vertical axis (from top to bottom)
in the middle of the matrix.
}
\seealso{
Other Array: 
\code{\link{DIM}()},
\code{\link{column_stack}()},
\code{\link{delete}()},
\code{\link{dimC<-}()},
\code{\link{embedseries}()},
\code{\link{empty}()},
\code{\link{expand_dims}()},
\code{\link{eye}()},
\code{\link{flatten}()},
\code{\link{full}()},
\code{\link{insert}()},
\code{\link{mabind}()},
\code{\link{mamatrix}()},
\code{\link{marray}()},
\code{\link{ndim}()},
\code{\link{nsize}()},
\code{\link{ones}()},
\code{\link{rearrange}()},
\code{\link{reshape.array}()},
\code{\link{rot90}()},
\code{\link{row_stack}()},
\code{\link{slice}()},
\code{\link{squeeze}()},
\code{\link{transpose}()},
\code{\link{vander}()},
\code{\link{zeros}()}
}
\concept{Array}
