% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepArray.r
\name{full}
\alias{full}
\title{Multidimensional array creation: array filled with value}
\usage{
full(dim = NULL, fill_value = NA, order = c("C", "F"))
}
\arguments{
\item{dim}{Shape of the new array.}

\item{fill_value}{Value to fill the array.}

\item{order}{Whether to store multidimensional array in C- or Fortran-contiguous (row- or column-wise) order.}
}
\value{
An array of \code{fill_value} with the given shape.
}
\description{
Multidimensional array creation: array filled with value
}
\details{
This function corresponds to \code{full()} from NumPy.
}
\seealso{
Other Array: 
\code{\link{column_stack}()},
\code{\link{dimC<-}()},
\code{\link{empty}()},
\code{\link{eye}()},
\code{\link{flatten}()},
\code{\link{insert}()},
\code{\link{mabind}()},
\code{\link{mamatrix}()},
\code{\link{marray}()},
\code{\link{ndim}()},
\code{\link{nsize}()},
\code{\link{ones}()},
\code{\link{reshape.array}()},
\code{\link{row_stack}()},
\code{\link{slice}()},
\code{\link{squeeze}()},
\code{\link{transpose}()},
\code{\link{vander}()},
\code{\link{zeros}()}
}
\concept{Array}
