% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepCNN.r
\name{get.CNN.image.X.channels}
\alias{get.CNN.image.X.channels}
\title{Get number of color channels from image feature tensor}
\usage{
get.CNN.image.X.channels(X.tensor)
}
\arguments{
\item{X.tensor}{A tensor of the image features produced by \code{as.CNN.image.X}.}
}
\value{
Number of color channels.
}
\description{
Get number of color channels from image feature tensor
}
\seealso{
\code{\link{as.CNN.image.X}}, \code{\link{get.CNN.image.X.samples}}, \code{\link{get.CNN.image.X.height}}, \code{\link{get.CNN.image.X.width}}.

Other Convolutional Neural Network (CNN): 
\code{\link{as.CNN.image.X}()},
\code{\link{as.CNN.image.Y}()},
\code{\link{get.CNN.image.X.height}()},
\code{\link{get.CNN.image.X.samples}()},
\code{\link{get.CNN.image.X.width}()},
\code{\link{get.CNN.image.Y.samples}()},
\code{\link{get.CNN.image.Y.units}()}
}
\concept{Convolutional Neural Network (CNN)}
