% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepCNN.r
\name{get.CNN.image.X.samples}
\alias{get.CNN.image.X.samples}
\title{Get number of samples from image feature tensor}
\usage{
get.CNN.image.X.samples(X.tensor)
}
\arguments{
\item{X.tensor}{A tensor of the image features produced by \code{as.CNN.image.X}.}
}
\value{
Number of images.
}
\description{
Get number of samples from image feature tensor
}
\seealso{
\code{\link{as.CNN.image.X}}, \code{\link{get.CNN.image.X.height}}, \code{\link{get.CNN.image.X.width}}, \code{\link{get.CNN.image.X.channels}}.

Other Convolutional Neural Network (CNN): 
\code{\link{as.CNN.image.X}()},
\code{\link{as.CNN.image.Y}()},
\code{\link{as.CNN.temp.X}()},
\code{\link{as.CNN.temp.Y}()},
\code{\link{get.CNN.image.X.channels}()},
\code{\link{get.CNN.image.X.height}()},
\code{\link{get.CNN.image.X.width}()},
\code{\link{get.CNN.image.Y.samples}()},
\code{\link{get.CNN.image.Y.units}()},
\code{\link{get.CNN.temp.X.samples}()},
\code{\link{get.CNN.temp.X.subsequences}()},
\code{\link{get.CNN.temp.X.timesteps}()},
\code{\link{get.CNN.temp.X.units}()},
\code{\link{get.CNN.temp.Y.samples}()},
\code{\link{get.CNN.temp.Y.timesteps}()},
\code{\link{get.CNN.temp.Y.units}()}
}
\concept{Convolutional Neural Network (CNN)}
