% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepCNN.r
\name{get.CNN.temp.X.timesteps}
\alias{get.CNN.temp.X.timesteps}
\title{Get number of input timesteps from feature tensor}
\usage{
get.CNN.temp.X.timesteps(X.tensor)
}
\arguments{
\item{X.tensor}{A tensor of the resampled feature matrix produced by \code{as.CNN.temp.X}.}
}
\value{
Number of input timesteps.
}
\description{
Get number of input timesteps from feature tensor
}
\seealso{
\code{\link{as.CNN.temp.X}}, \code{\link{get.CNN.temp.X.samples}}, \code{\link{get.CNN.temp.X.subsequences}}, \code{\link{get.CNN.temp.X.units}}.

Other Convolutional Neural Network (CNN): 
\code{\link{as.CNN.image.X}()},
\code{\link{as.CNN.image.Y}()},
\code{\link{as.CNN.temp.X}()},
\code{\link{as.CNN.temp.Y}()},
\code{\link{get.CNN.image.X.channels}()},
\code{\link{get.CNN.image.X.height}()},
\code{\link{get.CNN.image.X.samples}()},
\code{\link{get.CNN.image.X.width}()},
\code{\link{get.CNN.image.Y.samples}()},
\code{\link{get.CNN.image.Y.units}()},
\code{\link{get.CNN.temp.X.samples}()},
\code{\link{get.CNN.temp.X.subsequences}()},
\code{\link{get.CNN.temp.X.units}()},
\code{\link{get.CNN.temp.Y.samples}()},
\code{\link{get.CNN.temp.Y.timesteps}()},
\code{\link{get.CNN.temp.Y.units}()}
}
\concept{Convolutional Neural Network (CNN)}
