% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepRNN.r
\name{get.LSTM.period_shift}
\alias{get.LSTM.period_shift}
\title{Period shift}
\usage{
get.LSTM.period_shift(timesteps = 1, lag = 0, type = "univariate")
}
\arguments{
\item{timesteps}{The number of timesteps.}

\item{lag}{The number of considered lags on feature side.}

\item{type}{The type of time series: \code{univariate} or \code{multivariate}.}
}
\value{
The period shift.
}
\description{
\code{get.LSTM.period_shift} calculates the period shift for either a univariate or a multivariate time series.
  A period shift denotes the number of periods to go backwards to get features (X) for outcomes (Y).
  In other words: how many samples get lost respectively must be used as X for Y description/prediction?
}
\seealso{
Other Recurrent Neural Network (RNN), Long Short-Term Memory (LSTM): 
\code{\link{as.LSTM.X}()},
\code{\link{as.LSTM.Y}()},
\code{\link{as.LSTM.data.frame}()},
\code{\link{as.LSTM.period_outcome}()},
\code{\link{as.lag}()},
\code{\link{as.timesteps}()},
\code{\link{build.LSTM}()},
\code{\link{fit.LSTM}()},
\code{\link{get.LSTM.X.samples}()},
\code{\link{get.LSTM.X.timesteps}()},
\code{\link{get.LSTM.X.units}()},
\code{\link{get.LSTM.XY}()},
\code{\link{get.LSTM.Y.samples}()},
\code{\link{get.LSTM.Y.timesteps}()},
\code{\link{get.LSTM.Y.units}()},
\code{\link{predict.LSTM}()},
\code{\link{start.LSTM.invert_differencing}()}
}
\concept{Recurrent Neural Network (RNN), Long Short-Term Memory (LSTM)}
