% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepRNN.r
\name{get.LSTM.X.timesteps}
\alias{get.LSTM.X.timesteps}
\title{Get number of input timesteps from feature tensor}
\usage{
get.LSTM.X.timesteps(X.tensor)
}
\arguments{
\item{X.tensor}{A tensor of the resampled feature matrix produced by \code{as.LSTM.X}.}
}
\value{
Number of input timesteps; number of different periods within one sample of the resampled feature matrix.
}
\description{
Get number of input timesteps from feature tensor
}
\seealso{
\code{\link{as.LSTM.X}}, \code{\link{get.LSTM.X.samples}}, \code{\link{get.LSTM.X.units}},
  \code{\link{as.LSTM.Y}}, \code{\link{get.LSTM.Y.samples}}, \code{\link{get.LSTM.Y.units}}.

Other Recurrent Neural Network (RNN): 
\code{\link{as.LSTM.X}()},
\code{\link{as.LSTM.Y}()},
\code{\link{as.LSTM.data.frame}()},
\code{\link{as.LSTM.period_outcome}()},
\code{\link{as.lag}()},
\code{\link{as.timesteps}()},
\code{\link{build.LSTM}()},
\code{\link{fit.LSTM}()},
\code{\link{get.LSTM.X.samples}()},
\code{\link{get.LSTM.X.units}()},
\code{\link{get.LSTM.XY}()},
\code{\link{get.LSTM.Y.samples}()},
\code{\link{get.LSTM.Y.timesteps}()},
\code{\link{get.LSTM.Y.units}()},
\code{\link{get.period_shift}()},
\code{\link{predict.ANN}()},
\code{\link{start.invert_differencing}()}
}
\concept{Recurrent Neural Network (RNN)}
