% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepRNN.r
\name{get.LSTM.XY}
\alias{get.LSTM.XY}
\title{Extract features (X) and outcome (Y) vector or matrix from data/time series in LSTM preformat}
\usage{
get.LSTM.XY(
  dataset,
  x = NULL,
  y = 2,
  other_columns = NULL,
  timesteps = 1,
  x.lag = 0,
  y.lag = 0,
  y.lag_type = c("none", "tsteps", "plain")
)
}
\arguments{
\item{dataset}{A data set, usually a matrix or data.frame.}

\item{x}{The column indices which spawn the feature matrix.}

\item{y}{The column indices of the outcomes.}

\item{other_columns}{The column indices of other columns which play an important role, e.g. a datetime column.}

\item{timesteps}{The number of timesteps, is the number of different periods within one sample of upcoming resampled feature matrix produced by \code{as.LSTM.X}.}

\item{x.lag}{The considered lags on feature side. For a univariate time series at least a lag of 1 is needed.
For a multivariate time series no lag must be necessarily used. This argument can also be a vector of the same length as \code{x}.
In that case, each feature can have its own specific lag.}

\item{y.lag}{The considered lag on outcome side. If outcomes (\code{y}) should be treated as features, a lag for the outcomes must be specified.}

\item{y.lag_type}{Indicates whether lagged y-variables are used as features in a multivariate time series.
\code{none} doesn't include lagged y-variables.
\code{tsteps} The lagged y-variables retrieve in the periods (timesteps) of the later resampled feature matrix thru \code{as.LSTM.X}.
  Therefore, only one lagged y-variable with a lag order of adjusted \code{lag} is chosen.
\code{plain} The number of included lagged y-variables is equal to the value of the \code{timesteps} argument.}
}
\value{
A list with named elements
  \code{X}: A feature matrix in LSTM compatible preformat for usage with \code{as.LSTM.X}.
  \code{Y}: An outcome matrix in LSTM compatible preformat for usage with \code{as.LSTM.Y}.
  \code{other_columns}: A vector, matrix or data.frame of the selected \code{other_columns}.
}
\description{
\code{get.LSTM.XY} extracts features and outcomes from a data/time series in a LSTM compatible preformat.
  Within a univariate time series, y(t) is explained by past y(t-1), y(t-2) etc. Therefore the last sample of the
  feature set must be deleted, because there is no Y-value for it. Resampling of the Y-values must start at timesteps + 1.
  That is different to a multivariate time series. For y(t), the corresponding features at time t are already given.
  Resampling must start at timesteps.
}
\seealso{
\code{\link{as.LSTM.X}}, \code{\link{as.LSTM.Y}}.

Other Recurrent Neural Network (RNN), Long Short-Term Memory (LSTM): 
\code{\link{as.LSTM.X}()},
\code{\link{as.LSTM.Y}()},
\code{\link{as.LSTM.data.frame}()},
\code{\link{as.LSTM.period_outcome}()},
\code{\link{as.lag}()},
\code{\link{as.timesteps}()},
\code{\link{build.LSTM}()},
\code{\link{fit.LSTM}()},
\code{\link{get.LSTM.X.samples}()},
\code{\link{get.LSTM.X.timesteps}()},
\code{\link{get.LSTM.X.units}()},
\code{\link{get.LSTM.Y.samples}()},
\code{\link{get.LSTM.Y.units}()},
\code{\link{get.LSTM.period_shift}()},
\code{\link{predict.LSTM}()},
\code{\link{start.LSTM.invert_differencing}()}
}
\concept{Recurrent Neural Network (RNN), Long Short-Term Memory (LSTM)}
