% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepRNN.r
\name{get.LSTM.Y.samples}
\alias{get.LSTM.Y.samples}
\title{Get number of output samples from outcome tensor}
\usage{
get.LSTM.Y.samples(Y.tensor)
}
\arguments{
\item{Y.tensor}{A tensor of the outcome produced by \code{as.LSTM.Y}.}
}
\value{
Number of output samples.
}
\description{
Get number of output samples from outcome tensor
}
\seealso{
\code{\link{as.LSTM.Y}}, \code{\link{get.LSTM.Y.units}},
  \code{\link{as.LSTM.X}}, \code{\link{get.LSTM.X.samples}}, \code{\link{get.LSTM.X.timesteps}}, \code{\link{get.LSTM.X.units}}.

Other Recurrent Neural Network (RNN): 
\code{\link{as.LSTM.X}()},
\code{\link{as.LSTM.Y}()},
\code{\link{as.LSTM.data.frame}()},
\code{\link{as.LSTM.period_outcome}()},
\code{\link{as.lag}()},
\code{\link{as.timesteps}()},
\code{\link{build.LSTM}()},
\code{\link{fit.LSTM}()},
\code{\link{get.LSTM.X.samples}()},
\code{\link{get.LSTM.X.timesteps}()},
\code{\link{get.LSTM.X.units}()},
\code{\link{get.LSTM.XY}()},
\code{\link{get.LSTM.Y.timesteps}()},
\code{\link{get.LSTM.Y.units}()},
\code{\link{get.period_shift}()},
\code{\link{predict.ANN}()},
\code{\link{start.invert_differencing}()}
}
\concept{Recurrent Neural Network (RNN)}
