% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepRNN.r
\name{get.LSTM.Y.timesteps}
\alias{get.LSTM.Y.timesteps}
\title{Get number of timesteps from outcome tensor if outcome is a sequence}
\usage{
get.LSTM.Y.timesteps(Y.tensor)
}
\arguments{
\item{Y.tensor}{A tensor of the outcome produced by \code{as.LSTM.Y}.}
}
\value{
Number of output timesteps.
}
\description{
Get number of timesteps from outcome tensor if outcome is a sequence
}
\seealso{
\code{\link{as.LSTM.Y}}, \code{\link{get.LSTM.Y.samples}},
  \code{\link{as.LSTM.X}}, \code{\link{get.LSTM.X.samples}}, \code{\link{get.LSTM.X.timesteps}}, \code{\link{get.LSTM.X.units}}.

Other Recurrent Neural Network (RNN), Long Short-Term Memory (LSTM): 
\code{\link{as.LSTM.X}()},
\code{\link{as.LSTM.Y}()},
\code{\link{as.LSTM.data.frame}()},
\code{\link{as.LSTM.period_outcome}()},
\code{\link{as.lag}()},
\code{\link{as.timesteps}()},
\code{\link{build.LSTM}()},
\code{\link{fit.LSTM}()},
\code{\link{get.LSTM.X.samples}()},
\code{\link{get.LSTM.X.timesteps}()},
\code{\link{get.LSTM.X.units}()},
\code{\link{get.LSTM.XY}()},
\code{\link{get.LSTM.Y.samples}()},
\code{\link{get.LSTM.Y.units}()},
\code{\link{get.LSTM.period_shift}()},
\code{\link{predict.LSTM}()},
\code{\link{start.LSTM.invert_differencing}()}
}
\concept{Recurrent Neural Network (RNN), Long Short-Term Memory (LSTM)}
