% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepRNN.r
\name{get.LSTM.Y.units}
\alias{get.LSTM.Y.units}
\title{Get number of output units from 2-dimensional outcome tensor}
\usage{
get.LSTM.Y.units(Y.tensor)
}
\arguments{
\item{Y.tensor}{A two-dimensional array of the outcome produced by \code{as.LSTM.Y}.}
}
\value{
Number of output units or outcomes.
}
\description{
Get number of output units from 2-dimensional outcome tensor
}
\seealso{
\code{\link{as.LSTM.Y}}, \code{\link{get.LSTM.Y.samples}}, \code{\link{get.MLP.X.units}}, \code{\link{get.MLP.X.samples}}, \code{\link{get.MLP.X.timesteps}}.

Other Recurrent Neural Network (RNN), Long Short-Term Memory (LSTM): 
\code{\link{as.LSTM.X}()},
\code{\link{as.LSTM.Y}()},
\code{\link{as.LSTM.data.frame}()},
\code{\link{as.LSTM.period_outcome}()},
\code{\link{as.lag}()},
\code{\link{as.timesteps}()},
\code{\link{build.LSTM}()},
\code{\link{fit.LSTM}()},
\code{\link{get.LSTM.X.samples}()},
\code{\link{get.LSTM.X.timesteps}()},
\code{\link{get.LSTM.X.units}()},
\code{\link{get.LSTM.XY}()},
\code{\link{get.LSTM.Y.samples}()},
\code{\link{get.LSTM.period_shift}()},
\code{\link{predict.LSTM}()},
\code{\link{start.LSTM.invert_differencing}()}
}
\concept{Recurrent Neural Network (RNN), Long Short-Term Memory (LSTM)}
