% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepMLP.r
\name{get.MLP.X.samples}
\alias{get.MLP.X.samples}
\title{Get number of input samples from feature tensor}
\usage{
get.MLP.X.samples(X.tensor)
}
\arguments{
\item{X.tensor}{A two-dimensional array of the feature matrix produced by \code{as.MLP.X}.}
}
\value{
Number of input samples.
}
\description{
Get number of input samples from feature tensor
}
\seealso{
\code{\link{as.MLP.X}}, \code{\link{get.MLP.X.units}}.

Other Single & Multi Layer Perceptron (SLP, MLP): 
\code{\link{as.ANN.matrix}()},
\code{\link{as.MLP.X}()},
\code{\link{as.MLP.Y}()},
\code{\link{as.tensor.1D}()},
\code{\link{as.tensor.2D}()},
\code{\link{as.tensor.3D}()},
\code{\link{as.tensor}()},
\code{\link{build.MLP}()},
\code{\link{fit.MLP}()},
\code{\link{get.MLP.X.units}()},
\code{\link{get.MLP.Y.samples}()},
\code{\link{get.MLP.Y.units}()},
\code{\link{predict.ANN}()},
\code{\link{vector.as.ANN.matrix}()},
\code{\link{vector.as.numeric}()}
}
\concept{Single & Multi Layer Perceptron (SLP, MLP)}
