% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepMLP.r
\name{get.MLP.Y.samples}
\alias{get.MLP.Y.samples}
\title{Get number of output samples from outcome tensor}
\usage{
get.MLP.Y.samples(Y.tensor)
}
\arguments{
\item{Y.tensor}{A tensor produced by \code{as.MLP.Y}.}
}
\value{
Number of output samples.
}
\description{
Get number of output samples from outcome tensor
}
\seealso{
\code{\link{as.MLP.Y}}, \code{\link{get.MLP.Y.units}}.

Other Single & Multi Layer Perceptron (SLP, MLP): 
\code{\link{as.ANN.matrix}()},
\code{\link{as.MLP.X}()},
\code{\link{as.MLP.Y}()},
\code{\link{as.numpy}()},
\code{\link{as.tensor.1D}()},
\code{\link{as.tensor.2D}()},
\code{\link{as.tensor.3D}()},
\code{\link{as.tensor}()},
\code{\link{build.MLP}()},
\code{\link{fit.MLP}()},
\code{\link{get.MLP.X.samples}()},
\code{\link{get.MLP.X.units}()},
\code{\link{get.MLP.Y.units}()},
\code{\link{load_weights.ANN}()},
\code{\link{predict.ANN}()},
\code{\link{save_weights.ANN}()},
\code{\link{vector.as.ANN.matrix}()},
\code{\link{vector.as.numeric}()}
}
\concept{Single & Multi Layer Perceptron (SLP, MLP)}
