% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepMLP.r
\name{get.MLP.Y.units}
\alias{get.MLP.Y.units}
\title{Get number of output units from 2-dimensional outcome tensor}
\usage{
get.MLP.Y.units(Y.tensor)
}
\arguments{
\item{Y.tensor}{A two-dimensional array of the outcome produced by \code{as.MLP.Y}.}
}
\value{
Number of output units or outcomes.
}
\description{
Get number of output units from 2-dimensional outcome tensor
}
\seealso{
\code{\link{as.MLP.Y}}, \code{\link{get.MLP.X.units}}.

Other Single & Multi Layer Perceptron (SLP, MLP): 
\code{\link{as.ANN.matrix}()},
\code{\link{as.MLP.X}()},
\code{\link{as.MLP.Y}()},
\code{\link{as.tensor.1D}()},
\code{\link{as.tensor.2D}()},
\code{\link{as.tensor.3D}()},
\code{\link{build.MLP}()},
\code{\link{fit.MLP}()},
\code{\link{get.MLP.X.units}()},
\code{\link{predict.ANN}()},
\code{\link{vector.as.ANN.matrix}()}
}
\concept{Single & Multi Layer Perceptron (SLP, MLP)}
