% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepTimeSeries.r
\name{get.season}
\alias{get.season}
\title{Get season from given dates}
\usage{
get.season(dates)
}
\arguments{
\item{dates}{A vector with date values.}
}
\value{
A vector with same length as \code{dates} with seasonal values \code{Winter,Spring,Summer,Fall}.
}
\description{
Get season from given dates
}
\references{
\url{https://stackoverflow.com/questions/9500114/find-which-season-a-particular-date-belongs-to}.
}
\seealso{
Other TimeSeries: 
\code{\link{build.stationary}()},
\code{\link{invert_differencing}()},
\code{\link{lags}()},
\code{\link{period}()}
}
\concept{TimeSeries}
