% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepRNN.r
\name{get_LSTM_XY}
\alias{get_LSTM_XY}
\title{Extract features (X) and outcome (Y) vector or matrix from data/time series in LSTM preformat}
\usage{
get_LSTM_XY(
  dataset,
  x = NULL,
  y = 2L,
  other_columns = NULL,
  timesteps = 1L,
  xlag = 0L,
  y_as_feature = c("none", "plain", "timesteps"),
  ylag = 0L
)
}
\arguments{
\item{dataset}{A data set, usually a matrix or data frame.}

\item{x}{The names or indices of the feature columns.}

\item{y}{The names or indices of the outcome columns.}

\item{other_columns}{The names or indices of further columns, e.g. a periodic column.}

\item{timesteps}{A number or vector of timesteps for \code{x} and \code{y}. A timestep denotes the number of different periods of the values within one sample.
A feature does always have at least one timestep, but an outcome is either a scalar with one implicit timestep or a sequence with at least two timesteps.
If only one value is given, this value is used for the resampled feature tensor produced by \code{as_LSTM_X}. In this case, \code{y} will
be treated as a scalar outcome. If two values are given, the first value is used as before and the second value is the number of timesteps for the resampled
sequence or multi-step outcome produced by \code{as_LSTM_Y}.}

\item{xlag}{The considered lags on feature side. For a univariate time series at least a lag of 1 is needed.
For a multivariate time series no lag must be necessarily used. This argument can also be a vector of the same length as \code{x}.
In that case, each feature can have its own specified lag.}

\item{y_as_feature}{Indicates whether lagged outcomes are used as features in a multivariate time series.\cr
\code{none} doesn't include lagged y-variables (default).\cr
\code{plain} The lagged y-variables retrieve in the periods (timesteps) of the later resampled feature matrix thru \code{as_LSTM_X}.
Therefore, only one lagged y-variable with a lag order of adjusted \code{y_as_feature_lag} is chosen.\cr
\code{timesteps} The number of included lagged y-variables is equal to the value of the \code{timesteps} argument.}

\item{ylag}{The considered lag for lagged outcomes (\code{y}) as further features.}
}
\value{
A list with named elements
\code{X}: A feature matrix in LSTM compatible preformat for usage with \code{as_LSTM_X}.\cr
\code{Y}: An outcome matrix in LSTM compatible preformat for usage with \code{as_LSTM_Y}.\cr
\code{other_columns}: A data frame of the selected \code{other_columns}.
}
\description{
\code{get_LSTM_XY} extracts features and outcomes from a data/time series in a Long Short-Term Memory (LSTM) compatible preformat.
Within a univariate time series, y(t) is explained by past y(t-1), y(t-2) etc. Therefore the last sample of the
feature set must be deleted, because there is no Y-value for it. Resampling of the Y-values must start at timesteps + 1.
That is different to a multivariate time series. For y(t), the corresponding features at time t are already given.
Resampling must start at timesteps.
}
\seealso{
\code{\link{as_LSTM_X}}, \code{\link{as_LSTM_Y}}.

Other Recurrent Neural Network (RNN): 
\code{\link{as_LSTM_X}()},
\code{\link{as_LSTM_Y}()},
\code{\link{as_LSTM_data_frame}()},
\code{\link{as_LSTM_period_outcome}()},
\code{\link{as_lag}()},
\code{\link{as_timesteps}()},
\code{\link{build_LSTM}()},
\code{\link{fit_LSTM}()},
\code{\link{get_period_shift}()},
\code{\link{load_weights_ANN}()},
\code{\link{predict_ANN}()},
\code{\link{save_weights_ANN}()},
\code{\link{start_invert_differencing}()}
}
\concept{Recurrent Neural Network (RNN)}
