% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepRNN.r
\name{get_period_shift}
\alias{get_period_shift}
\title{Period shift}
\usage{
get_period_shift(timesteps = 1L, lag = 0L, type = "univariate")
}
\arguments{
\item{timesteps}{The number of timesteps.}

\item{lag}{The number of considered lags on feature side.}

\item{type}{The type of time series: \code{univariate} or \code{multivariate}.}
}
\value{
The period shift.
}
\description{

}
\details{
The period shift denotes the number of past periods starting from a certain period t, whose values of X are needed
  to describe Y in period t and which cannot be used to extract Y values. This number of past periods depends on the type of
  timeseries (univariate or multivariate), the number of timesteps and the underpinned number of lags. In other words, the
  period shift is the number of periods to go backwards to get features (X) for outcomes (Y) or as a question: how many samples
  get lost respectively must be used as X for Y description/prediction?
}
\seealso{
Other Recurrent Neural Network (RNN): 
\code{\link{as_LSTM_X}()},
\code{\link{as_LSTM_Y}()},
\code{\link{as_LSTM_data_frame}()},
\code{\link{as_LSTM_period_outcome}()},
\code{\link{as_lag}()},
\code{\link{as_timesteps}()},
\code{\link{build_LSTM}()},
\code{\link{fit_LSTM}()},
\code{\link{get_LSTM_XY}()},
\code{\link{load_weights_ANN}()},
\code{\link{predict_ANN}()},
\code{\link{save_weights_ANN}()},
\code{\link{start_invert_differencing}()}
}
\concept{Recurrent Neural Network (RNN)}
