% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepTimeSeries.r
\name{get_season}
\alias{get_season}
\title{Get season from given dates}
\usage{
get_season(dates)
}
\arguments{
\item{dates}{A vector with date values.}
}
\value{
A vector with same length as \code{dates} with seasonal values \code{Winter, Spring, Summer, Fall}.
}
\description{
Get season from given dates
}
\references{
\url{https://stackoverflow.com/questions/9500114/find-which-season-a-particular-date-belongs-to}.
}
\seealso{
Other Time Series: 
\code{\link{diff_log}()},
\code{\link{diff_percentage}()},
\code{\link{diffinv_log}()},
\code{\link{diffinv_percentage}()},
\code{\link{diffinv_simple}()},
\code{\link{invert_differencing}()},
\code{\link{lags}()},
\code{\link{partition}()},
\code{\link{period}()},
\code{\link{stationary}()}
}
\concept{Time Series}
