% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepQuality.r
\name{gini_impurity}
\alias{gini_impurity}
\title{Gini impurity}
\usage{
gini_impurity(x)
}
\arguments{
\item{x}{A vector of values, usually character labels as raw instances or as class frequencies.}
}
\value{
The Gini impurity.
}
\description{

}
\details{
Gini impurity is the probability of how often a randomly chosen element from a set \code{x} would be
  incorrectly labeled if it was randomly labeled according to the distribution of labels in the set. So, impurity is
  the probability of being incorrect if a label is randomly assigned to a sample of \code{x}.
}
\examples{
  gini_impurity(c("dog", "dog", "cat", "mouse"))
  gini_impurity(c(dog = 2, cat = 1, mouse = 1))
}
\references{
\url{https://victorzhou.com/blog/gini-impurity/}
}
\seealso{
Other Quality: 
\code{\link{accuracy}()},
\code{\link{cross_entropy}()},
\code{\link{entropy}()},
\code{\link{erfcinv}()},
\code{\link{erfc}()},
\code{\link{erfinv}()},
\code{\link{erf}()},
\code{\link{huber_loss}()},
\code{\link{log_cosh_loss}()},
\code{\link{mae}()},
\code{\link{mape}()},
\code{\link{mse}()},
\code{\link{msle}()},
\code{\link{quantile_loss}()},
\code{\link{rmse}()},
\code{\link{rmsle}()},
\code{\link{rmspe}()},
\code{\link{sse}()},
\code{\link{vc}()},
\code{\link{wape}()},
\code{\link{wmape}()}
}
\concept{Quality}
