% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepMisc.r
\name{\%null\%}
\alias{\%null\%}
\title{NULL assignment operator}
\usage{
x \%null\% y
}
\arguments{
\item{x}{object that is checked for \code{NULL}.}

\item{y}{result if \code{x} is \code{NULL}.}
}
\value{
Either \code{x} or \code{y}.
}
\description{
NULL assignment operator
}
\details{
Checking for \code{NULL} is very common practice in R. This operator shortens the handling of such checks and the return of corresponding values.
}
\examples{
x <- marray(1:24, dim = c(4, 3, 2))
x <- marray(1:24)
x <- 1:24
# The dim() result for the first x is 4 3 2, for the second x 24 and for the third x NULL.
# To get a corresponding behavior between a one-dimensional array and a simple vector you can write:
dim(x) \%null\% length(x)
}
