% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepQuality.r
\name{huber_loss}
\alias{huber_loss}
\title{Huber loss}
\usage{
huber_loss(y, yhat, delta = 1)
}
\arguments{
\item{y}{A numeric vector with actual values (to-be values).}

\item{yhat}{A numeric vector with estimated values (as-is values).}

\item{delta}{A parameter that shows the error difference and controls the calculation.}
}
\value{
Huber loss.
}
\description{
Huber loss
}
\details{
Huber loss is less sensitive to outliers than MSE.
}
\references{
Huber, Peter J. (1964): Robust Estimation of a Location Parameter. In: Annals of Mathematical Statistics, 35 (1964) 1, 73-101.
  Hasti, Trevor; Tibshirani, Robert; Friedman, Jerome (2009): The Elements of Statistical Learning. 2nd ed., 2009. New York: Springer. (p. 349).
}
\seealso{
Other Quality: 
\code{\link{accuracy}()},
\code{\link{log_cosh_loss}()},
\code{\link{mae}()},
\code{\link{mape}()},
\code{\link{mse}()},
\code{\link{quantile_loss}()},
\code{\link{rmse}()},
\code{\link{vc}()}
}
\concept{Quality}
