% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepCNN.r
\name{images_load}
\alias{images_load}
\title{Load images from different sources like from files or web}
\usage{
images_load(images, FUN, ...)
}
\arguments{
\item{images}{A vector or list of files, urls etc. containing images.}

\item{FUN}{The function to be applied for loading \code{images}. If no function is specified \code{image_load()} from keras is called.}

\item{...}{Optional arguments to \code{FUN}.}
}
\value{
A list of images.
}
\description{
Load images from different sources like from files or web
}
\examples{
  For an example see \code{\link{as_images_tensor4D}}.
}
\seealso{
\code{\link[base]{list.files}}, \code{\link[keras]{image_load}}.

Other Convolutional Neural Network (CNN): 
\code{\link{as.CNN.image.X}()},
\code{\link{as.CNN.image.Y}()},
\code{\link{as.CNN.temp.X}()},
\code{\link{as.CNN.temp.Y}()},
\code{\link{as_images_array}()},
\code{\link{as_images_tensor4D}()},
\code{\link{get.CNN.image.X.channels}()},
\code{\link{get.CNN.image.X.height}()},
\code{\link{get.CNN.image.X.samples}()},
\code{\link{get.CNN.image.X.width}()},
\code{\link{get.CNN.image.Y.samples}()},
\code{\link{get.CNN.image.Y.units}()},
\code{\link{get.CNN.temp.X.samples}()},
\code{\link{get.CNN.temp.X.subsequences}()},
\code{\link{get.CNN.temp.X.timesteps}()},
\code{\link{get.CNN.temp.X.units}()},
\code{\link{get.CNN.temp.Y.samples}()},
\code{\link{get.CNN.temp.Y.timesteps}()},
\code{\link{get.CNN.temp.Y.units}()},
\code{\link{images_resize}()}
}
\concept{Convolutional Neural Network (CNN)}
