% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepCNN.r
\name{images_resize}
\alias{images_resize}
\title{Resize loaded images}
\usage{
images_resize(imagelist, FUN, ...)
}
\arguments{
\item{imagelist}{A list of loaded images returned by \code{images_load()}.}

\item{FUN}{The function to be applied for resizing images within \code{imagelist}. If no function is specified the images within \code{imagelist} aren't resized.}

\item{...}{Optional arguments to \code{FUN}.}
}
\value{
A list of (resized) images.
}
\description{
Resize loaded images
}
\examples{
  For an example see \code{\link{as_images_tensor4D}}.
}
\seealso{
Other Convolutional Neural Network (CNN): 
\code{\link{as.CNN.image.X}()},
\code{\link{as.CNN.image.Y}()},
\code{\link{as.CNN.temp.X}()},
\code{\link{as.CNN.temp.Y}()},
\code{\link{as_images_array}()},
\code{\link{as_images_tensor4D}()},
\code{\link{get.CNN.image.X.channels}()},
\code{\link{get.CNN.image.X.height}()},
\code{\link{get.CNN.image.X.samples}()},
\code{\link{get.CNN.image.X.width}()},
\code{\link{get.CNN.image.Y.samples}()},
\code{\link{get.CNN.image.Y.units}()},
\code{\link{get.CNN.temp.X.samples}()},
\code{\link{get.CNN.temp.X.subsequences}()},
\code{\link{get.CNN.temp.X.timesteps}()},
\code{\link{get.CNN.temp.X.units}()},
\code{\link{get.CNN.temp.Y.samples}()},
\code{\link{get.CNN.temp.Y.timesteps}()},
\code{\link{get.CNN.temp.Y.units}()},
\code{\link{images_load}()}
}
\concept{Convolutional Neural Network (CNN)}
