% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepArray.r
\name{insert}
\alias{insert}
\title{Array insertion}
\usage{
insert(a, ..., axis = -1L, order = c("C", "F"))
}
\arguments{
\item{a}{An array.}

\item{...}{Any number of objects inserted into or appended to \code{a}.}

\item{axis}{The axis along which to insert the objects.}

\item{order}{The order according to which the respective elements of the objects are read during insertion or appending.
By default, the order is equivalent to the \code{C}-style ordering and means elements should be read in row-major order.
In opposite, the \code{Fortran}-style ordering means elements should be read in column-major order.}
}
\value{
The array \code{a} with objects inserted or appended.
}
\description{
Insert an object into an array.
}
\details{
This function corresponds partially to \code{insert()} and \code{append()} from NumPy.
}
\examples{
# original array
a <- array(seq.int(2 * 3 * 4), dim = c(2, 3, 4))
# slice to be added to the second axis
b <- array(100L + seq.int(2 * 1 * 4), dim = c(2, 1, 4))
insert(a, b, axis = 2L)

}
\seealso{
Other Array: 
\code{\link{DIM}()},
\code{\link{column_stack}()},
\code{\link{delete}()},
\code{\link{dimC<-}()},
\code{\link{embedseries}()},
\code{\link{empty}()},
\code{\link{expand_dims}()},
\code{\link{eye}()},
\code{\link{flatten}()},
\code{\link{flip}()},
\code{\link{full}()},
\code{\link{mabind}()},
\code{\link{mamatrix}()},
\code{\link{marray}()},
\code{\link{ndim}()},
\code{\link{nsize}()},
\code{\link{ones}()},
\code{\link{rearrange}()},
\code{\link{reshape.array}()},
\code{\link{rot90}()},
\code{\link{row_stack}()},
\code{\link{slice}()},
\code{\link{squeeze}()},
\code{\link{transpose}()},
\code{\link{vander}()},
\code{\link{zeros}()}
}
\concept{Array}
