% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepTimeSeries.r
\name{invert_differencing}
\alias{invert_differencing}
\title{Invert a differenced data series}
\usage{
invert_differencing(delta, origin, type = c("simple", "log", "percentage"))
}
\arguments{
\item{delta}{A differenced numeric vector.}

\item{origin}{A scalar or numeric vector with original value(s) to invert \code{delta}.}

\item{type}{The type of differencing to be used. Available types are \code{simple}, \code{log} and \code{percentage}.}
}
\value{
The inverted \code{delta}.
}
\description{
Invert a differenced data series
}
\seealso{
\code{\link{build.stationary}}.

Other Time Series: 
\code{\link{build.stationary}()},
\code{\link{diff.log}()},
\code{\link{diff.percentage}()},
\code{\link{diffinv.log}()},
\code{\link{diffinv.percentage}()},
\code{\link{diffinv.simple}()},
\code{\link{get.season}()},
\code{\link{lags}()},
\code{\link{period}()}
}
\concept{Time Series}
