% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepMetrics.r
\name{iou}
\alias{iou}
\title{Intersection-over-Union (IoU, Jaccard Index)}
\usage{
iou(actuals, preds, smooth = 1)
}
\arguments{
\item{actuals}{A multidimensional array of actual values.}

\item{preds}{A multidimensional array of prediction values.}

\item{smooth}{A smoothing factor to avoid division by zero.}
}
\value{
Intersection-over-Union (IoU, Jaccard Index).
}
\description{
Intersection-over-Union (IoU, Jaccard Index)
}
\details{
This metric is used for evaluation of the results within image segmentation. \code{actuals} as well as \code{preds} are binary encoded
image data masks in form of a n-dimensional array, mainly a two-dimensional array with the dimensions height and width for every channel. A value of
\code{1} indicates the background (e.g. white color), a value equal \code{0} indicates the object (e.g. black color).
}
\seealso{
Other Metrics: 
\code{\link{accuracy}()},
\code{\link{cross_entropy}()},
\code{\link{dice}()},
\code{\link{entropy}()},
\code{\link{erfcinv}()},
\code{\link{erfc}()},
\code{\link{erfinv}()},
\code{\link{erf}()},
\code{\link{gini_impurity}()},
\code{\link{huber_loss}()},
\code{\link{log_cosh_loss}()},
\code{\link{mae}()},
\code{\link{mape}()},
\code{\link{mse}()},
\code{\link{msle}()},
\code{\link{quantile_loss}()},
\code{\link{rmse}()},
\code{\link{rmsle}()},
\code{\link{rmspe}()},
\code{\link{sse}()},
\code{\link{stderror}()},
\code{\link{vc}()},
\code{\link{wape}()},
\code{\link{wmape}()}
}
\concept{Metrics}
