% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepEncoder.r
\docType{class}
\name{LabelBinarizer}
\alias{LabelBinarizer}
\title{class LabelBinarizer}
\description{
A class for encoding a categorical variable as a label indicator matrix.
}
\examples{
x <- factor(c("small", "small", "medium", "large", "large", "medium", "small"), levels = c("small", "medium", "large"))
lb <- LabelBinarizer$new()
xt <- lb$fit_transform(x)
lb$inverse_transform(xt)
}
\section{Super classes}{
\code{\link[deepANN:Encoder]{deepANN::Encoder}} -> \code{\link[deepANN:OneHotEncoder]{deepANN::OneHotEncoder}} -> \code{LabelBinarizer}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LabelBinarizer-clone}{\code{LabelBinarizer$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="deepANN" data-topic="OneHotEncoder" data-id="fit"><a href='../../deepANN/html/OneHotEncoder.html#method-OneHotEncoder-fit'><code>deepANN::OneHotEncoder$fit()</code></a></span></li>
<li><span class="pkg-link" data-pkg="deepANN" data-topic="OneHotEncoder" data-id="fit_transform"><a href='../../deepANN/html/OneHotEncoder.html#method-OneHotEncoder-fit_transform'><code>deepANN::OneHotEncoder$fit_transform()</code></a></span></li>
<li><span class="pkg-link" data-pkg="deepANN" data-topic="OneHotEncoder" data-id="initialize"><a href='../../deepANN/html/OneHotEncoder.html#method-OneHotEncoder-initialize'><code>deepANN::OneHotEncoder$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="deepANN" data-topic="OneHotEncoder" data-id="inverse_transform"><a href='../../deepANN/html/OneHotEncoder.html#method-OneHotEncoder-inverse_transform'><code>deepANN::OneHotEncoder$inverse_transform()</code></a></span></li>
<li><span class="pkg-link" data-pkg="deepANN" data-topic="OneHotEncoder" data-id="transform"><a href='../../deepANN/html/OneHotEncoder.html#method-OneHotEncoder-transform'><code>deepANN::OneHotEncoder$transform()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LabelBinarizer-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LabelBinarizer-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LabelBinarizer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
