% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepEncoder.r
\docType{class}
\name{LabelEncoder}
\alias{LabelEncoder}
\title{class LabelEncoder}
\description{
A class for encoding labels with values between 0 and n_labels-1.
}
\examples{
x <- factor(c("small", "small", "medium", "large", "large", "medium", "small"), levels = c("small", "medium", "large"))
le <- LabelEncoder$new()
xt <- le$fit_transform(x)
le$inverse_transform(xt)
}
\section{Super class}{
\code{\link[deepANN:Encoder]{deepANN::Encoder}} -> \code{LabelEncoder}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{categories}}{('NULL' | character vector)\cr
Holds the labels for the input.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LabelEncoder-new}{\code{LabelEncoder$new()}}
\item \href{#method-LabelEncoder-fit}{\code{LabelEncoder$fit()}}
\item \href{#method-LabelEncoder-transform}{\code{LabelEncoder$transform()}}
\item \href{#method-LabelEncoder-fit_transform}{\code{LabelEncoder$fit_transform()}}
\item \href{#method-LabelEncoder-inverse_transform}{\code{LabelEncoder$inverse_transform()}}
\item \href{#method-LabelEncoder-clone}{\code{LabelEncoder$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LabelEncoder-new"></a>}}
\if{latex}{\out{\hypertarget{method-LabelEncoder-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LabelEncoder$new()}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{feature_range}}{Desired range of transformed data.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LabelEncoder-fit"></a>}}
\if{latex}{\out{\hypertarget{method-LabelEncoder-fit}{}}}
\subsection{Method \code{fit()}}{
Extract labels from input.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LabelEncoder$fit(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{The input, usually a \code{\link{factor}}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The instance \code{self} itself.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LabelEncoder-transform"></a>}}
\if{latex}{\out{\hypertarget{method-LabelEncoder-transform}{}}}
\subsection{Method \code{transform()}}{
Encode input to numeric values between 0 and n_labels-1.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LabelEncoder$transform(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{The input that will be encoded.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The encoded \code{x}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LabelEncoder-fit_transform"></a>}}
\if{latex}{\out{\hypertarget{method-LabelEncoder-fit_transform}{}}}
\subsection{Method \code{fit_transform()}}{
Run \code{fit()} and \code{transform()}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LabelEncoder$fit_transform(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{The input that will be encoded.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The encoded \code{x}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LabelEncoder-inverse_transform"></a>}}
\if{latex}{\out{\hypertarget{method-LabelEncoder-inverse_transform}{}}}
\subsection{Method \code{inverse_transform()}}{
Transform labels back to original encoding.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LabelEncoder$inverse_transform(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{The data that will be transformed back.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The original encoded \code{x}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LabelEncoder-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LabelEncoder-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LabelEncoder$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
