% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepTimeSeries.r
\name{lags}
\alias{lags}
\title{Build a lagged data set or series}
\usage{
lags(x, k = 1, between = FALSE, na = NA)
}
\arguments{
\item{x}{A vector for building a lagged data series or an entire data set.}

\item{k}{The number of lags. A lag denotes a delay in time or a period shift. A lagged variable, also known as delayed variable, holds the values of past or future periodic values of the variable itself, depending on the direction of the shift.
The sign of \code{k} specifies the direction of the shift. A positive lag shifts sequence values forward in time, so into the future. A negative lag shifts sequence values back in time, so into the past.}

\item{between}{Controls whether only the desired \code{lag} is considered (\code{FALSE}) or all other lags between 1 to \code{lag} (\code{TRUE}).}

\item{na}{A sign which is used to indicate a missing value in a shifted period.}
}
\value{
A lagged data series.
}
\description{
\code{lags} builds a lagged data series or entire data set.
}
\seealso{
Other Time Series: 
\code{\link{diff_log}()},
\code{\link{diff_percentage}()},
\code{\link{diffinv_log}()},
\code{\link{diffinv_percentage}()},
\code{\link{diffinv_simple}()},
\code{\link{get_season}()},
\code{\link{invert_differencing}()},
\code{\link{partition}()},
\code{\link{period}()},
\code{\link{stationary}()}
}
\concept{Time Series}
