% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepCNN.r
\name{load.image_features}
\alias{load.image_features}
\title{Load images from files and create a 4-dimensional image array}
\usage{
load.image_features(filelist, width, height, channels = 3)
}
\arguments{
\item{filelist}{A list of file names, e.g. returned by \code{list.files}.}

\item{width}{The width of an image, identical with the number of columns.}

\item{height}{The height of an image, identical with the number of rows.}

\item{channels}{The channel of an image. A color channel is a primary color (like red, green and blue), 
equal to a color valence (denotes how light effects the color sensation of an eye or in common of the brain).
Primary colors can be mixed to produce any color. 
A channel equal \code{1} indicates a grayscaled image, \code{3} a colored image.}
}
\value{
A 4-dimensional array with the dimensions samples (number of images), height, width and channel.
}
\description{
Load images from files and create a 4-dimensional image array
}
\seealso{
\code{\link[base]{list.files}},
  \code{\link[keras]{image_load}}, \code{\link[keras]{image_to_array}}, \code{\link[reticulate]{array_reshape}}.
}
\concept{Convolutional Neural Network (CNN)}
