% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepQuality.r
\name{log_cosh}
\alias{log_cosh}
\title{Log-Cosh loss}
\usage{
log_cosh(actual, predicted)
}
\arguments{
\item{actual}{A numeric vector with to-be values.}

\item{predicted}{A numeric vector with as-is values.}
}
\value{
Log-Cosh loss.
}
\description{
Log-Cosh loss
}
\seealso{
Other Quality: 
\code{\link{accuracy}()},
\code{\link{huber}()},
\code{\link{mae}()},
\code{\link{mape}()},
\code{\link{mse}()},
\code{\link{rmse}()},
\code{\link{vc}()}
}
\concept{Quality}
