% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepQuality.r
\name{log_cosh_loss}
\alias{log_cosh_loss}
\title{Log-Cosh loss}
\usage{
log_cosh_loss(y, yhat)
}
\arguments{
\item{y}{A numeric vector with actual values (to-be values).}

\item{yhat}{A numeric vector with estimated values (as-is values).}
}
\value{
Log-Cosh loss.
}
\description{
Log-Cosh loss
}
\seealso{
Other Quality: 
\code{\link{accuracy}()},
\code{\link{huber_loss}()},
\code{\link{mae}()},
\code{\link{mape}()},
\code{\link{mse}()},
\code{\link{quantile_loss}()},
\code{\link{rmse}()},
\code{\link{vc}()}
}
\concept{Quality}
