% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepArray.r
\name{mabind}
\alias{mabind}
\title{Array binding}
\usage{
mabind(..., input_shape = NULL, axis = -1, order = c("C", "F"))
}
\arguments{
\item{...}{Any number of objects that are combined into an array. The objects can be packed into a \code{list}. The dimensions of these objects must be equal, excluding axis, if they are not to be coerced into a certain \code{input_shape}.}

\item{input_shape}{The dimension the input objects are to be coerced. By default \code{NULL}, the original dimensions are used.}

\item{axis}{The dimension along the objects are combined. By default (\code{-1}), the last dimension is used for binding the arrays.}

\item{order}{The order in which elements of the objects should be read during coercing to arrays.
By default, the order is equivalent to the \code{C}-style ordering and means elements should be read in row-major order.
In opposite, the \code{Fortran}-style ordering means elements should be read in column-major order.}
}
\value{
An array as a combination of all input arrays along a specified dimension.
}
\description{
Combine arrays along a specified dimension.
}
\details{
This function corresponds to \code{concatenate()} from NumPy.
}
\examples{
a1 <- marray(1:24, dim = c(4, 3, 2), order = "F"); a1
a2 <- marray(-1:-24, dim = c(4, 3, 2), order = "F"); a2
a3 <- marray(sample(24), dim = c(4, 3, 2), order = "F"); a3
mabind(a1, a2, a3) # output is an 4x3x6 array
mabind(a1, a2, a3, axis = 1) # output is an 12x3x2 array
mabind(a1, a2, a3, axis = 2) # output is an 4x9x2 array
}
\seealso{
Other Array: 
\code{\link{DIM}()},
\code{\link{column_stack}()},
\code{\link{delete}()},
\code{\link{dimC<-}()},
\code{\link{embedseries}()},
\code{\link{empty}()},
\code{\link{expand_dims}()},
\code{\link{eye}()},
\code{\link{flatten}()},
\code{\link{flip}()},
\code{\link{full}()},
\code{\link{insert}()},
\code{\link{mamatrix}()},
\code{\link{marray}()},
\code{\link{ndim}()},
\code{\link{nsize}()},
\code{\link{ones}()},
\code{\link{rearrange}()},
\code{\link{reshape.array}()},
\code{\link{rot90}()},
\code{\link{row_stack}()},
\code{\link{slice}()},
\code{\link{squeeze}()},
\code{\link{transpose}()},
\code{\link{vander}()},
\code{\link{zeros}()}
}
\concept{Array}
