% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepUtils.r
\name{mbind}
\alias{mbind}
\alias{mbind.array}
\alias{mbind.marray}
\title{Combine matrices of a multidimensional array}
\usage{
mbind(x, ...)

\method{mbind}{array}(x, order = c("C", "F"))

\method{mbind}{marray}(x, order = c("C", "F"))
}
\arguments{
\item{x}{A multidimensional array.}

\item{...}{Additional arguments to be passed to the method.}

\item{order}{The order in which elements of data should be read during combination.
By default, the order is equivalent to the \code{C}-style ordering and means elements should be read in row-major order.
In opposite, the \code{Fortran}-style ordering means elements should be read in column-major order.}
}
\value{
A two-dimensional array with combined matrices of the first two dimensions of \code{x}.
}
\description{
This function combines the respective first two dimensions of a multidimensional array by columns or rows.
}
\seealso{
Other Utils: 
\code{\link{as_ANN_matrix}()},
\code{\link{as_tensor_1D}()},
\code{\link{as_tensor_2D}()},
\code{\link{as_tensor_3D}()},
\code{\link{degree}()},
\code{\link{distance}()},
\code{\link{flatten}()},
\code{\link{insert}()},
\code{\link{list_as_numeric}()},
\code{\link{marray}()},
\code{\link{ndim}()},
\code{\link{nsamples}()},
\code{\link{nsubsequences}()},
\code{\link{ntimesteps}()},
\code{\link{nunits}()},
\code{\link{probability}()},
\code{\link{radian}()},
\code{\link{random_seed}()},
\code{\link{re.factor}()},
\code{\link{sd_pop}()},
\code{\link{similarity}()},
\code{\link{ta}()},
\code{\link{tensor}()},
\code{\link{var_pop}()},
\code{\link{vector_as_ANN_matrix}()},
\code{\link{vector_as_numeric}()}
}
\concept{Utils}
