% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepML.r
\name{moving_average}
\alias{moving_average}
\title{Weighted moving average}
\usage{
moving_average(x, n = 3L, weights = NULL)
}
\arguments{
\item{x}{A numeric vector.}

\item{n}{The order of the moving average.}

\item{weights}{Optional weights.}
}
\value{
A vector with the (weighted) moving average.
}
\description{
Weighted moving average
}
\examples{
  x <- c(855, 847, 1000, 635, 346, 2146, 1328, 1322, 3124, 1012, 1280, 2435, 1016, 3465, 1107, 1172, 3432, 836, 142, 345, 2603, 739, 716, 880, 1008, 112, 361)
  moving_average(x)
  moving_average(x, weights = c(1L, 2L, 3L))
}
\seealso{
Other Machine Learning: 
\code{\link{cross_validation_split}()},
\code{\link{erfcinv}()},
\code{\link{erfc}()},
\code{\link{erfinv}()},
\code{\link{erf}()},
\code{\link{euclidean_distance}()},
\code{\link{k_nearest_neighbors}()},
\code{\link{naive_forecast}()},
\code{\link{predict.kmeans}()}
}
\concept{Machine Learning}
