% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepQuality.r
\name{mse}
\alias{mse}
\title{Mean squared error}
\usage{
mse(actual, predicted)
}
\arguments{
\item{actual}{A vector with to-be values.}

\item{predicted}{A vector with as-is values.}
}
\value{
Mean squared error (MSE).
}
\description{
Mean squared error
}
\seealso{
\code{\link{mae}} for mean absolute error, \code{\link{mape}} for mean absolute percentage error,
  \code{\link{rmse}} for root mean square error, \code{\link{vc}} for variance coefficient.
}
\concept{Quality

\code{mse} calculates the mean squared error.}
