% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepQuality.r
\name{msle}
\alias{msle}
\title{Mean squared logarithmic error (MSLE)}
\usage{
msle(actuals, preds, alpha = 1, na.rm = FALSE)
}
\arguments{
\item{actuals}{A numeric vector of actual values.}

\item{preds}{A numeric vector of prediction values.}

\item{alpha}{A numeric value (default \code{1}) to prevent taking a negative or zero log.}

\item{na.rm}{A logical value indicating whether actual and prediction pairs with at least one NA value should be ignored.}
}
\value{
Mean squared logarithmic error.
}
\description{
Mean squared logarithmic error (MSLE)
}
\seealso{
Other Quality: 
\code{\link{accuracy}()},
\code{\link{entropy}()},
\code{\link{gini_impurity}()},
\code{\link{huber_loss}()},
\code{\link{log_cosh_loss}()},
\code{\link{mae}()},
\code{\link{mape}()},
\code{\link{mse}()},
\code{\link{quantile_loss}()},
\code{\link{rmse}()},
\code{\link{rmsle}()},
\code{\link{vc}()}
}
\concept{Quality}
