% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepML.r
\name{naive_forecast}
\alias{naive_forecast}
\title{Naive forecasting methods}
\usage{
naive_forecast(x, drift = 0, na = NA)
}
\arguments{
\item{x}{A vector with numbers.}

\item{drift}{The number of periods used to calculate the change over time (it's called a drift).
If drift is more than 1 the mean value of the changes over time is used; default \code{0}.}

\item{na}{A value that indicates that no value is given; default \code{NA}.}
}
\value{
A series of naive predicted values for a given vector.
}
\description{
\code{naive_forecast} offers three naive forecast approaches: plain random walk and random walk with drifts.
}
\seealso{
Other Machine Learning: 
\code{\link{cross_validation_split}()},
\code{\link{distance}()},
\code{\link{erfcinv}()},
\code{\link{erfc}()},
\code{\link{erfinv}()},
\code{\link{erf}()},
\code{\link{k_nearest_neighbors}()},
\code{\link{moving_average}()},
\code{\link{predict.kmeans}()},
\code{\link{similarity}()}
}
\concept{Machine Learning}
