% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepScaling.r
\name{normalize}
\alias{normalize}
\title{Min-Max-Scaling}
\usage{
normalize(x, minx = NULL, maxx = NULL)
}
\arguments{
\item{x}{A numeric vector whose elements should be normalized thru min-max-scaling.}

\item{minx}{A given minimum value for scaling. If minimum value is \code{NULL} then the minimum of \code{x} is used.}

\item{maxx}{A given maximum value for scaling. If maximum value is \code{NULL} then the maximum of \code{x} is used.}
}
\value{
A vector of same length as \code{x} with min-max-scaled values; values are between 0 and 1.
}
\description{
Min-Max-Scaling
}
\seealso{
\code{\link{denormalize}}.

Other Scaling: 
\code{\link{denormalize}()}
}
\concept{Scaling}
