% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepScaling.r
\name{normalize_data}
\alias{normalize_data}
\title{Normalization of training and test data set}
\usage{
normalize_data(trainset, testset)
}
\arguments{
\item{trainset}{A training data set.}

\item{testset}{A test data set.}
}
\value{
A named list with following labels
  \code{min}: vector with minima of each column of \code{trainset},
  \code{max}: vector with maxima of each column of \code{trainset},
  \code{train}: min-max-scaled \code{trainset},
  \code{test}: min-max-scaled \code{testset}.
}
\description{
Normalization of training and test data set
}
\seealso{
\code{\link{normalize}}, \code{\link{denormalize}}.
}
\concept{Scaling

\code{normalize_data} scales a training and a test data set whereby the test data are normalized with train data scales (min,max).}
