% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepMLP.r
\name{nsubsequences}
\alias{nsubsequences}
\title{Number of subsequences within an array}
\usage{
nsubsequences(a)
}
\arguments{
\item{a}{An array.}
}
\value{
Number of subsequences.
}
\description{
Number of subsequences within an array
}
\details{
The number of subsequences is stored in the second dimension of a four-dimensional \code{a}, usually used for a temporal CNN.
}
\seealso{
Other Single & Multi Layer Perceptron (SLP, MLP): 
\code{\link{as_MLP_X}()},
\code{\link{as_MLP_Y}()},
\code{\link{as_tensor_1d}()},
\code{\link{as_tensor_2d}()},
\code{\link{as_tensor_3d}()},
\code{\link{build_MLP}()},
\code{\link{fit_MLP}()},
\code{\link{load_weights_ANN}()},
\code{\link{nsamples}()},
\code{\link{ntimesteps}()},
\code{\link{nunits}()},
\code{\link{predict_ANN}()},
\code{\link{save_weights_ANN}()}
}
\concept{Single & Multi Layer Perceptron (SLP, MLP)}
