% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepDummy.r
\name{one_hot_encode}
\alias{one_hot_encode}
\title{One-hot encoding}
\usage{
one_hot_encode(x)
}
\arguments{
\item{x}{A vector with categorical values, so-called levels, of a non-metric variable.}
}
\value{
A matrix with all levels as columns with either 0 or 1 values.
}
\description{
\code{one_hot_encode} rebuilds a categorical variable to a so-called 'one-hot vector';
  within a sample (row) of a one-hot-vector each level of the variable is rebuild in the form \code{(0|1,0|1,0|1,...)}.
}
\seealso{
\code{\link{one_hot_decode}}.

Other Dummyfication: 
\code{\link{append_rows}()},
\code{\link{dummify}()},
\code{\link{effectcoding}()},
\code{\link{one_hot_decode}()},
\code{\link{remove_columns}()},
\code{\link{resample.imbalanced}()}
}
\concept{Dummyfication}
