% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepOutlier.r
\name{outlier.dataset}
\alias{outlier.dataset}
\title{Replace outliers in columns of a data set with \code{NA}}
\usage{
outlier.dataset(dataset, columns = NULL, type = c("tukey", "ml"), ...)
}
\arguments{
\item{dataset}{A data set, usually a data frame.}

\item{columns}{The names of the columns whose outlier values are to be replaced. if \code{NULL} (default), all corresponding columns are examined.}

\item{type}{The type of outlier definition and detection. 
\code{tukey} refers to the method of Tukey (1977).
\code{ml} denotes maximum likelihood estimation.}

\item{...}{Dependent on \code{type}. 
For \code{tukey} the constant \code{k} can be specified, otherwise it's value is \code{1.5}. 
For \code{ml} the constant \code{k} can be specified, otherwise it's value is \code{2}.}
}
\value{
The \code{dataset} with replaced outliers.
}
\description{
Replace outliers in columns of a data set with \code{NA}
}
\seealso{
\code{\link{outlier}}.

Other Outlier: 
\code{\link{outlier}()},
\code{\link{winsorize}()}
}
\concept{Outlier}
