% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepOutlier.r
\name{outlier}
\alias{outlier}
\title{Definition and detection of outliers}
\usage{
outlier(x, type = c("tukey", "mle"), na.replace = FALSE, ...)
}
\arguments{
\item{x}{A numeric vector.}

\item{type}{The type of outlier definition and detection. 
\code{tukey} refers to the method of Tukey (1977).
\code{mle} denotes maximum likelihood estimation.}

\item{na.replace}{Boolean that indicates whether outliers should be replaced by \code{NA}.}

\item{...}{Depends on the type value. 
For \code{tukey} the constant \code{k} can be specified, otherwise it's value is \code{1.5}. 
For \code{mle} the constant \code{k} can be specified, otherwise it's value is \code{2}.}
}
\value{
A named list of lower and upper boundaries and values, and of \code{x} with replaced outliers where appropriate.
}
\description{
Definition and detection of outliers
}
\references{
Tukey, John W. (1977): Exploratory Data Analysis. 1977. Reading: Addison-Wesley.
}
\seealso{
\code{\link[stats]{quantile}}, \code{\link[stats]{IQR}}, \code{\link{winsorize}}.

Other Outlier: 
\code{\link{winsorize}()}
}
\concept{Outlier}
