% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepOutlier.r
\name{outlier_dataset}
\alias{outlier_dataset}
\title{Replace outliers in columns of a data set with \code{NA}}
\usage{
outlier_dataset(
  dataset,
  columns = NULL,
  type = c("iqr", "mean", "median"),
  ...
)
}
\arguments{
\item{dataset}{A data set, usually a data frame.}

\item{columns}{The names or indices of the columns whose outlier values are to be replaced; if \code{NULL} (default), all corresponding columns are examined.}

\item{type}{The type of outlier definition and detection.
\code{iqr} refers to the method of Tukey (1977).
\code{mean} denotes maximum likelihood estimation.
\code{median} denotes scaled median absolute deviation.}

\item{...}{Dependent on \code{type}.
For \code{iqr} the constant \code{k} can be specified, otherwise it's value is \code{1.5}.
For \code{mean} the constant \code{k} can be specified, otherwise it's value is \code{3}.
For \code{median} the constant \code{k} can be specified, otherwise it's value is \code{3}.}
}
\value{
The \code{dataset} with replaced outliers.
}
\description{
Replace outliers in columns of a data set with \code{NA}
}
\seealso{
\code{\link{outlier}}.

Other Outlier: 
\code{\link{outlier}()},
\code{\link{winsorize}()}
}
\concept{Outlier}
