% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepMisc.r
\name{\%>\%}
\alias{\%>\%}
\title{Pipe operator}
\usage{
lhs %>% rhs
}
\description{
Pipe operator
}
\details{
R v4.1.0 directly provides simple native forward pipe operator |>. The simple form of the forward pipe inserts the left-hand side as the first argument in the right-hand side call.
}
