% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepRNN.r
\name{predict.ANN}
\alias{predict.ANN}
\title{Predict with ANN model}
\usage{
\method{predict}{ANN}(
  model,
  X.tensor,
  batch_size = 1,
  scale_type = NULL,
  scaler = NULL,
  diff_type = NULL,
  timesteps = 1,
  lag = 0,
  differences = 1,
  invert_first_row = NULL,
  Y.actual = NULL,
  type = "univariate"
)
}
\arguments{
\item{model}{A model object, returned by \code{fit.MLP} or \code{fit.LSTM} in the list element \code{model}.}

\item{X.tensor}{A feature tensor returned by \code{as.MLP.X} or \code{as.LSTM.X}.}

\item{batch_size}{Batch size, the number of samples used per gradient update.}

\item{scale_type}{Type of scaling with supported techniques min-max scaling (\code{minmax}), z-score scaling (\code{zscore}) and log transformation (\code{log}).
Per default (\code{NULL}) no inverted scaling is done.}

\item{scaler}{Scaling factors for the different scaling types. The type min-max scaling needs a list with vectors with min and max values for each outcome,
z-score scaling needs a list with vectors with mean and sd values for each outcome, log transformation needs no scaler.}

\item{diff_type}{Type of differentiation with supported techniques simple-differentiation (\code{simple}), log-differentiation (\code{log}) and percentage-differentiation (\code{percentage}).
\code{NULL} (default) indicates that no inverted differentiation is done.}

\item{timesteps}{The number of timesteps; stands for the number of different periods within one sample (record) of the resampled feature matrix, returned by \code{as.LSTM.X}.}

\item{lag}{The number of considered lags on feature side.}

\item{differences}{The number of differences.}

\item{invert_first_row}{The row index of the first row of the training or test data set regarding to the raw data set before differencing.}

\item{Y.actual}{A vector, matrix or data frame of raw data outcome values used for invert differentiation.}

\item{type}{The type of time series: \code{univariate} or \code{multivariate}.}
}
\value{
A two- or three-dimensional array with predicted outcome values.
  A two-dimensional array results for \code{fit.MLP} or for \code{fit.LSTM} if \code{return_sequences = FALSE}.
  A three-dimensional array results for \code{fit.LSTM} if \code{return_sequences = TRUE}.
}
\description{
Predict with ANN model
}
\seealso{
\code{\link{fit.MLP}}, \code{\link{as.MLP.X}}, \code{\link{fit.LSTM}}, \code{\link{as.LSTM.X}},
  \code{\link[stats]{predict}}, \code{\link{scale.datasets}}.

Other Single & Multi Layer Perceptron (SLP, MLP): 
\code{\link{as.ANN.matrix}()},
\code{\link{as.MLP.X}()},
\code{\link{as.MLP.Y}()},
\code{\link{as.tensor.1D}()},
\code{\link{as.tensor.2D}()},
\code{\link{as.tensor.3D}()},
\code{\link{as.tensor}()},
\code{\link{build.MLP}()},
\code{\link{fit.MLP}()},
\code{\link{get.MLP.X.samples}()},
\code{\link{get.MLP.X.units}()},
\code{\link{get.MLP.Y.samples}()},
\code{\link{get.MLP.Y.units}()},
\code{\link{vector.as.ANN.matrix}()},
\code{\link{vector.as.numeric}()}

Other Recurrent Neural Network (RNN): 
\code{\link{as.LSTM.X}()},
\code{\link{as.LSTM.Y}()},
\code{\link{as.LSTM.data.frame}()},
\code{\link{as.LSTM.period_outcome}()},
\code{\link{as.lag}()},
\code{\link{as.timesteps}()},
\code{\link{build.LSTM}()},
\code{\link{fit.LSTM}()},
\code{\link{get.LSTM.X.samples}()},
\code{\link{get.LSTM.X.timesteps}()},
\code{\link{get.LSTM.X.units}()},
\code{\link{get.LSTM.XY}()},
\code{\link{get.LSTM.Y.samples}()},
\code{\link{get.LSTM.Y.timesteps}()},
\code{\link{get.LSTM.Y.units}()},
\code{\link{get.period_shift}()},
\code{\link{start.invert_differencing}()}
}
\concept{Recurrent Neural Network (RNN)}
\concept{Single & Multi Layer Perceptron (SLP, MLP)}
