% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepMLP.r
\name{predict.MLP}
\alias{predict.MLP}
\title{Predict with SLP/MLP model}
\usage{
\method{predict}{MLP}(mlp, X, batch_size = 1, y.min = NULL, y.max = NULL)
}
\arguments{
\item{mlp}{A model object, returned by \code{fit.MLP} in the list element \code{model}.}

\item{X}{A feature data set, usually a matrix or data.frame.}

\item{batch_size}{Batch size, the number of samples used per gradient update.}

\item{y.min}{A vector with minima of outcomes, returned by \code{normalize_data}, otherwise \code{NULL}.}

\item{y.max}{A vector with maxima of outcomes, returned by \code{normalize_data}, otherwise \code{NULL}.}
}
\value{
A matrix with predicted outcome values per column.
}
\description{
Predict with SLP/MLP model
}
\seealso{
\code{\link{fit.MLP}}, \code{\link[stats]{predict}}, \code{\link{normalize_data}}.

Other Single & Multi Layer Perceptron (SLP, MLP): 
\code{\link{as.ANN.matrix}()},
\code{\link{as.MLP.X}()},
\code{\link{as.MLP.Y}()},
\code{\link{build.MLP}()},
\code{\link{fit.MLP}()},
\code{\link{get.MLP.X.units}()},
\code{\link{get.MLP.Y.units}()}
}
\concept{Single & Multi Layer Perceptron (SLP, MLP)}
