% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepML.r
\name{predict.naivebayes}
\alias{predict.naivebayes}
\title{Prediction for naive Bayes}
\usage{
\method{predict}{naivebayes}(object, x, ...)
}
\arguments{
\item{object}{R object.}

\item{x}{A matrix or data frame with feature values.}

\item{...}{Optional arguments.}
}
\value{
Numeric values for classifying the features within \code{x} to the levels of \code{y} stored in \code{object}.
}
\description{

}
\seealso{
Other Machine Learning: 
\code{\link{cross_validation_split}()},
\code{\link{k_nearest_neighbors}()},
\code{\link{moving_average}()},
\code{\link{naive_bayes}()},
\code{\link{naive_forecast}()},
\code{\link{predict.kmeans}()}
}
\concept{Machine Learning}
