% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepUtils.r
\name{probability}
\alias{probability}
\alias{probability.character}
\alias{probability.factor}
\alias{probability.logical}
\alias{probability.numeric}
\title{Probability}
\usage{
probability(x, ...)

\method{probability}{character}(x, y, laplace = 0)

\method{probability}{factor}(x, y, laplace = 0)

\method{probability}{logical}(x, y, laplace = 0)

\method{probability}{numeric}(x, FUN, ...)
}
\arguments{
\item{x}{A scalar or vector the probability is calculated for.}

\item{...}{Optional arguments to \code{FUN}.}

\item{y}{A factor variable the probability distribution is calculated to get the probability of a categorical \code{x}.}

\item{laplace}{A value for Laplace smoothing to avoid zero probability problem, default \code{0} is equal to no smoothing.}

\item{FUN}{The function to be applied to compute the probability distribution. If no function is specified, gaussian distribution density function \code{dnorm()} is applied to calculate the probability of a continuous \code{x}.}
}
\value{
Probability of \code{x}.
}
\description{

}
\details{
\code{x} is a scalar or vector with values from either a categorical or a continuous variable, the probability is calculated for.
}
\examples{
  # Getting the probability distribution of a factor variable y
  probability(levels(y), y)
  # Calculate the probability of every entry within a factor variable y
  probability(y, y)
  # Calculate the probability of certain levels or characters of a factor variable y
  probability(c("Oma", "Opa"), y)

  # Compute the probability of a numeric variable underlying a gaussian probability density function
  # with mean and standard deviation of x
  x <- 1:10
  probability(x)
  # with pregiven mean and standard deviation
  probability(x, mean = 0, sd = 1)
}
\seealso{
\code{\link[stats]{Normal}}

Other Utils: 
\code{\link{as_ANN_matrix}()},
\code{\link{as_tensor_1d}()},
\code{\link{as_tensor_2d}()},
\code{\link{as_tensor_3d}()},
\code{\link{degree}()},
\code{\link{dimC<-}()},
\code{\link{distance}()},
\code{\link{eye}()},
\code{\link{flatten}()},
\code{\link{insert}()},
\code{\link{list_as_numeric}()},
\code{\link{mbind}()},
\code{\link{ndim}()},
\code{\link{nsamples}()},
\code{\link{nsubsequences}()},
\code{\link{ntimesteps}()},
\code{\link{nunits}()},
\code{\link{radian}()},
\code{\link{random_seed}()},
\code{\link{re.factor}()},
\code{\link{reshape.array}()},
\code{\link{sd_pop}()},
\code{\link{similarity}()},
\code{\link{slice}()},
\code{\link{squeeze}()},
\code{\link{ta}()},
\code{\link{tensor}()},
\code{\link{vander}()},
\code{\link{var_pop}()},
\code{\link{vector_as_ANN_matrix}()},
\code{\link{vector_as_numeric}()}
}
\concept{Utils}
