% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepQuality.r
\name{quantile_loss}
\alias{quantile_loss}
\title{Quantile loss}
\usage{
quantile_loss(y, yhat, q = 0.5)
}
\arguments{
\item{y}{A numeric vector with actual values (to-be values).}

\item{yhat}{A numeric vector with estimated values (as-is values).}

\item{q}{A quantile fraction between 0 and 1.}
}
\value{
Quantile loss.
}
\description{
Quantile loss
}
\seealso{
Other Quality: 
\code{\link{accuracy}()},
\code{\link{huber}()},
\code{\link{log_cosh}()},
\code{\link{mae}()},
\code{\link{mape}()},
\code{\link{mse}()},
\code{\link{rmse}()},
\code{\link{vc}()}
}
\concept{Quality}
