% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepUtils.r
\name{radian}
\alias{radian}
\title{Degree to radian}
\usage{
radian(degree)
}
\arguments{
\item{degree}{Degree.}
}
\value{
The radian of \code{degree}.
}
\description{
Degree to radian
}
\examples{
  radian(180) # pi
}
\seealso{
Other Utils: 
\code{\link{as_ANN_matrix}()},
\code{\link{as_tensor_1d}()},
\code{\link{as_tensor_2d}()},
\code{\link{as_tensor_3d}()},
\code{\link{degree}()},
\code{\link{dimC<-}()},
\code{\link{distance}()},
\code{\link{expand_dims}()},
\code{\link{eye}()},
\code{\link{flatten}()},
\code{\link{insert}()},
\code{\link{list_as_numeric}()},
\code{\link{mabind}()},
\code{\link{mbind}()},
\code{\link{ndim}()},
\code{\link{nsamples}()},
\code{\link{nsubsequences}()},
\code{\link{ntimesteps}()},
\code{\link{nunits}()},
\code{\link{probability}()},
\code{\link{random_seed}()},
\code{\link{re.factor}()},
\code{\link{reshape.array}()},
\code{\link{sd_pop}()},
\code{\link{similarity}()},
\code{\link{slice}()},
\code{\link{squeeze}()},
\code{\link{ta}()},
\code{\link{tensor}()},
\code{\link{vander}()},
\code{\link{var_pop}()},
\code{\link{vector_as_ANN_matrix}()},
\code{\link{vector_as_numeric}()}
}
\concept{Utils}
