% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepSampler.r
\docType{class}
\name{RandomOverSampler}
\alias{RandomOverSampler}
\title{RandomOverSampler class}
\description{
A class for random oversampling datasets.
}
\section{Super class}{
\code{\link[deepANN:Sampler]{deepANN::Sampler}} -> \code{RandomOverSampler}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RandomOverSampler-fit_resample}{\code{RandomOverSampler$fit_resample()}}
\item \href{#method-RandomOverSampler-clone}{\code{RandomOverSampler$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="deepANN" data-topic="Sampler" data-id="fit"><a href='../../deepANN/html/Sampler.html#method-Sampler-fit'><code>deepANN::Sampler$fit()</code></a></span></li>
<li><span class="pkg-link" data-pkg="deepANN" data-topic="Sampler" data-id="initialize"><a href='../../deepANN/html/Sampler.html#method-Sampler-initialize'><code>deepANN::Sampler$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RandomOverSampler-fit_resample"></a>}}
\if{latex}{\out{\hypertarget{method-RandomOverSampler-fit_resample}{}}}
\subsection{Method \code{fit_resample()}}{
Random oversample the dataset.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RandomOverSampler$fit_resample(X, y)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X}}{Feature data set, usually a \code{\link{data.frame}} or an \code{\link{array}} of shape (n_samples, n_features).}

\item{\code{y}}{Target data set, usually a column of a \code{\link{data.frame}} or an \code{\link{array}} which holds the corresponding label for each sample in \code{X}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list with \code{X_resampled} and \code{y_resampled}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RandomOverSampler-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RandomOverSampler-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RandomOverSampler$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
