% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepSampler.r
\docType{class}
\name{RandomUnderSampler}
\alias{RandomUnderSampler}
\title{RadomUnderSampler class}
\description{
A class for random undersampling datasets.
}
\section{Super class}{
\code{\link[deepANN:Sampler]{deepANN::Sampler}} -> \code{RandomUnderSampler}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RandomUnderSampler-fit_resample}{\code{RandomUnderSampler$fit_resample()}}
\item \href{#method-RandomUnderSampler-clone}{\code{RandomUnderSampler$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="deepANN" data-topic="Sampler" data-id="fit"><a href='../../deepANN/html/Sampler.html#method-Sampler-fit'><code>deepANN::Sampler$fit()</code></a></span></li>
<li><span class="pkg-link" data-pkg="deepANN" data-topic="Sampler" data-id="initialize"><a href='../../deepANN/html/Sampler.html#method-Sampler-initialize'><code>deepANN::Sampler$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RandomUnderSampler-fit_resample"></a>}}
\if{latex}{\out{\hypertarget{method-RandomUnderSampler-fit_resample}{}}}
\subsection{Method \code{fit_resample()}}{
Random undersample the dataset.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RandomUnderSampler$fit_resample(X, y)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X}}{Feature data set, usually a \code{\link{data.frame}} or an \code{\link{array}} of shape (n_samples, n_features).}

\item{\code{y}}{Target data set, usually a column of a \code{\link{data.frame}} or an \code{\link{array}} which holds the corresponding label for each sample in \code{X}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list with \code{X_resampled} and \code{y_resampled}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RandomUnderSampler-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RandomUnderSampler-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RandomUnderSampler$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
