% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepUtils.r
\name{re.factor}
\alias{re.factor}
\title{Renew an (ordered) factor object}
\usage{
re.factor(x)
}
\arguments{
\item{x}{An (ordered) factor object.}
}
\value{
The renewed (ordered) factor object \code{x}.
}
\description{
Renew an (ordered) factor object
}
\seealso{
\code{\link{factor}}.

Other Utils: 
\code{\link{as_ANN_matrix}()},
\code{\link{degree}()},
\code{\link{distance}()},
\code{\link{list_as_numeric}()},
\code{\link{probability}()},
\code{\link{radian}()},
\code{\link{random_seed}()},
\code{\link{sd_pop}()},
\code{\link{similarity}()},
\code{\link{var_pop}()},
\code{\link{vector_as_ANN_matrix}()},
\code{\link{vector_as_numeric}()}
}
\concept{Utils}
