% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepDummy.r
\name{remove_columns}
\alias{remove_columns}
\title{Remove columns with only one specific value}
\usage{
remove_columns(dataset, value = 0)
}
\arguments{
\item{dataset}{A data set, usually a data frame.}

\item{value}{The specified values searched for.}
}
\value{
The \code{dataset} without those columns that contain only one specific value.
}
\description{
Remove columns with only one specific value
}
\seealso{
Other Dummifying: 
\code{\link{append_rows}()},
\code{\link{dummify_multilabel}()},
\code{\link{dummify}()},
\code{\link{effectcoding}()},
\code{\link{one_hot_decode}()},
\code{\link{one_hot_encode}()},
\code{\link{resample_imbalanced}()},
\code{\link{sparse_encode}()}
}
\concept{Dummifying}
