% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepDummy.r
\name{resample.imbalanced}
\alias{resample.imbalanced}
\title{Resampling imbalanced data for classification problems}
\usage{
resample.imbalanced(dataset, x, y, n = 1, k = 1, type = "smote")
}
\arguments{
\item{dataset}{An imbalanced data set, usually a data.frame.}

\item{x}{The column indices which spawn the feature matrix.}

\item{y}{The column index of the target vector with class labels (categories).}

\item{n}{The number of newly created samples or the percentage of deleted samples.}

\item{k}{The number of nearest neighbors, only relevant for SMOTE technique.}

\item{type}{The technique to be used for creating a balanced data set.
\code{oversampling}: copy \code{n} rows of minority class (under-represented category)
\code{undersampling}: delete \code{n}% rows of majority class (over-represented category)
\code{smote}: Synthetic Minority Oversampling Technique (SMOTE): create \code{n} synthetic rows of minority class of \code{k} nearest neighbors}
}
\value{
A balanced data set.
}
\description{
\code{resample.imbalanced} resamples an imbalanced data set to get a balanced data set.
}
\references{
Chawla, Nitesh V., Bowyer, Kevin W., Hall, Lawrence O., Kegelmeyer, W. Philip (2002): SMOTE: Synthetic Minority Over-sampling Technique. In: Journal of Artificial Intelligence Research, 16 (2002), 321-357. https://doi.org/10.1613/jair.953;
 \url{https://www.cs.cmu.edu/afs/cs/project/jair/pub/volume16/chawla02a-html/chawla2002.html},
 \url{http://rikunert.com/SMOTE_explained}.
}
\seealso{
Other Dummyfication: 
\code{\link{dummify}()},
\code{\link{effectcoding}()},
\code{\link{one_hot_decode}()},
\code{\link{one_hot_encode}()},
\code{\link{remove_columns}()}
}
\concept{Dummyfication}
