% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepDummy.r
\name{resample_imbalanced}
\alias{resample_imbalanced}
\title{Resampling imbalanced data for classification problems}
\usage{
resample_imbalanced(
  dataset,
  x,
  y,
  n = 1L,
  k = 1L,
  type = c("oversampling", "undersampling", "smote")
)
}
\arguments{
\item{dataset}{An imbalanced data set, usually a data frame.}

\item{x}{The names or indices of the feature columns within \code{dataset}.}

\item{y}{The names or indices of the target columns with class labels (categories) within \code{dataset}.}

\item{n}{The number of newly created samples or the percentage of deleted samples.}

\item{k}{The number of nearest neighbors, only relevant for type \code{smote}.}

\item{type}{The technique to be used for creating a balanced data set.\cr
\code{oversampling}: copy \code{n} rows of minority class (under-represented category)\cr
\code{undersampling}: delete \code{n}\% rows of majority class (over-represented category)\cr
\code{smote}: Synthetic Minority Oversampling Technique (SMOTE): create \code{n} synthetic rows of minority class of \code{k} nearest neighbors}
}
\value{
A balanced data set.
}
\description{
\code{resample_imbalanced} resamples an imbalanced data set to get a balanced data set.
}
\references{
Chawla, Nitesh V., Bowyer, Kevin W., Hall, Lawrence O., Kegelmeyer, W. Philip (2002): SMOTE: Synthetic Minority Over-sampling Technique. In: Journal of Artificial Intelligence Research, 16 (2002), 321-357. https://doi.org/10.1613/jair.953;
\url{https://www.cs.cmu.edu/afs/cs/project/jair/pub/volume16/chawla02a-html/chawla2002.html},
\url{http://rikunert.com/SMOTE_explained}.
}
\seealso{
Other Dummifying: 
\code{\link{append_rows}()},
\code{\link{dummify}()},
\code{\link{dummify_multilabel}()},
\code{\link{effectcoding}()},
\code{\link{one_hot_decode}()},
\code{\link{one_hot_encode}()},
\code{\link{remove_columns}()},
\code{\link{sparse_encode}()}
}
\concept{Dummifying}
