% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepArray.r
\name{reshape.array}
\alias{reshape.array}
\alias{reshape.marray}
\title{Reshape an array}
\usage{
reshape.array(a, dim = NULL, order = c("C", "F"))

reshape.marray(a, dim = NULL, order = c("C", "F"))
}
\arguments{
\item{a}{An array.}

\item{dim}{A integerish vector of new dimensions to be set on the array.}

\item{order}{The order in which elements of \code{a} should be read during rearrangement.
By default, the order is equivalent to the \code{C}-style ordering and means elements should be read in row-major order.
In opposite, the \code{Fortran}-style ordering means elements should be read in column-major order.}
}
\value{
The (redimensioned) array \code{a}.
}
\description{
Reshape an array
}
\details{
This function corresponds to \code{reshape()} from NumPy.
}
\seealso{
Other Array: 
\code{\link{column_stack}()},
\code{\link{delete}()},
\code{\link{dimC<-}()},
\code{\link{empty}()},
\code{\link{expand_dims}()},
\code{\link{eye}()},
\code{\link{flatten}()},
\code{\link{flip}()},
\code{\link{full}()},
\code{\link{insert}()},
\code{\link{mabind}()},
\code{\link{mamatrix}()},
\code{\link{marray}()},
\code{\link{ndim}()},
\code{\link{nsize}()},
\code{\link{ones}()},
\code{\link{rot90}()},
\code{\link{row_stack}()},
\code{\link{slice}()},
\code{\link{squeeze}()},
\code{\link{transpose}()},
\code{\link{vander}()},
\code{\link{zeros}()}
}
\concept{Array}
