% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepArray.r
\name{rot90}
\alias{rot90}
\title{Array rotation}
\usage{
rot90(a, k = 1L, axes = c(1L, 2L))
}
\arguments{
\item{a}{An array.}

\item{k}{Number of times the array is rotated by 90 degree. Positive numbers represent clockwise rotation, negative numbers counterclockwise rotation.}

\item{axes}{The array is rotated in the plane defined by the axes. Axes must be different.}
}
\value{
A rotated view of \code{a}.
}
\description{
Rotate an array by 90 degrees in the plane specified by axes.
}
\details{
This function corresponds to \code{rot90()} from NumPy.
}
\seealso{
Other Array: 
\code{\link{DIM}()},
\code{\link{column_stack}()},
\code{\link{delete}()},
\code{\link{dimC<-}()},
\code{\link{embedseries}()},
\code{\link{empty}()},
\code{\link{expand_dims}()},
\code{\link{eye}()},
\code{\link{flatten}()},
\code{\link{flip}()},
\code{\link{full}()},
\code{\link{insert}()},
\code{\link{mabind}()},
\code{\link{mamatrix}()},
\code{\link{marray}()},
\code{\link{ndim}()},
\code{\link{nsize}()},
\code{\link{ones}()},
\code{\link{reshape.array}()},
\code{\link{row_stack}()},
\code{\link{slice}()},
\code{\link{squeeze}()},
\code{\link{transpose}()},
\code{\link{vander}()},
\code{\link{zeros}()}
}
\concept{Array}
