% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepSampler.r
\docType{class}
\name{Sampler}
\alias{Sampler}
\title{Base Sampler class}
\description{
This is an abstract class for sampling datasets.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{label}}{('character()')}

\item{\code{mode}}{('character()')}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Sampler-new}{\code{Sampler$new()}}
\item \href{#method-Sampler-fit}{\code{Sampler$fit()}}
\item \href{#method-Sampler-fit_resample}{\code{Sampler$fit_resample()}}
\item \href{#method-Sampler-clone}{\code{Sampler$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Sampler-new"></a>}}
\if{latex}{\out{\hypertarget{method-Sampler-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sampler$new(label = NA_character_, mode = c("auto"))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{label}}{The label for the sampler object.}

\item{\code{mode}}{Currently unused member.

Note that this object is typically constructed by a derived class.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Sampler-fit"></a>}}
\if{latex}{\out{\hypertarget{method-Sampler-fit}{}}}
\subsection{Method \code{fit()}}{
Check inputs and statistics of the sampler. Normally, you should use \code{fit_resample} in all cases.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sampler$fit(X, y)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X}}{Feature dataset, usually a \code{\link{data.frame}} or an \code{\link{array}}.}

\item{\code{y}}{Target dataset, usually a column of a \code{\link{data.frame}} or an \code{\link{array}}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list with \code{X_resampled} and \code{y_resampled}.

Note that this object is typically constructed by a derived class.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Sampler-fit_resample"></a>}}
\if{latex}{\out{\hypertarget{method-Sampler-fit_resample}{}}}
\subsection{Method \code{fit_resample()}}{
Resample the dataset.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sampler$fit_resample(X, y)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X}}{Feature data set, usually a \code{\link{data.frame}} or an \code{\link{array}}.}

\item{\code{y}}{Target data set, usually a column of a \code{\link{data.frame}} or an \code{\link{array}}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list with \code{X_resampled} and \code{y_resampled}.

Note that this object is typically constructed by a derived class.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Sampler-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Sampler-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sampler$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
