% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepSampler.r
\name{oversampling}
\alias{oversampling}
\alias{random_oversampling}
\alias{smote}
\alias{undersampling}
\alias{random_undersampling}
\title{Wrapper function for class Oversampler.}
\usage{
oversampling(X, y, mode = c("auto"))

random_oversampling(X, y, mode = c("auto"))

smote(X, y, mode = c("auto"), k_neighbors = 3L)

undersampling(X, y, mode = c("auto"))

random_undersampling(X, y, mode = c("auto"))
}
\arguments{
\item{X}{Feature data set, usually a \code{\link{data.frame}} or an \code{\link{array}} of shape (n_samples, n_features).}

\item{y}{Target data set, usually a column of a \code{\link{data.frame}} or an \code{\link{array}} which holds the corresponding label for each sample in \code{X}.}

\item{k_neighbors}{The number of nearest neighbors used to define the neighborhood of samples to use to generate the synthetic samples.}
}
\value{
A list with \code{X_resampled} and \code{y_resampled}.
}
\description{
Wrapper function for class Oversampler.

Wrapper function for class RandomOversampler.

Wrapper function for class SMOTE.

Wrapper function for class UnderSampler.

Wrapper function for class RandomUnderSampler.
}
