% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepRNN.r
\name{save_weights_ANN}
\alias{save_weights_ANN}
\title{Save model weights to file}
\usage{
save_weights_ANN(model, filename)
}
\arguments{
\item{model}{A model object.}

\item{filename}{The file name.}
}
\value{
The model object.
}
\description{

}
\seealso{
\code{\link{load_weights_ANN}}, \code{\link[base]{files}},
  \code{\link[keras]{save_model_weights_hdf5}}.

Other Single & Multi Layer Perceptron (SLP, MLP): 
\code{\link{as_MLP_X}()},
\code{\link{as_MLP_Y}()},
\code{\link{build_MLP}()},
\code{\link{fit_MLP}()},
\code{\link{load_weights_ANN}()},
\code{\link{predict_ANN}()}

Other Recurrent Neural Network (RNN): 
\code{\link{as_LSTM_X}()},
\code{\link{as_LSTM_Y}()},
\code{\link{as_LSTM_data_frame}()},
\code{\link{as_LSTM_period_outcome}()},
\code{\link{as_lag}()},
\code{\link{as_timesteps}()},
\code{\link{build_LSTM}()},
\code{\link{fit_LSTM}()},
\code{\link{get_LSTM_XY}()},
\code{\link{get_period_shift}()},
\code{\link{load_weights_ANN}()},
\code{\link{predict_ANN}()},
\code{\link{start_invert_differencing}()}
}
\concept{Recurrent Neural Network (RNN)}
\concept{Single & Multi Layer Perceptron (SLP, MLP)}
