% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepScaling.r
\name{scale.center}
\alias{scale.center}
\title{(Mean) Centering}
\usage{
\method{scale}{center}(x, use.attr = TRUE, invert = FALSE, const = NULL)
}
\arguments{
\item{x}{A numeric vector whose elements should be scaled thru centering.}

\item{use.attr}{A logical value indicating whether the scaling factor \code{const} is stored as named attribute of the scaled vector \code{x}.}

\item{invert}{A logical value indicating the direction of scaling. If set to \code{TRUE}, \code{x} is an already scaled vector and scaling will be inverted.}

\item{const}{A constant value that is subtracted from all values in \code{x}. By default (\code{NULL}), the mean of \code{x} is used as the constant value.}
}
\value{
A vector of same length as \code{x} with (mean) centered values.
}
\description{
(Mean) Centering
}
\seealso{
\code{\link{scaling}}.

Other Scaling: 
\code{\link{scale.datasets}()},
\code{\link{scale.dataset}()},
\code{\link{scale.log}()},
\code{\link{scale.minmax}()},
\code{\link{scale.zscore}()},
\code{\link{scaling}()}
}
\concept{Scaling}
