% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepScaling.r
\name{scale.dataset}
\alias{scale.dataset}
\title{Scaling of a data set}
\usage{
\method{scale}{dataset}(
  dataset,
  columns = NULL,
  type = c("minmax", "zscore", "log"),
  use.attr = TRUE
)
}
\arguments{
\item{dataset}{A data set, usually a matrix or data frame.}

\item{columns}{The names or indices of the columns to be scaled.}

\item{type}{Type of scaling with supported techniques min-max scaling (\code{minmax}), z-score scaling (\code{zscore}) and log transformation (\code{log}).}

\item{use.attr}{A logical value indicating whether scaling factors like \code{min}, \code{max}, \code{mean} and \code{sd} are stored as named attributes of the scaled columns of \code{dataset}.}
}
\value{
The scaled dataset if \code{use.attr = T} (default), or a named list dependent on the \code{type}.
  \code{minmax}: The first element stores the min value, the second element the max value, and the third element the scaled dataset.
  \code{zscore}: The first element stores the mean value, the second element the sd value, and the third element the scaled dataset.
  \code{log}: The first element stores the scaled dataset.
}
\description{
Scaling of a data set
}
\seealso{
\code{\link{scaling}}

Other Scaling: 
\code{\link{scale.datasets}()},
\code{\link{scale.log}()},
\code{\link{scale.minmax}()},
\code{\link{scale.zscore}()},
\code{\link{scaling}()}
}
\concept{Scaling}
