% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepScaling.r
\name{scale.log}
\alias{scale.log}
\title{Log Transformation}
\usage{
\method{scale}{log}(x, invert = FALSE)
}
\arguments{
\item{x}{A numeric vector whose elements should be log transformed.}

\item{invert}{A logical value indicating the direction of scaling. If set to \code{TRUE}, \code{x} is an already scaled vector and scaling will be inverted.}
}
\value{
A vector of same length as \code{x} with log transformed values.
}
\description{
Log Transformation
}
\seealso{
\code{\link{scaling}}.

Other Scaling: 
\code{\link{scale.datasets}()},
\code{\link{scale.minmax}()},
\code{\link{scale.zscore}()},
\code{\link{scaling}()}
}
\concept{Scaling}
