% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepScaling.r
\name{scale.zscore}
\alias{scale.zscore}
\title{Z-Score Scaling}
\usage{
\method{scale}{zscore}(x, use.attr = TRUE, invert = FALSE, meanx = NULL, sdx = NULL)
}
\arguments{
\item{x}{A numeric vector whose elements should be scaled thru z-score scaling.}

\item{use.attr}{A logical value indicating whether scaling factors \code{mean} and \code{sd} are stored as named attributes of the scaled vector \code{x}.}

\item{invert}{A logical value indicating the direction of scaling. If set to \code{TRUE}, \code{x} is an already scaled vector and scaling will be inverted.}

\item{meanx}{A given mean value for scaling. If this value is \code{NULL} the mean of \code{x} is used.}

\item{sdx}{A given standard deviation value for scaling. If this value is \code{NULL} the standard deviation of \code{x} is used.}
}
\value{
A vector of same length as \code{x} with z-score scaled values.
}
\description{
Z-Score Scaling
}
\seealso{
\code{\link{scaling}}.

Other Scaling: 
\code{\link{scale.datasets}()},
\code{\link{scale.log}()},
\code{\link{scale.minmax}()},
\code{\link{scaling}()}
}
\concept{Scaling}
