% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepScaling.r
\name{scale_center}
\alias{scale_center}
\title{(Mean) Centering}
\usage{
scale_center(x, use_attr = TRUE, invert = FALSE, const = NULL)
}
\arguments{
\item{x}{A numeric object whose elements should be scaled thru centering.}

\item{use_attr}{A logical value indicating whether the scaling factor \code{const} is stored as named attribute of the scaled \code{x}.}

\item{invert}{A logical value indicating the direction of scaling. If set to \code{TRUE}, \code{x} is already scaled and scaling will be inverted.}

\item{const}{A constant value that is subtracted from all values in \code{x}. By default (\code{NULL}), the mean of \code{x} is used as the constant value.}
}
\value{
Transformed \code{x} with (mean) centered values.
}
\description{
(Mean) Centering
}
\seealso{
\code{\link{scaling}}.

Other Scaling: 
\code{\link{scale_dataset}()},
\code{\link{scale_log}()},
\code{\link{scale_minmax}()},
\code{\link{scale_train_test}()},
\code{\link{scale_zscore}()},
\code{\link{scaling}()}
}
\concept{Scaling}
