% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepScaling.r
\name{scale_dataset}
\alias{scale_dataset}
\title{Scaling of a data set}
\usage{
scale_dataset(
  dataset,
  columns = NULL,
  type = c("minmax", "zscore", "log", "center"),
  invert = FALSE,
  ...
)
}
\arguments{
\item{dataset}{A data set, usually a matrix or data frame.}

\item{columns}{The names or indices of the columns to be scaled. If \code{NULL} (default), all columns are used.}

\item{type}{Type of scaling with supported techniques min-max scaling (\code{minmax}), z-score scaling (\code{zscore}), log transformation (\code{log}) and centering \code{center}.}

\item{invert}{A logical value indicating the direction of scaling. If set to \code{TRUE}, \code{columns} are already scaled vectors and scaling will be inverted.}

\item{...}{Further arguments depend on the type of scaling. Min-max scaling, z-score scaling and centering need two or one further arguments if necessary.}
}
\value{
The scaled \code{dataset}.
}
\description{

}
\seealso{
\code{\link{scaling}}

Other Scaling: 
\code{\link{scale_center}()},
\code{\link{scale_log}()},
\code{\link{scale_minmax}()},
\code{\link{scale_train_test}()},
\code{\link{scale_zscore}()},
\code{\link{scaling}()}
}
\concept{Scaling}
