% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepScaling.r
\name{scale_log}
\alias{scale_log}
\title{Log Transformation}
\usage{
scale_log(x, invert = FALSE)
}
\arguments{
\item{x}{A numeric vector whose elements should be log transformed.}

\item{invert}{A logical value indicating the direction of scaling. If set to \code{TRUE}, \code{x} is an already scaled vector and scaling will be inverted.}
}
\value{
A vector of same length as \code{x} with log transformed values.
}
\description{

}
\seealso{
\code{\link{scaling}}.

Other Scaling: 
\code{\link{scale_center}()},
\code{\link{scale_dataset}()},
\code{\link{scale_minmax}()},
\code{\link{scale_train_test}()},
\code{\link{scale_zscore}()},
\code{\link{scaling}()}
}
\concept{Scaling}
