% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepScaling.r
\name{scale_minmax}
\alias{scale_minmax}
\title{Min-Max Scaling}
\usage{
scale_minmax(x, use_attr = TRUE, invert = FALSE, minx = NULL, maxx = NULL)
}
\arguments{
\item{x}{A numeric vector whose elements should be scaled thru min-max scaling.}

\item{use_attr}{A logical value indicating whether scaling factors \code{min} and \code{max} are stored as named attributes of the scaled vector \code{x}.}

\item{invert}{A logical value indicating the direction of scaling. If set to \code{TRUE}, \code{x} is an already scaled vector and scaling will be inverted.}

\item{minx}{A given minimum value for scaling. If this value is \code{NULL} the minimum of \code{x} is used.}

\item{maxx}{A given maximum value for scaling. If this value is \code{NULL} the maximum of \code{x} is used.}
}
\value{
A vector of same length as \code{x} with min-max scaled values between 0 and 1.
}
\description{

}
\seealso{
\code{\link{scaling}}.

Other Scaling: 
\code{\link{scale_center}()},
\code{\link{scale_dataset}()},
\code{\link{scale_log}()},
\code{\link{scale_train_test}()},
\code{\link{scale_zscore}()},
\code{\link{scaling}()}
}
\concept{Scaling}
