% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepScaling.r
\name{scale_train_test}
\alias{scale_train_test}
\title{Scaling of a train and test data set}
\usage{
scale_train_test(
  trainset,
  testset,
  columns = NULL,
  type = c("minmax", "zscore", "log", "center")
)
}
\arguments{
\item{trainset}{A train data set.}

\item{testset}{A test data set.}

\item{columns}{The names or indices of the columns to be scaled. If \code{NULL} (default), all columns are used.}

\item{type}{Type of scaling with supported techniques min-max scaling (\code{minmax}), z-score scaling (\code{zscore}), log transformation (\code{log}) and centering \code{center}.}
}
\value{
A named list dependent on the \code{type}.\cr
  \code{minmax}: The first element stores the min value, the second element the max value, the third element the scaled train data set and the fourth element the scaled test data set.\cr
  \code{zscore}: The first element stores the mean value, the second element the sd value, the third element the scaled train data set and the fourth element the scaled test data set.\cr
  \code{log}: The first element stores the scaled train data set and the second element the scaled test data set.\cr
  \code{center}: The first element stores the constant value, the second element the scaled train data set and the third element the scaled test data set.
}
\description{

}
\seealso{
\code{\link{scaling}}

Other Scaling: 
\code{\link{scale_center}()},
\code{\link{scale_dataset}()},
\code{\link{scale_log}()},
\code{\link{scale_minmax}()},
\code{\link{scale_zscore}()},
\code{\link{scaling}()}
}
\concept{Scaling}
