% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepScaling.r
\name{scale_zscore}
\alias{scale_zscore}
\title{Z-Score Scaling}
\usage{
scale_zscore(x, use_attr = TRUE, invert = FALSE, meanx = NULL, sdx = NULL)
}
\arguments{
\item{x}{A numeric object whose elements should be scaled thru z-score scaling.}

\item{use_attr}{A logical value indicating whether scaling factors \code{mean} and \code{sd} are stored as named attributes of the scaled \code{x}.}

\item{invert}{A logical value indicating the direction of scaling. If set to \code{TRUE}, \code{x} is already scaled and scaling will be inverted.}

\item{meanx}{A given mean value for scaling. If this value is \code{NULL} the mean of \code{x} is used.}

\item{sdx}{A given standard deviation value for scaling. If this value is \code{NULL} the standard deviation of \code{x} is used.}
}
\value{
Transformed \code{x} with z-score scaled values.
}
\description{
Z-Score Scaling
}
\seealso{
\code{\link{scaling}}.

Other Scaling: 
\code{\link{scale_center}()},
\code{\link{scale_dataset}()},
\code{\link{scale_log}()},
\code{\link{scale_minmax}()},
\code{\link{scale_train_test}()},
\code{\link{scaling}()}
}
\concept{Scaling}
