% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepScaling.r
\docType{class}
\name{Scaler}
\alias{Scaler}
\alias{is.Scaler}
\title{Class Scaler}
\usage{
is.Scaler(object)
}
\description{
This is the abstract base class for scaler objects.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Scaler-clone}{\code{Scaler$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Scaler-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Scaler-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Scaler$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
