% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepML.r
\name{sd.pop}
\alias{sd.pop}
\title{Population standard deviation}
\usage{
sd.pop(x, na.rm = FALSE)
}
\arguments{
\item{x}{A numeric vector, matrix or data frame.}

\item{na.rm}{A logical value indicating whether missing values should be removed or not (default).}
}
\value{
The population standard deviation.
}
\description{
Population standard deviation
}
\details{
The population standard deviation and sample standard deviation, implemented in \code{stats} package, differ in the denominator.
  The value of denominator in the formula for standard deviation in case of population data is \code{n}, while it is \code{n-1} for sample data.
}
\examples{
  x <- sample(1000)
  sd.pop(x)
}
\seealso{
\code{\link[stats]{sd}}.

Other Machine Learning: 
\code{\link{as.degree}()},
\code{\link{as.radian}()},
\code{\link{cross_validation_split}()},
\code{\link{distance}()},
\code{\link{erfcinv}()},
\code{\link{erfc}()},
\code{\link{erfinv}()},
\code{\link{erf}()},
\code{\link{k_nearest_neighbors}()},
\code{\link{moving_average}()},
\code{\link{naive_forecast}()},
\code{\link{predict.kmeans}()},
\code{\link{similarity}()},
\code{\link{var.pop}()}
}
\concept{Machine Learning}
