% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepUtils.r
\name{sd_pop}
\alias{sd_pop}
\title{Population standard deviation}
\usage{
sd_pop(x, na.rm = FALSE)
}
\arguments{
\item{x}{A numeric vector, matrix or data frame.}

\item{na.rm}{A logical value indicating whether missing values should be removed or not (default).}
}
\value{
The population standard deviation.
}
\description{

}
\details{
The population standard deviation and sample standard deviation, implemented in \code{stats} package, differ in the denominator.
  The value of denominator in the formula for standard deviation in case of population data is \code{n}, while it is \code{n-1} for sample data.
}
\examples{
  x <- sample(1000)
  sd_pop(x)
}
\seealso{
\code{\link[stats]{sd}}.

Other Utils: 
\code{\link{as_ANN_matrix}()},
\code{\link{as_tensor_1D}()},
\code{\link{as_tensor_2D}()},
\code{\link{as_tensor_3D}()},
\code{\link{degree}()},
\code{\link{distance}()},
\code{\link{flatten}()},
\code{\link{insert}()},
\code{\link{list_as_numeric}()},
\code{\link{marray}()},
\code{\link{mbind}()},
\code{\link{ndim}()},
\code{\link{nsamples}()},
\code{\link{nsubsequences}()},
\code{\link{ntimesteps}()},
\code{\link{nunits}()},
\code{\link{probability}()},
\code{\link{radian}()},
\code{\link{re.factor}()},
\code{\link{similarity}()},
\code{\link{ta}()},
\code{\link{tensor}()},
\code{\link{var_pop}()},
\code{\link{vector_as_ANN_matrix}()},
\code{\link{vector_as_numeric}()}
}
\concept{Utils}
