% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepML.r
\name{similarity}
\alias{similarity}
\title{Similarity}
\usage{
similarity(x1, x2, type = c("simple", "jaccard", "tanimoto"))
}
\arguments{
\item{x1}{A numeric or logical vector.}

\item{x2}{A numeric or logical vector.}

\item{type}{The type of the similarity measure: \code{simple} (default), \code{jaccard} or \code{tanimoto}.}
}
\value{
The similarity between \code{x1} and \code{x2}.
}
\description{
Similarity
}
\examples{
  similarity(c(1L, 1L, 0L), c(0L, 1L, 1L), type = "simple")
  similarity(c(1L, 1L, 0L), c(0L, 1L, 1L), type = "jaccard")
  similarity(c(1L, 1L, 0L), c(0L, 1L, 1L), type = "tanimoto")
}
\seealso{
Other Machine Learning: 
\code{\link{as.degree}()},
\code{\link{as.radian}()},
\code{\link{cross_validation_split}()},
\code{\link{distance}()},
\code{\link{erfcinv}()},
\code{\link{erfc}()},
\code{\link{erfinv}()},
\code{\link{erf}()},
\code{\link{k_nearest_neighbors}()},
\code{\link{moving_average}()},
\code{\link{naive_forecast}()},
\code{\link{predict.kmeans}()},
\code{\link{sd.pop}()},
\code{\link{var.pop}()}
}
\concept{Machine Learning}
