% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepSampler.r
\docType{class}
\name{SMOTE}
\alias{SMOTE}
\title{SMOTE class}
\description{
A class for to perform oversampling using SMOTE (Synthetic Minority Oversampling Technique).
}
\references{
Chawla, N.V., Bowyer, K.W., Hall, L.O., & Kegelmeyer, W.P. (2000). SMOTE: Synthetic Minority Over-sampling Technique. In International Conference of Knowledge Based Computer Systems, pp. 46-57. National Center for Software Technology, Mumbai, India, Allied Press.
Chawla, N.V., Bowyer, K.W., Hall, L.O., & Kegelmeyer, W.P. (2002). SMOTE: Synthetic Minority Over-sampling Technique. In Journal of Artificial Intelligence Research, 16 (1), 321-357. https://doi.org/10.1613/jair.953.
}
\section{Super class}{
\code{\link[deepANN:Sampler]{deepANN::Sampler}} -> \code{SMOTE}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{label}}{('integer()')}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{label}}{('integer()')}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SMOTE-new}{\code{SMOTE$new()}}
\item \href{#method-SMOTE-fit_resample}{\code{SMOTE$fit_resample()}}
\item \href{#method-SMOTE-clone}{\code{SMOTE$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="deepANN" data-topic="Sampler" data-id="fit"><a href='../../deepANN/html/Sampler.html#method-Sampler-fit'><code>deepANN::Sampler$fit()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SMOTE-new"></a>}}
\if{latex}{\out{\hypertarget{method-SMOTE-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SMOTE$new(label = NA_character_, mode = c("auto"), k_neighbors = 3L)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{label}}{The label for the sampler object.}

\item{\code{mode}}{Currently unused member.}

\item{\code{k_neighbors}}{The number of nearest neighbors used to define the neighborhood of samples to use to generate the synthetic samples.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SMOTE-fit_resample"></a>}}
\if{latex}{\out{\hypertarget{method-SMOTE-fit_resample}{}}}
\subsection{Method \code{fit_resample()}}{
SMOTE oversample the dataset.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SMOTE$fit_resample(X, y)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X}}{Feature data set, usually a \code{\link{data.frame}} or an \code{\link{array}} of shape (n_samples, n_features).}

\item{\code{y}}{Target data set, usually a column of a \code{\link{data.frame}} or an \code{\link{array}} which holds the corresponding label for each sample in \code{X}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list with \code{X_resampled} and \code{y_resampled}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SMOTE-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SMOTE-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SMOTE$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
