% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepDummy.r
\name{sparse_encode}
\alias{sparse_encode}
\title{Sparse encoding}
\usage{
sparse_encode(x)
}
\arguments{
\item{x}{A vector with values (levels) of a categorical variable.}
}
\value{
An array with numeric encoded levels of \code{x}.
}
\description{
Sparse encoding
}
\examples{
  x <- factor(sample(lvls <- c("dog", "cat", "mouse"), 10, TRUE), levels = lvls)
  sparse_encode(x)

}
\seealso{
\code{\link{one_hot_encode}}.

Other Dummifying: 
\code{\link{append_rows}()},
\code{\link{dummify_multilabel}()},
\code{\link{dummify}()},
\code{\link{effectcoding}()},
\code{\link{one_hot_decode}()},
\code{\link{one_hot_encode}()},
\code{\link{remove_columns}()},
\code{\link{resample_imbalanced}()}
}
\concept{Dummifying}
