% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepScaling.r
\docType{class}
\name{StandardScaler}
\alias{StandardScaler}
\alias{zscore}
\alias{inverse_zscore}
\title{class StandardScaler}
\usage{
zscore(x, .mean, .sd)

inverse_zscore(x, .mean, .sd)
}
\description{
A class for transforming features by removing the mean and scaling to unit standard deviation.
}
\examples{
a <- random_int(dim = c(6, 4), min = 1, max = 30)
scaler <- StandardScaler$new()
xt <- scaler$fit_transform(a)
scaler$inverse_transform(xt)

a <- random_int(dim = c(4, 3, 2), min = 1, max = 30)
scaler <- StandardScaler$new()
xt <- scaler$fit_transform(a, axis = c(1, 2))
scaler$inverse_transform(xt, axis = c(1, 2))
}
\section{Super class}{
\code{\link[deepANN:Scaler]{deepANN::Scaler}} -> \code{StandardScaler}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{mean}}{('NULL' | numeric vector)\cr
The mean value for each feature.}

\item{\code{std}}{('NULL' | numeric vector)\cr
The standard deviation for each feature.}

\item{\code{with_mean}}{('NULL' | logical)\cr
If true, center the data before scaling.}

\item{\code{with_std}}{('NULL' | logical)\cr
If true, scale the data to unit standard deviation.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-StandardScaler-new}{\code{StandardScaler$new()}}
\item \href{#method-StandardScaler-fit}{\code{StandardScaler$fit()}}
\item \href{#method-StandardScaler-transform}{\code{StandardScaler$transform()}}
\item \href{#method-StandardScaler-fit_transform}{\code{StandardScaler$fit_transform()}}
\item \href{#method-StandardScaler-inverse_transform}{\code{StandardScaler$inverse_transform()}}
\item \href{#method-StandardScaler-clone}{\code{StandardScaler$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StandardScaler-new"></a>}}
\if{latex}{\out{\hypertarget{method-StandardScaler-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StandardScaler$new(with_mean = TRUE, with_std = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{with_mean}}{If true, center the data before scaling.}

\item{\code{with_std}}{if true, scale the data to unit standard deviation}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StandardScaler-fit"></a>}}
\if{latex}{\out{\hypertarget{method-StandardScaler-fit}{}}}
\subsection{Method \code{fit()}}{
Compute the mean and std to be used for later scaling.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StandardScaler$fit(X, axis = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X}}{The data used to compute the per-feature mean and std used for later scaling along the features axis.}

\item{\code{axis}}{The axis along computation is applied.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The instance \code{self} itself.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StandardScaler-transform"></a>}}
\if{latex}{\out{\hypertarget{method-StandardScaler-transform}{}}}
\subsection{Method \code{transform()}}{
Perform standardization by centering and scaling.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StandardScaler$transform(X, axis = 1, order = c("C", "F"))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X}}{The data that will be transformed.}

\item{\code{axis}}{The axis along scaling is applied.}

\item{\code{order}}{The order in which elements of data should be read during scaling. By default, the order is equivalent to the C-style ordering and means elements should be read in row-major order. In opposite, the Fortran-style ordering means elements should be read in column-major order.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The transformed \code{X}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StandardScaler-fit_transform"></a>}}
\if{latex}{\out{\hypertarget{method-StandardScaler-fit_transform}{}}}
\subsection{Method \code{fit_transform()}}{
Run \code{fit()} and \code{transform()}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StandardScaler$fit_transform(X, axis = 1, order = c("C", "F"))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X}}{The data that will be transformed.}

\item{\code{axis}}{The axis along scaling is applied.}

\item{\code{order}}{The order in which elements of data should be read during scaling. By default, the order is equivalent to the C-style ordering and means elements should be read in row-major order. In opposite, the Fortran-style ordering means elements should be read in column-major order.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The transformed \code{X}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StandardScaler-inverse_transform"></a>}}
\if{latex}{\out{\hypertarget{method-StandardScaler-inverse_transform}{}}}
\subsection{Method \code{inverse_transform()}}{
Scale back \code{X} to the original representation.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StandardScaler$inverse_transform(X, axis = 1, order = c("C", "F"))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X}}{The data that will be transformed.}

\item{\code{axis}}{The axis along scaling is applied.}

\item{\code{order}}{The order in which elements of data should be read during scaling. By default, the order is equivalent to the C-style ordering and means elements should be read in row-major order. In opposite, the Fortran-style ordering means elements should be read in column-major order.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The transformed \code{X}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StandardScaler-clone"></a>}}
\if{latex}{\out{\hypertarget{method-StandardScaler-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StandardScaler$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
