% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepRNN.r
\name{start.invert_differencing}
\alias{start.invert_differencing}
\title{Start row index/period for invert differencing}
\usage{
\method{start}{invert_differencing}(
  invert_first_row,
  differences = 1,
  timesteps = 1,
  lag = 0,
  type = "univariate"
)
}
\arguments{
\item{invert_first_row}{The row index of the first row of the training or test data set regarding to the raw data set before differencing.}

\item{differences}{The number of differences.}

\item{timesteps}{The number of timesteps.}

\item{lag}{The number of considered lags on feature side.}

\item{type}{The type of time series: \code{univariate} or \code{multivariate}.}
}
\value{
The start row index of raw data outcome values before differencing.
}
\description{
\code{start.invert_differencing} calculates the row index where to start invert differencing.
}
\seealso{
\code{\link{get.period_shift}}.

Other Recurrent Neural Network (RNN): 
\code{\link{as.LSTM.X}()},
\code{\link{as.LSTM.Y}()},
\code{\link{as.LSTM.data.frame}()},
\code{\link{as.LSTM.period_outcome}()},
\code{\link{as.lag}()},
\code{\link{as.timesteps}()},
\code{\link{build.LSTM}()},
\code{\link{fit.LSTM}()},
\code{\link{get.LSTM.X.samples}()},
\code{\link{get.LSTM.X.timesteps}()},
\code{\link{get.LSTM.X.units}()},
\code{\link{get.LSTM.XY}()},
\code{\link{get.LSTM.Y.samples}()},
\code{\link{get.LSTM.Y.timesteps}()},
\code{\link{get.LSTM.Y.units}()},
\code{\link{get.period_shift}()},
\code{\link{predict.ANN}()}
}
\concept{Recurrent Neural Network (RNN)}
