% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepRNN.r
\name{start_invert_differencing}
\alias{start_invert_differencing}
\title{Start row index/period for invert differencing}
\usage{
start_invert_differencing(
  invert_first_row,
  differences = 1L,
  timesteps = 1L,
  lag = 0L,
  type = "univariate"
)
}
\arguments{
\item{invert_first_row}{The row index of the first row of the training or test data set regarding to the raw data set before differencing.}

\item{differences}{The number of differences.}

\item{timesteps}{The number of timesteps.}

\item{lag}{The number of considered lags on feature side.}

\item{type}{The type of time series: \code{univariate} or \code{multivariate}.}
}
\value{
The start row index of raw data outcome values before differencing.
}
\description{
\code{start_invert_differencing} calculates the row index where to start invert differencing.
}
\seealso{
\code{\link{get_period_shift}}.

Other Recurrent Neural Network (RNN): 
\code{\link{as_LSTM_X}()},
\code{\link{as_LSTM_Y}()},
\code{\link{as_LSTM_data_frame}()},
\code{\link{as_LSTM_period_outcome}()},
\code{\link{as_lag}()},
\code{\link{as_timesteps}()},
\code{\link{build_LSTM}()},
\code{\link{fit_LSTM}()},
\code{\link{get_LSTM_XY}()},
\code{\link{get_period_shift}()},
\code{\link{load_weights_ANN}()},
\code{\link{predict_ANN}()},
\code{\link{save_weights_ANN}()}
}
\concept{Recurrent Neural Network (RNN)}
