% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepUtils.r
\name{ta}
\alias{ta}
\alias{ta.array}
\alias{ta.marray}
\title{Transpose multidimensional array}
\usage{
ta(a, ...)

\method{ta}{array}(a, perm = NULL)

\method{ta}{marray}(a, perm = NULL)
}
\arguments{
\item{a}{A multidimensional array.}

\item{perm}{The permutation vector of the dimensions. The default \code{NULL} indicates to reverse the order of the dimensions.}
}
\value{
The array \code{a} with swapped dimensions.
}
\description{
This function transposes a multidimensional array.
}
\seealso{
\code{\link{t}}.

Other Utils: 
\code{\link{as_ANN_matrix}()},
\code{\link{as_tensor_1d}()},
\code{\link{as_tensor_2d}()},
\code{\link{as_tensor_3d}()},
\code{\link{degree}()},
\code{\link{dimC<-}()},
\code{\link{distance}()},
\code{\link{eye}()},
\code{\link{flatten}()},
\code{\link{insert}()},
\code{\link{list_as_numeric}()},
\code{\link{mbind}()},
\code{\link{ndim}()},
\code{\link{nsamples}()},
\code{\link{nsubsequences}()},
\code{\link{ntimesteps}()},
\code{\link{nunits}()},
\code{\link{probability}()},
\code{\link{radian}()},
\code{\link{random_seed}()},
\code{\link{re.factor}()},
\code{\link{reshape.array}()},
\code{\link{sd_pop}()},
\code{\link{similarity}()},
\code{\link{slice}()},
\code{\link{squeeze}()},
\code{\link{tensor}()},
\code{\link{vander}()},
\code{\link{var_pop}()},
\code{\link{vector_as_ANN_matrix}()},
\code{\link{vector_as_numeric}()}
}
\concept{Utils}
