% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepUtils.r
\name{tensor}
\alias{tensor}
\alias{as.tensor}
\alias{as.tensor.default}
\alias{is.tensor}
\title{Tensor}
\usage{
tensor(data, ...)

as.tensor(data, ...)

\method{as.tensor}{default}(
  data,
  dim = NULL,
  dimnames = NULL,
  order = c("C", "F"),
  numeric = TRUE,
  reverse = FALSE
)

is.tensor(x)
}
\arguments{
\item{data}{A data set, e.g. vector, array, matrix, data frame, tibble, data.table.}

\item{...}{Additional arguments to be passed to or from methods.}

\item{dim}{The new dimensions to be set on the tensor.}

\item{dimnames}{Either \code{NULL} or the names of the dimensions. This must be a list with one component for each dimension, either \code{NULL} or a character vector of the length given by \code{dim} for that dimension.}

\item{order}{The order in which elements of data should be read during rearrangement.
By default, the order is equivalent to the \code{C}-style ordering and means elements should be read in row-major order.
In opposite, the \code{Fortran}-style ordering means elements should be read in column-major order.}

\item{numeric}{A logical value indicating whether the elements should be coerced as numeric elements.}

\item{reverse}{Controls the order of the elements in the (reshaped) tensor. By default, they are used in the given order, but they can also be used in reverse order.}

\item{x}{An R object.}
}
\value{
A tensor (n-dimensional array).
}
\description{
\code{tensor(data, ...)} creates a (reshaped) tensor (a n-dimensional array).\cr
  \code{as.tensor(data, ...)} attempts to turn its argument into a tensor.\cr
  \code{is.tensor(x)} tests if its argument is a tensor.\cr
}
\details{
The function \code{array_reshape} from reticulate package differs from the base function \code{dim}.
  While \code{dim} will fill new dimensions in column-major (Fortran-style) ordering, \code{array_reshape} allows both row-major (C-style) ordering and column-major (Fortran-style) ordering.

  \code{tensor} creates a tensor.

  \code{as.tensor} is a generic function. The default method will return a recreated tensor.

  \code{is.tensor} returns \code{TRUE} or \code{FALSE} depending on whether its argument is a tensor.
}
\seealso{
\code{\link{dim}}, \code{\link[reticulate]{array_reshape}}.

Other Utils: 
\code{\link{as_ANN_matrix}()},
\code{\link{as_tensor_1D}()},
\code{\link{as_tensor_2D}()},
\code{\link{as_tensor_3D}()},
\code{\link{degree}()},
\code{\link{distance}()},
\code{\link{flatten}()},
\code{\link{nsamples}()},
\code{\link{nsubsequences}()},
\code{\link{ntimesteps}()},
\code{\link{numpy}()},
\code{\link{nunits}()},
\code{\link{probability}()},
\code{\link{radian}()},
\code{\link{sd_pop}()},
\code{\link{similarity}()},
\code{\link{var_pop}()},
\code{\link{vector_as_ANN_matrix}()},
\code{\link{vector_as_numeric}()}
}
\concept{Utils}
