% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepUtils.r
\name{vander}
\alias{vander}
\title{Multidimensional array creation: Vandermonde matrix}
\usage{
vander(data, n)
}
\arguments{
\item{data}{The data to be reshaped to a Vandermonde matrix.}

\item{n}{The number of columns of the resulting matrix.}
}
\value{
A Vandermonde matrix.
}
\description{
Multidimensional array creation: Vandermonde matrix
}
\details{
The Python package NumPy offers an analog array creation function \code{np.vander(x, n)}.
}
\seealso{
Other Utils: 
\code{\link{as_ANN_matrix}()},
\code{\link{as_tensor_1d}()},
\code{\link{as_tensor_2d}()},
\code{\link{as_tensor_3d}()},
\code{\link{degree}()},
\code{\link{distance}()},
\code{\link{eye}()},
\code{\link{flatten}()},
\code{\link{insert}()},
\code{\link{list_as_numeric}()},
\code{\link{marray}()},
\code{\link{mbind}()},
\code{\link{ndim}()},
\code{\link{nsamples}()},
\code{\link{nsubsequences}()},
\code{\link{ntimesteps}()},
\code{\link{nunits}()},
\code{\link{probability}()},
\code{\link{radian}()},
\code{\link{random_seed}()},
\code{\link{re.factor}()},
\code{\link{sd_pop}()},
\code{\link{similarity}()},
\code{\link{slice}()},
\code{\link{squeeze}()},
\code{\link{ta}()},
\code{\link{tensor}()},
\code{\link{var_pop}()},
\code{\link{vector_as_ANN_matrix}()},
\code{\link{vector_as_numeric}()}
}
\concept{Utils}
