% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepML.r
\name{var.pop}
\alias{var.pop}
\title{Population variance}
\usage{
var.pop(x, y = NULL, na.rm = FALSE, use)
}
\arguments{
\item{x}{A numeric vector, matrix or data frame.}

\item{y}{\code{NULL} (default) or a vector, matrix or data frame with compatible dimensions to x. The default is equivalent to \code{y = x} (but more efficient).}

\item{na.rm}{A logical value indicating whether missing values should be removed or not (default).}

\item{use}{An optional character string giving a method for computing covariances in the presence of missing values.
This must be (an abbreviation of) one of the strings \code{everything}, \code{all.obs}, \code{complete.obs}, \code{na.or.complete}, or \code{pairwise.complete.obs}.}
}
\value{
The population variance.
}
\description{
Population variance
}
\details{
The population variance and sample variance, implemented in \code{stats} package, differ in the denominator.
  The value of denominator in the formula for variance in case of population data is \code{n}, while it is \code{n-1} for sample data.
}
\examples{
  x <- sample(1000)
  var.pop(x)
}
\seealso{
\code{\link[stats]{cor}}.

Other Machine Learning: 
\code{\link{as.degree}()},
\code{\link{as.radian}()},
\code{\link{cross_validation_split}()},
\code{\link{distance}()},
\code{\link{erfcinv}()},
\code{\link{erfc}()},
\code{\link{erfinv}()},
\code{\link{erf}()},
\code{\link{k_nearest_neighbors}()},
\code{\link{moving_average}()},
\code{\link{naive_forecast}()},
\code{\link{predict.kmeans}()},
\code{\link{sd.pop}()},
\code{\link{similarity}()}
}
\concept{Machine Learning}
