% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepUtils.r
\name{var_pop}
\alias{var_pop}
\title{Population variance}
\usage{
var_pop(x, y = NULL, na.rm = FALSE, use)
}
\arguments{
\item{x}{A numeric vector, matrix or data frame.}

\item{y}{\code{NULL} (default) or a vector, matrix or data frame with compatible dimensions to x. The default is equivalent to \code{y = x} (but more efficient).}

\item{na.rm}{A logical value indicating whether missing values should be removed or not (default).}

\item{use}{An optional character string giving a method for computing covariances in the presence of missing values.
This must be (an abbreviation of) one of the strings \code{everything}, \code{all.obs}, \code{complete.obs}, \code{na.or.complete}, or \code{pairwise.complete.obs}.}
}
\value{
The population variance.
}
\description{

}
\details{
The population variance and sample variance, implemented in \code{stats} package, differ in the denominator.
  The value of denominator in the formula for variance in case of population data is \code{n}, while it is \code{n-1} for sample data.
}
\examples{
  x <- sample(1000)
  var_pop(x)
}
\seealso{
\code{\link[stats]{cor}}.

Other Utils: 
\code{\link{as_ANN_matrix}()},
\code{\link{as_tensor_1d}()},
\code{\link{as_tensor_2d}()},
\code{\link{as_tensor_3d}()},
\code{\link{degree}()},
\code{\link{distance}()},
\code{\link{eye}()},
\code{\link{flatten}()},
\code{\link{insert}()},
\code{\link{list_as_numeric}()},
\code{\link{marray}()},
\code{\link{mbind}()},
\code{\link{ndim}()},
\code{\link{nsamples}()},
\code{\link{nsubsequences}()},
\code{\link{ntimesteps}()},
\code{\link{nunits}()},
\code{\link{probability}()},
\code{\link{radian}()},
\code{\link{random_seed}()},
\code{\link{re.factor}()},
\code{\link{sd_pop}()},
\code{\link{similarity}()},
\code{\link{slice}()},
\code{\link{ta}()},
\code{\link{tensor}()},
\code{\link{vander}()},
\code{\link{vector_as_ANN_matrix}()},
\code{\link{vector_as_numeric}()}
}
\concept{Utils}
