% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepMLP.r
\name{vector.as.ANN.matrix}
\alias{vector.as.ANN.matrix}
\title{Transform a vector into a ANN compatible matrix}
\usage{
vector.as.ANN.matrix(
  x,
  ncol = 1,
  reverse = FALSE,
  by = c("row", "col", "step")
)
}
\arguments{
\item{x}{A numeric vector.}

\item{ncol}{The number of columns in the resulting matrix. If \code{by = step}, the number of columns is equal to the number of timesteps used for a LSTM.}

\item{reverse}{Controls the order of the values in the transformed vector \code{X}. By default they are used in the given order, but they can also be used in reverse order.}

\item{by}{Controls the transformation process. The options \code{row} and \code{col} lead to a matrix whereby the values are structured row-wise or column-wise.
The option \code{step} stands for a stepwise order of the values row by row (e.g. 1 2 3, 2 3 4, 4 5 6 etc.).}
}
\value{
The transformed or resampled vector \code{x} into a matrix.
}
\description{
Transform a vector into a ANN compatible matrix
}
\seealso{
\code{\link{as.tensor.3D}}.

Other Single & Multi Layer Perceptron (SLP, MLP): 
\code{\link{as.ANN.matrix}()},
\code{\link{as.MLP.X}()},
\code{\link{as.MLP.Y}()},
\code{\link{as.tensor.1D}()},
\code{\link{as.tensor.2D}()},
\code{\link{as.tensor.3D}()},
\code{\link{as.tensor}()},
\code{\link{build.MLP}()},
\code{\link{fit.MLP}()},
\code{\link{get.MLP.X.samples}()},
\code{\link{get.MLP.X.units}()},
\code{\link{get.MLP.Y.samples}()},
\code{\link{get.MLP.Y.units}()},
\code{\link{predict.ANN}()},
\code{\link{vector.as.numeric}()}
}
\concept{Single & Multi Layer Perceptron (SLP, MLP)}
