% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepMLP.r
\name{vector.as.numeric}
\alias{vector.as.numeric}
\title{Transform a vector to a numeric vector}
\usage{
vector.as.numeric(x, adjust = NULL)
}
\arguments{
\item{x}{A vector.}

\item{adjust}{A number that is added to or subtracted from a factor level value, or even not (\code{NULL}).}
}
\value{
The vector \code{x} as numeric vector.
}
\description{
Transform a vector to a numeric vector
}
