% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepUtils.r
\name{vector_as_numeric}
\alias{vector_as_numeric}
\title{Transform a vector to a numeric vector}
\usage{
vector_as_numeric(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
The vector \code{x} as numeric vector.
}
\description{

}
\seealso{
Other Utils: 
\code{\link{as_ANN_matrix}()},
\code{\link{as_tensor_1d}()},
\code{\link{as_tensor_2d}()},
\code{\link{as_tensor_3d}()},
\code{\link{degree}()},
\code{\link{distance}()},
\code{\link{eye}()},
\code{\link{flatten}()},
\code{\link{insert}()},
\code{\link{list_as_numeric}()},
\code{\link{marray}()},
\code{\link{mbind}()},
\code{\link{ndim}()},
\code{\link{nsamples}()},
\code{\link{nsubsequences}()},
\code{\link{ntimesteps}()},
\code{\link{nunits}()},
\code{\link{probability}()},
\code{\link{radian}()},
\code{\link{random_seed}()},
\code{\link{re.factor}()},
\code{\link{sd_pop}()},
\code{\link{similarity}()},
\code{\link{slice}()},
\code{\link{ta}()},
\code{\link{tensor}()},
\code{\link{vander}()},
\code{\link{var_pop}()},
\code{\link{vector_as_ANN_matrix}()}
}
\concept{Utils}
