% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepOutlier.r
\name{winsorize}
\alias{winsorize}
\title{Winsorize outliers}
\usage{
winsorize(x, quantile.low = 0.05)
}
\arguments{
\item{x}{A numeric vector.}

\item{quantile.low}{A lower quantile limit (default \code{0.05}). The upper quantile limit is calculated by \code{1-quantile.low}.}
}
\value{
A winsorized version of the given \code{x} vector.
}
\description{
\code{winsorize} sets extremely low or high values, so-called outliers, to quantile limits.
}
\seealso{
\code{\link[stats]{quantile}}, \code{\link{outlier}}.

Other Outlier: 
\code{\link{outlier}()}
}
\concept{Outlier}
