% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepArray.r
\name{zeros}
\alias{zeros}
\title{Multidimensional array creation: array filled with zeros}
\usage{
zeros(dim = NULL)
}
\arguments{
\item{dim}{Shape of the new array.}
}
\value{
An array of zeros with the given shape.
}
\description{
Multidimensional array creation: array filled with zeros
}
\details{
This function corresponds to \code{zeros()} from NumPy.
}
\seealso{
Other Array: 
\code{\link{column_stack}()},
\code{\link{dimC<-}()},
\code{\link{empty}()},
\code{\link{eye}()},
\code{\link{flatten}()},
\code{\link{full}()},
\code{\link{insert}()},
\code{\link{mabind}()},
\code{\link{mamatrix}()},
\code{\link{marray}()},
\code{\link{ndim}()},
\code{\link{nsize}()},
\code{\link{ones}()},
\code{\link{reshape.array}()},
\code{\link{row_stack}()},
\code{\link{slice}()},
\code{\link{squeeze}()},
\code{\link{transpose}()},
\code{\link{vander}()}
}
\concept{Array}
