% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepQuality.r
\name{accuracy}
\alias{accuracy}
\title{Classification accuracy}
\usage{
accuracy(actual, predicted)
}
\arguments{
\item{actual}{A vector, matrix or n-dimensional array with to-be values}

\item{predicted}{A vector, matrix or n-dimensional array with as-is values}
}
\value{
The fraction of right predictions within total predictions.
}
\description{
Classification accuracy
}
\details{
\eqn{Accuracy = Number of correct predictions / Total number of predictions}
}
\seealso{
Other Quality: 
\code{\link{mae}()},
\code{\link{mape}()},
\code{\link{rmse}()},
\code{\link{vc}()}
}
\concept{Quality}
