% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepMetrics.r
\name{accuracy}
\alias{accuracy}
\title{Classification accuracy}
\usage{
accuracy(
  actuals,
  preds,
  type = c("standard", "precision", "recall", "F1", "tpr", "fpr", "kappa",
    "misclassification"),
  na.rm = FALSE
)
}
\arguments{
\item{actuals}{Numeric data (vector, array, matrix, data frame or list) of actual values.}

\item{preds}{Numeric data (vector, array, matrix, data frame or list) of prediction values.}

\item{type}{Denotes the calculated type (\code{standard} (default), \code{precision}, \code{recall}, \code{F1}-score, \code{tpr}, \code{fpr}, \code{kappa} or \code{misclassification}-error of accuracy derivative from confusion matrix.}

\item{na.rm}{A logical value indicating whether actual and prediction pairs with at least one NA value should be ignored.}
}
\value{
The type-specific accuracy of a classification problem.
}
\description{

}
\details{
\eqn{Standard Accuracy = Number of correct predictions / Total number of predictions}.\cr
  \eqn{Precision = True Positives / (True Positives + False Positives)}. The denominator is the total predicted positives.\cr
  \eqn{Recall = True Positives / (True Positives + False Negatives)}. The denominator is the total actual positives.\cr
  \eqn{F1 = 2 * Precision * Recall / (Precision + Recall)}.\cr
  \eqn{TPR (True Positive Rate) or sensitivity = True Positives / (True Positives + False Negatives)}.\cr
  \eqn{FPR (False Positive Rate) or specificity = True Negatives / (True Negatives + False Positives)}.\cr
  \eqn{Kappa statistics = (p0 - pe) / (1 - pe)}.\cr
  \eqn{Misclassification error = Number of incorrect predictions / Total number of predictions}.\cr

  Standard accuracy and misclassification error are mainly used for single-label classification problems, while the others can also be used for multi-label classification problems.
}
\seealso{
Other Metrics: 
\code{\link{cross_entropy}()},
\code{\link{entropy}()},
\code{\link{erfcinv}()},
\code{\link{erfc}()},
\code{\link{erfinv}()},
\code{\link{erf}()},
\code{\link{gini_impurity}()},
\code{\link{huber_loss}()},
\code{\link{log_cosh_loss}()},
\code{\link{mae}()},
\code{\link{mape}()},
\code{\link{mse}()},
\code{\link{msle}()},
\code{\link{quantile_loss}()},
\code{\link{rmse}()},
\code{\link{rmsle}()},
\code{\link{rmspe}()},
\code{\link{sse}()},
\code{\link{vc}()},
\code{\link{wape}()},
\code{\link{wmape}()}
}
\concept{Metrics}
