% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepQuality.r
\name{accuracy}
\alias{accuracy}
\title{Classification accuracy}
\usage{
accuracy(y, yhat)
}
\arguments{
\item{y}{Numeric data (vector, array, matrix, data frame or list) with actual values (to-be values).}

\item{yhat}{Numeric data (vector, array, matrix, data frame or list) with estimated values (as-is values).}
}
\value{
The fraction of right predictions within total predictions.
}
\description{
Classification accuracy
}
\details{
\eqn{Accuracy = Number of correct predictions / Total number of predictions}
}
\seealso{
Other Quality: 
\code{\link{huber_loss}()},
\code{\link{log_cosh_loss}()},
\code{\link{mae}()},
\code{\link{mape}()},
\code{\link{mse}()},
\code{\link{quantile_loss}()},
\code{\link{rmse}()},
\code{\link{vc}()}
}
\concept{Quality}
