% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepQuality.r
\name{accuracy}
\alias{accuracy}
\title{Classification accuracy}
\usage{
accuracy(actuals, preds, na.rm = FALSE)
}
\arguments{
\item{actuals}{Numeric data (vector, array, matrix, data frame or list) of actual values.}

\item{preds}{Numeric data (vector, array, matrix, data frame or list) of prediction values.}

\item{na.rm}{A logical value indicating whether actual and prediction pairs with at least one NA value should be ignored.}
}
\value{
The fraction of right predictions within total number of predictions.
}
\description{
Classification accuracy
}
\details{
\eqn{Accuracy = Number of correct predictions / Total number of predictions}
}
\seealso{
Other Quality: 
\code{\link{gini_impurity}()},
\code{\link{huber_loss}()},
\code{\link{log_cosh_loss}()},
\code{\link{mae}()},
\code{\link{mape}()},
\code{\link{mse}()},
\code{\link{msle}()},
\code{\link{quantile_loss}()},
\code{\link{rmse}()},
\code{\link{rmsle}()},
\code{\link{vc}()}
}
\concept{Quality}
