% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepMLP.r
\name{as.ANN.matrix}
\alias{as.ANN.matrix}
\title{Convert data into an ANN compatible matrix with only numbers}
\usage{
as.ANN.matrix(data)
}
\arguments{
\item{data}{A data set, usually a matrix or data frame.}
}
\value{
A matrix with only numbers.
}
\description{
Convert data into an ANN compatible matrix with only numbers
}
\seealso{
Other Single & Multi Layer Perceptron (SLP, MLP): 
\code{\link{as.MLP.X}()},
\code{\link{as.MLP.Y}()},
\code{\link{as.tensor.1D}()},
\code{\link{as.tensor.2D}()},
\code{\link{as.tensor.3D}()},
\code{\link{as.tensor}()},
\code{\link{build.MLP}()},
\code{\link{fit.MLP}()},
\code{\link{get.MLP.X.samples}()},
\code{\link{get.MLP.X.units}()},
\code{\link{get.MLP.Y.samples}()},
\code{\link{get.MLP.Y.units}()},
\code{\link{predict.ANN}()},
\code{\link{vector.as.ANN.matrix}()},
\code{\link{vector.as.numeric}()}
}
\concept{Single & Multi Layer Perceptron (SLP, MLP)}
