% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepCNN.r
\name{as.CNN.image.Y}
\alias{as.CNN.image.Y}
\title{Create a one-hot encoded vector for image labels}
\usage{
as.CNN.image.Y(labels)
}
\arguments{
\item{labels}{The labels of the images}
}
\value{
A one-hot encoded vector for the image labels
}
\description{
Create a one-hot encoded vector for image labels
}
\seealso{
\code{\link{one_hot_encode}}

Other Convolutional Neural Network (CNN): 
\code{\link{as.CNN.image.X}()}
}
\concept{Convolutional Neural Network (CNN)}
