% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepCNN.r
\name{as.CNN.image.Y}
\alias{as.CNN.image.Y}
\title{Create a one-hot vector for image labels (output)}
\usage{
as.CNN.image.Y(labels)
}
\arguments{
\item{labels}{The labels of the images either as factors for single-label classification or as a numeric or logical matrix for multi-label classification.}
}
\value{
A one-hot encoded vector or matrix for the image labels.
}
\description{
Create a one-hot vector for image labels (output)
}
\seealso{
\code{\link{one_hot_encode}}, \code{\link{as.CNN.image.X}}

Other Convolutional Neural Network (CNN): 
\code{\link{as.CNN.image.X}()},
\code{\link{as.CNN.temp.X}()},
\code{\link{as.CNN.temp.Y}()},
\code{\link{get.CNN.image.X.channels}()},
\code{\link{get.CNN.image.X.height}()},
\code{\link{get.CNN.image.X.samples}()},
\code{\link{get.CNN.image.X.width}()},
\code{\link{get.CNN.image.Y.samples}()},
\code{\link{get.CNN.image.Y.units}()},
\code{\link{get.CNN.temp.X.samples}()},
\code{\link{get.CNN.temp.X.subsequences}()},
\code{\link{get.CNN.temp.X.timesteps}()},
\code{\link{get.CNN.temp.X.units}()},
\code{\link{get.CNN.temp.Y.samples}()},
\code{\link{get.CNN.temp.Y.timesteps}()},
\code{\link{get.CNN.temp.Y.units}()}
}
\concept{Convolutional Neural Network (CNN)}
