% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepRNN.r
\name{as.LSTM.data.frame}
\alias{as.LSTM.data.frame}
\title{Recreation of a data.frame based on preformatted X and Y data sets}
\usage{
as.LSTM.data.frame(
  X,
  Y,
  names_X,
  names_Y,
  timesteps = 1,
  reverse = FALSE,
  suffix = "_t"
)
}
\arguments{
\item{X}{A feature data set, usually a matrix or data.frame, returned by \code{get.LSTM.XY}.}

\item{Y}{An outcome data set, usually a vector, matrix or data.frame, returned by \code{get.LSTM.XY}.}

\item{names_X}{Names of the features.}

\item{names_Y}{Names of the outcomes.}

\item{timesteps}{Number of timesteps; stands for the number of different periods within one sample (record) of the result, the resampled feature matrix \code{X}.}

\item{reverse}{Controls the order of the values in the resampled feature matrix \code{X} and the resampled outcome matrix \code{Y}. By default they are used in the given order (forward in time), but they can also be used in reverse order (backward in time).}

\item{suffix}{The suffix for every feature per timestep or period.}
}
\value{
A data.frame with outcome column(s) and a further resampled feature matrix.
  The feature matrix within this data.frame has the following forward oriented form:
  x1(t1), x1(t2), x1(t3)...x2(t1), x2(t2), x2(t3)...x3(t1), x3(t2), x3(t3)...
}
\description{
Recreation of a data.frame based on preformatted X and Y data sets
}
\seealso{
\code{\link{get.LSTM.XY}}.

Other Recurrent Neural Network (RNN), Long Short-Term Memory (LSTM): 
\code{\link{as.LSTM.X}()},
\code{\link{as.LSTM.Y}()},
\code{\link{as.LSTM.period_outcome}()},
\code{\link{as.lag}()},
\code{\link{as.timesteps}()},
\code{\link{build.LSTM}()},
\code{\link{fit.LSTM}()},
\code{\link{get.LSTM.X.samples}()},
\code{\link{get.LSTM.X.timesteps}()},
\code{\link{get.LSTM.X.units}()},
\code{\link{get.LSTM.XY}()},
\code{\link{get.LSTM.Y.samples}()},
\code{\link{get.LSTM.Y.timesteps}()},
\code{\link{get.LSTM.Y.units}()},
\code{\link{get.LSTM.period_shift}()},
\code{\link{predict.LSTM}()},
\code{\link{start.LSTM.invert_differencing}()}
}
\concept{Recurrent Neural Network (RNN), Long Short-Term Memory (LSTM)}
