% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepRNN.r
\name{as.LSTM.period_outcome}
\alias{as.LSTM.period_outcome}
\title{Combination of periods and actual outcome values within a matrix for quality control or graphical representation}
\usage{
as.LSTM.period_outcome(
  dataset,
  p,
  y,
  timesteps = 1,
  lag = 0,
  type = "univariate"
)
}
\arguments{
\item{dataset}{A matrix or data frame with training or test data.}

\item{p}{The column index for the period variable.}

\item{y}{The column indices for the actual outcomes.}

\item{timesteps}{The number of timesteps; stands for the number of different periods within one sample (record) of the resampled feature matrix, returned by \code{as.LSTM.X}.}

\item{lag}{The number of considered lags on feature side.}

\item{type}{The type of time series: \code{univariate} or \code{multivariate}.}
}
\value{
A data frame with period and actual outcome values that can be used for quality assurance or
  for graphical representation together with the predicted values produced by \code{predict.LSTM}.
}
\description{
Combination of periods and actual outcome values within a matrix for quality control or graphical representation
}
\seealso{
\code{\link{as.LSTM.X}}, \code{\link{predict.LSTM}}.

Other Recurrent Neural Network (RNN): 
\code{\link{as.LSTM.X}()},
\code{\link{as.LSTM.Y}()},
\code{\link{as.LSTM.data.frame}()},
\code{\link{as.lag}()},
\code{\link{as.timesteps}()},
\code{\link{build.LSTM}()},
\code{\link{fit.LSTM}()},
\code{\link{get.LSTM.X.samples}()},
\code{\link{get.LSTM.X.timesteps}()},
\code{\link{get.LSTM.X.units}()},
\code{\link{get.LSTM.XY}()},
\code{\link{get.LSTM.Y.samples}()},
\code{\link{get.LSTM.Y.timesteps}()},
\code{\link{get.LSTM.Y.units}()},
\code{\link{get.period_shift}()},
\code{\link{predict.ANN}()},
\code{\link{start.invert_differencing}()}
}
\concept{Recurrent Neural Network (RNN)}
