% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepMLP.r
\name{as.MLP.X}
\alias{as.MLP.X}
\title{Features (X) data format for SLP/MLP}
\usage{
as.MLP.X(X)
}
\arguments{
\item{X}{A feature data set, usually a matrix or data frame.}
}
\value{
A two-dimensional array of the feature matrix \code{X}.
}
\description{
Features (X) data format for SLP/MLP
}
\seealso{
\code{\link{as.MLP.Y}}, \code{\link{as.ANN.matrix}}.

Other Single & Multi Layer Perceptron (SLP, MLP): 
\code{\link{as.ANN.matrix}()},
\code{\link{as.MLP.Y}()},
\code{\link{as.numpy}()},
\code{\link{as.tensor.1D}()},
\code{\link{as.tensor.2D}()},
\code{\link{as.tensor.3D}()},
\code{\link{as.tensor}()},
\code{\link{build.MLP}()},
\code{\link{fit.MLP}()},
\code{\link{flatten}()},
\code{\link{get.MLP.X.samples}()},
\code{\link{get.MLP.X.units}()},
\code{\link{get.MLP.Y.samples}()},
\code{\link{get.MLP.Y.units}()},
\code{\link{load_weights.ANN}()},
\code{\link{predict.ANN}()},
\code{\link{save_weights.ANN}()},
\code{\link{vector.as.ANN.matrix}()},
\code{\link{vector.as.numeric}()}
}
\concept{Single & Multi Layer Perceptron (SLP, MLP)}
