% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepMLP.r
\name{as.tensor.3D}
\alias{as.tensor.3D}
\title{Transform data into a tensor with three ranks or dimensions.}
\usage{
as.tensor.3D(data, timesteps = 1, reverse = FALSE)
}
\arguments{
\item{data}{A dataset, usually a matrix or data.frame.}

\item{reverse}{Controls the order of the values in the transformed \code{data}. By default they are used in the given order, but they can also be used in reverse order.}
}
\value{
A 3D-tensor (three-dimensional array).
}
\description{
Transform data into a tensor with three ranks or dimensions.
}
\seealso{
\code{\link{as.tensor.1D}}, \code{\link{as.tensor.2D}}.

Other Single & Multi Layer Perceptron (SLP, MLP): 
\code{\link{as.ANN.matrix}()},
\code{\link{as.MLP.X}()},
\code{\link{as.MLP.Y}()},
\code{\link{as.tensor.1D}()},
\code{\link{as.tensor.2D}()},
\code{\link{as.tensor}()},
\code{\link{as.vector.timesteps}()},
\code{\link{build.MLP}()},
\code{\link{fit.MLP}()},
\code{\link{get.MLP.X.units}()},
\code{\link{get.MLP.Y.units}()},
\code{\link{predict.MLP}()}
}
\concept{Single & Multi Layer Perceptron (SLP, MLP)}
