% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepMLP.r
\name{as.tensor}
\alias{as.tensor}
\title{Converts data to a tensor with specific rank.}
\usage{
as.tensor(data, adjust = NULL, rank = 2, timesteps = NULL, reverse = FALSE)
}
\arguments{
\item{data}{The data, usually a matrix or data.frame.}

\item{adjust}{A number that is added to or subtracted from a factor level value within \code{data}, or even not (\code{NULL}).}

\item{rank}{The rank or number of dimensions of the tensor.}

\item{timesteps}{The number of timesteps; different periods within one sample (record) of the resampled \code{data}.}

\item{reverse}{Controls the order of the values in the resampled \code{data}. By default they are used in the given order, but they can also be used in reverse order.}
}
\value{
A tensor with specific rank.
}
\description{
Converts data to a tensor with specific rank.
}
\seealso{
\code{\link{as.ANN.matrix}}.

Other Single & Multi Layer Perceptron (SLP, MLP): 
\code{\link{as.ANN.matrix}()},
\code{\link{as.MLP.X}()},
\code{\link{as.MLP.Y}()},
\code{\link{as.tensor.1D}()},
\code{\link{as.tensor.2D}()},
\code{\link{as.tensor.3D}()},
\code{\link{build.MLP}()},
\code{\link{fit.MLP}()},
\code{\link{get.MLP.X.units}()},
\code{\link{get.MLP.Y.units}()},
\code{\link{predict.MLP}()},
\code{\link{vector.as.ANN.matrix}()}
}
\concept{Single & Multi Layer Perceptron (SLP, MLP)}
