% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepMLP.r
\name{as.vector.timesteps}
\alias{as.vector.timesteps}
\title{Transform a vector into a vector with timesteps}
\usage{
\method{as.vector}{timesteps}(x, timesteps = 1, reverse = FALSE)
}
\arguments{
\item{x}{A numeric vector.}

\item{timesteps}{The number of timesteps. A timestep denotes a period within a row.}

\item{reverse}{Controls the order of the values in the transformed vector \code{X}. By default they are used in the given order, but they can also be used in reverse order.}
}
\value{
The transformed or resampled vector \code{x}.
}
\description{
Transform a vector into a vector with timesteps
}
\seealso{
\code{\link{as.tensor.3D}}.

Other Single & Multi Layer Perceptron (SLP, MLP): 
\code{\link{as.ANN.matrix}()},
\code{\link{as.MLP.X}()},
\code{\link{as.MLP.Y}()},
\code{\link{as.tensor.1D}()},
\code{\link{as.tensor.2D}()},
\code{\link{as.tensor.3D}()},
\code{\link{as.tensor}()},
\code{\link{build.MLP}()},
\code{\link{fit.MLP}()},
\code{\link{get.MLP.X.units}()},
\code{\link{get.MLP.Y.units}()},
\code{\link{predict.MLP}()}
}
\concept{Single & Multi Layer Perceptron (SLP, MLP)}
