% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepCNN.r
\name{as_CNN_image_Y}
\alias{as_CNN_image_Y}
\title{Create a one-hot vector for image labels (output)}
\usage{
as_CNN_image_Y(y, encoding = c("one_hot", "sparse"))
}
\arguments{
\item{y}{The labels of the images either as factors for single-label classification or as a numeric or logical matrix for multi-label classification.}

\item{encoding}{The type of encoding: one-hot encoding or sparse encoding.}
}
\value{
A one-hot encoded vector or matrix for the image labels.
}
\description{
Create a one-hot vector for image labels (output)
}
\seealso{
\code{\link{one_hot_encode}}, \code{\link{as_CNN_image_X}}

Other Convolutional Neural Network (CNN): 
\code{\link{as_CNN_image_X}()},
\code{\link{as_CNN_temp_X}()},
\code{\link{as_CNN_temp_Y}()},
\code{\link{as_images_array}()},
\code{\link{as_images_tensor}()},
\code{\link{build_CNN_alexnet}()},
\code{\link{build_CNN_inception_resnet_v2}()},
\code{\link{build_CNN_inception_v3}()},
\code{\link{build_CNN_lenet5}()},
\code{\link{build_CNN_mobilenet_v2}()},
\code{\link{build_CNN_mobilenet_v3}()},
\code{\link{build_CNN_mobilenet}()},
\code{\link{build_CNN_nasnet}()},
\code{\link{build_CNN_resnet50}()},
\code{\link{build_CNN_unet}()},
\code{\link{build_CNN_vgg16}()},
\code{\link{build_CNN_vgg19}()},
\code{\link{build_CNN_xception}()},
\code{\link{build_CNN_zfnet}()},
\code{\link{images_load}()},
\code{\link{images_resize}()}
}
\concept{Convolutional Neural Network (CNN)}
