% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepCNN.r
\name{as_CNN_temp_Y}
\alias{as_CNN_temp_Y}
\title{Outcomes (Y) data format for a temporal CNN}
\usage{
as_CNN_temp_Y(y, timesteps = 1L, reverse = FALSE)
}
\arguments{
\item{y}{An outcome data set, usually a vector, matrix or data frame, returned by \code{get_LSTM_XY}.}

\item{timesteps}{Number of timesteps; stands for the number of different periods within one sample (record) of the result, the resampled outcome matrix \code{y}.}

\item{reverse}{A logical value indicating the order of the values in the resampled outcome matrix \code{y}. The values can be in given order (forward in time) or in reverse order (backward in time).}
}
\value{
Dependent on the type of \code{y} and timesteps. If \code{y} is a factor, the result is a one-hot vector.
  If \code{timesteps = NULL|1} a 2D-array with the dimensions samples and number of output units, representing a scalar outcome;
  if \code{timesteps >= 2} a 3D-array with the dimensions samples, timesteps and number of output units, representing a sequence or multi-step outcome.
}
\description{
Outcomes (Y) data format for a temporal CNN
}
\seealso{
\code{\link{get_LSTM_XY}}, \code{\link{as_CNN_temp_X}}.

Other Convolutional Neural Network (CNN): 
\code{\link{as_CNN_image_X}()},
\code{\link{as_CNN_image_Y}()},
\code{\link{as_CNN_temp_X}()},
\code{\link{as_images_array}()},
\code{\link{as_images_tensor}()},
\code{\link{build_CNN_alexnet}()},
\code{\link{build_CNN_inception_resnet_v2}()},
\code{\link{build_CNN_inception_v3}()},
\code{\link{build_CNN_lenet5}()},
\code{\link{build_CNN_mobilenet_v2}()},
\code{\link{build_CNN_mobilenet_v3}()},
\code{\link{build_CNN_mobilenet}()},
\code{\link{build_CNN_nasnet}()},
\code{\link{build_CNN_resnet50}()},
\code{\link{build_CNN_unet}()},
\code{\link{build_CNN_vgg16}()},
\code{\link{build_CNN_vgg19}()},
\code{\link{build_CNN_xception}()},
\code{\link{build_CNN_zfnet}()},
\code{\link{images_load}()},
\code{\link{images_resize}()}
}
\concept{Convolutional Neural Network (CNN)}
