% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepCNN.r
\name{as_images_array}
\alias{as_images_array}
\title{Convert (resized) images to 3D arrays}
\usage{
as_images_array(imagelist, FUN, ...)
}
\arguments{
\item{imagelist}{A list of (resized) images returned by either \code{images_load()} or \code{images_resize()}.}

\item{FUN}{The function to be applied for changing the representation of the images within \code{imagelist}. If no function is specified \code{image_to_array()} from keras is called.}

\item{...}{Optional arguments to \code{FUN}.}
}
\value{
A list of images represented in 3D arrays with dimensions height, width and channels.
}
\description{
Convert (resized) images to 3D arrays
}
\examples{
  For an example see \code{\link{as_images_tensor}}.
}
\seealso{
\code{\link[keras3]{image_to_array}}.

Other Convolutional Neural Network (CNN): 
\code{\link{alexnet}()},
\code{\link{as_CNN_image_X}()},
\code{\link{as_CNN_image_Y}()},
\code{\link{as_CNN_temp_X}()},
\code{\link{as_CNN_temp_Y}()},
\code{\link{as_images_tensor}()},
\code{\link{images_load}()},
\code{\link{images_resize}()},
\code{\link{inception_resnet_v2}()},
\code{\link{inception_v3}()},
\code{\link{lenet5}()},
\code{\link{mobilenet}()},
\code{\link{mobilenet_v2}()},
\code{\link{mobilenet_v3}()},
\code{\link{nasnet}()},
\code{\link{resnet}},
\code{\link{unet}()},
\code{\link{vgg}},
\code{\link{xception}()},
\code{\link{zfnet}()}
}
\concept{Convolutional Neural Network (CNN)}
