% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepMLP.r
\name{as_MLP_Y}
\alias{as_MLP_Y}
\title{Outcomes (Y) data format for SLP/MLP}
\usage{
as_MLP_Y(y, encoding = c("one_hot", "sparse"))
}
\arguments{
\item{y}{An outcome data set, usually a vector, matrix or data frame.}

\item{encoding}{The type of encoding: one-hot encoding or sparse encoding.}
}
\value{
A two-dimensional array of the outcome \code{y}. For a factor outcome, the result is an encoded matrix.
}
\description{
Outcomes (Y) data format for SLP/MLP
}
\seealso{
\code{\link{as_MLP_X}}, \code{\link{one_hot_encode}}, \code{\link{sparse_encode}}.

Other Single & Multi Layer Perceptron (SLP, MLP): 
\code{\link{as_MLP_X}()},
\code{\link{as_tensor_1d}()},
\code{\link{as_tensor_2d}()},
\code{\link{as_tensor_3d}()},
\code{\link{build_MLP}()},
\code{\link{fit_MLP}()},
\code{\link{load_weights_ANN}()},
\code{\link{nsamples}()},
\code{\link{nsubsequences}()},
\code{\link{ntimesteps}()},
\code{\link{nunits}()},
\code{\link{predict_ANN}()},
\code{\link{save_weights_ANN}()}
}
\concept{Single & Multi Layer Perceptron (SLP, MLP)}
