% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepUtils.r
\name{as_tensor_2D}
\alias{as_tensor_2D}
\title{Transform data into a tensor with two ranks or dimensions.}
\usage{
as_tensor_2D(data, reverse = FALSE)
}
\arguments{
\item{data}{A data set, usually a matrix or data frame.}

\item{reverse}{Controls the order of the values in the transformed \code{data}. By default, they are used in the given order, but they can also be used in reverse order.}
}
\value{
A 2D-tensor (two-dimensional array equal to a matrix).
}
\description{

}
\seealso{
\code{\link{as_tensor_1D}}, \code{\link{as_tensor_3D}}.

Other Utils: 
\code{\link{as_ANN_matrix}()},
\code{\link{as_tensor_1D}()},
\code{\link{as_tensor_3D}()},
\code{\link{degree}()},
\code{\link{distance}()},
\code{\link{flatten}()},
\code{\link{nsamples}()},
\code{\link{nsubsequences}()},
\code{\link{ntimesteps}()},
\code{\link{numpy}()},
\code{\link{nunits}()},
\code{\link{probability}()},
\code{\link{radian}()},
\code{\link{sd_pop}()},
\code{\link{similarity}()},
\code{\link{tensor}()},
\code{\link{var_pop}()},
\code{\link{vector_as_ANN_matrix}()},
\code{\link{vector_as_numeric}()}
}
\concept{Utils}
