% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepMetrics.r
\name{coerce_dimension}
\alias{coerce_dimension}
\title{Coerce data to an array with no trailing dimension of 1 or to a matrix}
\usage{
coerce_dimension(x)
}
\arguments{
\item{x}{A data structure like vector, matrix, array, data frame or list.}
}
\value{
The coerced data structure, either an array with no trailing dimension of 1 or a matrix.
}
\description{
Coerce data to an array with no trailing dimension of 1 or to a matrix
}
