% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepML.r
\name{cross_validation_split}
\alias{cross_validation_split}
\title{K-fold cross validation}
\usage{
cross_validation_split(dataset, folds = 3, foldname = "fold", shuffle = FALSE)
}
\arguments{
\item{dataset}{A data set, usually a data frame.}

\item{folds}{Number of created folds.}

\item{foldname}{A prefix name for each created fold followed by a number, e.g. 1,2,3.}

\item{shuffle}{Controls whether the samples of the data set should be randomly shuffled before fold creation.
For time series data, this argument must be set equal to \code{FALSE} because the order of the samples can't be changed.}
}
\value{
A named list with folds.
}
\description{
\code{cross_validation} splits a data set in partial sets, so-called folds, and creates a list of folds.
}
\seealso{
Other Machine Learning: 
\code{\link{euclidean_distance}()},
\code{\link{k_nearest_neighbors}()},
\code{\link{naive_forecast}()}
}
\concept{Machine Learning}
