% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepTimeSeries.r
\name{diff.log}
\alias{diff.log}
\title{Log-differentiation of a numeric vector}
\usage{
\method{diff}{log}(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
The log-differentiation of \code{x}.
}
\description{
Log-differentiation of a numeric vector
}
\seealso{
\code{\link{build.stationary}}, \code{\link{diffinv.log}}.

Other TimeSeries: 
\code{\link{build.stationary}()},
\code{\link{diff.percentage}()},
\code{\link{diffinv.log}()},
\code{\link{diffinv.percentage}()},
\code{\link{diffinv.simple}()},
\code{\link{get.season}()},
\code{\link{invert_differencing}()},
\code{\link{lags}()},
\code{\link{period}()}
}
\concept{TimeSeries}
