% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepTimeSeries.r
\name{diff.percentage}
\alias{diff.percentage}
\title{Percentage-differencing of a numeric vector}
\usage{
\method{diff}{percentage}(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
The percentage-differenced \code{x}.
}
\description{
Percentage-differencing of a numeric vector
}
\seealso{
\code{\link{build.stationary}}, \code{\link{diffinv.percentage}}.

Other Time Series: 
\code{\link{build.stationary}()},
\code{\link{diff.log}()},
\code{\link{diffinv.log}()},
\code{\link{diffinv.percentage}()},
\code{\link{diffinv.simple}()},
\code{\link{get.season}()},
\code{\link{invert_differencing}()},
\code{\link{lags}()},
\code{\link{period}()}
}
\concept{Time Series}
