% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepTimeSeries.r
\name{diffinv.log}
\alias{diffinv.log}
\title{Invert a log-differentiated vector}
\usage{
\method{diffinv}{log}(delta, origin)
}
\arguments{
\item{delta}{A log-differentiated numeric vector.}

\item{origin}{A scalar or numeric vector with original value(s) to invert the differentiated data series.}
}
\value{
The inverted log-differentiated \code{delta}.
}
\description{
Invert a log-differentiated vector
}
\seealso{
\code{\link{diff.log}}.

Other TimeSeries: 
\code{\link{build.stationary}()},
\code{\link{diff.log}()},
\code{\link{diff.percentage}()},
\code{\link{diffinv.percentage}()},
\code{\link{diffinv.simple}()},
\code{\link{get.season}()},
\code{\link{invert_differencing}()},
\code{\link{lags}()},
\code{\link{period}()}
}
\concept{TimeSeries}
