% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepTimeSeries.r
\name{diffinv.percentage}
\alias{diffinv.percentage}
\title{Invert a percentage-differenced vector}
\usage{
\method{diffinv}{percentage}(delta, origin)
}
\arguments{
\item{delta}{A percentage-differenced numeric vector.}

\item{origin}{A scalar or numeric vector with original value(s) to invert \code{delta}.}
}
\value{
The inverted \code{delta}.
}
\description{
Invert a percentage-differenced vector
}
\seealso{
\code{\link{diff.percentage}}.

Other Time Series: 
\code{\link{build.stationary}()},
\code{\link{diff.log}()},
\code{\link{diff.percentage}()},
\code{\link{diffinv.log}()},
\code{\link{diffinv.simple}()},
\code{\link{get.season}()},
\code{\link{invert_differencing}()},
\code{\link{lags}()},
\code{\link{period}()}
}
\concept{Time Series}
