% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepTimeSeries.r
\name{diffinv.simple}
\alias{diffinv.simple}
\title{Invert a simple-differenced vector}
\usage{
\method{diffinv}{simple}(delta, origin)
}
\arguments{
\item{delta}{A simple-differenced numeric vector.}

\item{origin}{A scalar or numeric vector with original value(s) to invert \code{delta}.}
}
\value{
A vector whose elements are the cumulative sums of \code{delta} and \code{origin}.
}
\description{
Invert a simple-differenced vector
}
\seealso{
\code{\link{build.stationary}}, \code{\link[base]{cumsum}}.

Other Time Series: 
\code{\link{build.stationary}()},
\code{\link{diff.log}()},
\code{\link{diff.percentage}()},
\code{\link{diffinv.log}()},
\code{\link{diffinv.percentage}()},
\code{\link{get.season}()},
\code{\link{invert_differencing}()},
\code{\link{lags}()},
\code{\link{period}()}
}
\concept{Time Series}
