% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepArray.r
\name{dimC<-}
\alias{dimC<-}
\title{Dimensions of an object}
\usage{
dimC(x) <- value
}
\arguments{
\item{x}{An object to get or set dimensions on.}

\item{value}{An integerish vector of new dimensions.}
}
\value{
The (redimensioned) object \code{x}.
}
\description{
Retrieve or set the dimensions of an object in row-major order.
}
\references{
Implementation credits go to \url{https://github.com/t-kalinowski/listarrays}.
}
\seealso{
Other Array: 
\code{\link{DIM}()},
\code{\link{column_stack}()},
\code{\link{delete}()},
\code{\link{embedseries}()},
\code{\link{empty}()},
\code{\link{expand_dims}()},
\code{\link{eye}()},
\code{\link{flatten}()},
\code{\link{flip}()},
\code{\link{full}()},
\code{\link{insert}()},
\code{\link{mabind}()},
\code{\link{mamatrix}()},
\code{\link{marray}()},
\code{\link{ndim}()},
\code{\link{nsize}()},
\code{\link{ones}()},
\code{\link{reshape.array}()},
\code{\link{rot90}()},
\code{\link{row_stack}()},
\code{\link{slice}()},
\code{\link{squeeze}()},
\code{\link{transpose}()},
\code{\link{vander}()},
\code{\link{zeros}()}
}
\concept{Array}
