% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepML.r
\name{distance}
\alias{distance}
\title{Distance}
\usage{
distance(x1, x2, type = c("euclidean", "squared_euclidean"))
}
\arguments{
\item{x1}{A numeric vector.}

\item{x2}{A numeric vector.}
}
\value{
The distance between \code{x1} and \code{x2}.
}
\description{
Distance
}
\examples{
  x1 <- c(20, 1, 41, 13, 5, 69)
  x2 <- c(11, 2, 23, 4, 10, 67)
  distance(x1, x2)
}
\seealso{
Other Machine Learning: 
\code{\link{cross_validation_split}()},
\code{\link{erfcinv}()},
\code{\link{erfc}()},
\code{\link{erfinv}()},
\code{\link{erf}()},
\code{\link{k_nearest_neighbors}()},
\code{\link{moving_average}()},
\code{\link{naive_forecast}()},
\code{\link{predict.kmeans}()},
\code{\link{similarity}()}
}
\concept{Machine Learning}
