% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepUtils.r
\name{distance}
\alias{distance}
\title{Distance}
\usage{
distance(
  x,
  y,
  type = c("euclidean", "squared euclidean", "manhattan", "minkowski", "chebyshev",
    "jaccard", "soergel", "geographical"),
  ...
)
}
\arguments{
\item{x}{A numeric vector.}

\item{y}{A numeric vector.}

\item{type}{The type of the distance measure.}

\item{...}{Optional arguments.}
}
\value{
The distance between \code{x} and \code{y}.
}
\description{

}
\details{
The following types of distance measures are implemented:
\itemize{
\item Euclidean: \eqn{\sqrt(\sum(x_i - y_i)^2)}
\item Squared Euclidean: \eqn{\sum(x_i - y_i)^2}
\item Manhattan: \eqn{\sum |x_i - y_i|}
\item Minkowski: \eqn{(\sum |x_i - y_i|^p)^(1/p)}
\item Chebyshev: \eqn{max |x_i - y_i|}
\item Jaccard: \eqn{1 - Jaccard similarity}
\item Soergel: \eqn{1 - Ruzicka similarity}
\item Geographical distance based on longitude and latitude values expected for both \code{x} and \code{y}. Usually longitude and latitude values are given in degree, an automatically conversion to radian is made.
}
}
\examples{
  # Euclidean distance
  x <- c(20, 1, 41, 13, 5, 69)
  y <- c(11, 2, 23, 4, 10, 67)
  distance(x, y)

  # Geographical distance
  geo_coord <- c("longitude", "latitude")
  regensburg <- setNames(c(49.013432, 12.101624), geo_coord)
  kiel <- setNames(c(54.323292, 10.122765), geo_coord)
  distance(regensburg, kiel, type = "geographical")
}
\seealso{
Other Utils: 
\code{\link{as_ANN_matrix}()},
\code{\link{as_tensor_1D}()},
\code{\link{as_tensor_2D}()},
\code{\link{as_tensor_3D}()},
\code{\link{degree}()},
\code{\link{flatten}()},
\code{\link{insert}()},
\code{\link{list_as_numeric}()},
\code{\link{marray}()},
\code{\link{mbind}()},
\code{\link{ndim}()},
\code{\link{nsamples}()},
\code{\link{nsubsequences}()},
\code{\link{ntimesteps}()},
\code{\link{nunits}()},
\code{\link{probability}()},
\code{\link{radian}()},
\code{\link{random_seed}()},
\code{\link{re.factor}()},
\code{\link{sd_pop}()},
\code{\link{similarity}()},
\code{\link{slice}()},
\code{\link{ta}()},
\code{\link{tensor}()},
\code{\link{var_pop}()},
\code{\link{vector_as_ANN_matrix}()},
\code{\link{vector_as_numeric}()}
}
\concept{Utils}
