% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepDummy.r
\name{dummify}
\alias{dummify}
\title{Dummyfication of categorical (nominal or ordinal) variables}
\usage{
dummify(
  dataset,
  columns = NULL,
  remove_level = c("first", "last", "most", "least", "none"),
  effectcoding = FALSE,
  remove_columns = FALSE
)
}
\arguments{
\item{dataset}{A data set with factor and/or character variables.}

\item{columns}{The column names of the variables which shell be dummyfied; if \code{NULL} all corresponding columns are encoded.}

\item{remove_level}{Controls which level of a factor or character variable is removed.
\code{first} removes the first level.
\code{last} removes the last level.
\code{most} removes the level with the most occurrences within the samples.
\code{least} removes the level with the least occurrences within the samples.
\code{none} removes none of the levels.}

\item{effectcoding}{Instead of using default 0/1 value pairs for dummy variables, effectcoding allows to set -1/1 pairs.}

\item{remove_columns}{Controls whether the factor/character variables shell be removed from the data set after they are encoded in dummy variables.}
}
\value{
The data set with encoded dummy variables.
}
\description{
\code{dummify} encodes non-metric variables within a data set.
}
\seealso{
\code{\link{effectcoding}}.

Other Dummyfication: 
\code{\link{effectcoding}()},
\code{\link{one_hot_decode}()},
\code{\link{one_hot_encode}()},
\code{\link{remove_columns}()},
\code{\link{resample.imbalanced}()}
}
\concept{Dummyfication}
