% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepDummy.r
\name{effectcoding}
\alias{effectcoding}
\title{Effectcoding}
\usage{
effectcoding(x)
}
\arguments{
\item{x}{An already binary encoded variable with 0/1 value pairs.}
}
\value{
\code{x} encoded with -1/1 value pairs.
}
\description{
Effectcoding
}
\references{
\url{http://www.faqs.org/faqs/ai-faq/neural-nets/part2/}.
}
\seealso{
\code{\link{dummify}}.

Other Dummifying: 
\code{\link{append_rows}()},
\code{\link{dummify}()},
\code{\link{dummify_multilabel}()},
\code{\link{one_hot_decode}()},
\code{\link{one_hot_encode}()},
\code{\link{remove_columns}()},
\code{\link{resample_imbalanced}()},
\code{\link{sparse_encode}()}
}
\concept{Dummifying}
