% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepMetrics.r
\name{entropy}
\alias{entropy}
\title{Shannon entropy}
\usage{
entropy(x, base = NULL)
}
\arguments{
\item{x}{A vector of values, usually character labels as raw instances or as class frequencies.}

\item{base}{A positive or complex number: the base with respect to which logarithms are computed.
Defaults to \code{NULL} indicates that the base is automatically determined by the number of class levels of \code{x}.}
}
\value{
Entropy.
}
\description{
Shannon entropy
}
\details{
Shannon entropy is a concept from information theory and represents a quantification of the level
of impurity or randomness that exists within a partition with class levels of \code{x}.
}
\examples{
  entropy(c("no", "no", "yes", "yes", "yes", "no", "yes", "no", "yes", "yes", "yes", "yes", "yes", "no"))
  entropy(c("no" = 5, "yes" = 9))
}
\seealso{
Other Metrics: 
\code{\link{accuracy}()},
\code{\link{cross_entropy}()},
\code{\link{dice}()},
\code{\link{erf}()},
\code{\link{erfc}()},
\code{\link{erfcinv}()},
\code{\link{erfinv}()},
\code{\link{gini_impurity}()},
\code{\link{huber_loss}()},
\code{\link{iou}()},
\code{\link{log_cosh_loss}()},
\code{\link{mae}()},
\code{\link{mape}()},
\code{\link{mse}()},
\code{\link{msle}()},
\code{\link{quantile_loss}()},
\code{\link{rmse}()},
\code{\link{rmsle}()},
\code{\link{rmspe}()},
\code{\link{sse}()},
\code{\link{stderror}()},
\code{\link{vc}()},
\code{\link{wape}()},
\code{\link{wmape}()}
}
\concept{Metrics}
