% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepML.r
\name{erf}
\alias{erf}
\title{Error function (from MATLAB)}
\usage{
erf(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
Error function as the integral of the Gaussian distribution with 0 mean and variance 1/2.
}
\description{
Error function (from MATLAB)
}
\seealso{
Other Machine Learning: 
\code{\link{cross_validation_split}()},
\code{\link{distance}()},
\code{\link{erfcinv}()},
\code{\link{erfc}()},
\code{\link{erfinv}()},
\code{\link{k_nearest_neighbors}()},
\code{\link{moving_average}()},
\code{\link{naive_forecast}()},
\code{\link{predict.kmeans}()},
\code{\link{sd.pop}()},
\code{\link{similarity}()},
\code{\link{var.pop}()}
}
\concept{Machine Learning}
