% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepMetrics.r
\name{erf}
\alias{erf}
\title{Error function (from MATLAB)}
\usage{
erf(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
Error function as the integral of the Gaussian distribution with 0 mean and variance 1/2.
}
\description{
Error function (from MATLAB)
}
\seealso{
Other Metrics: 
\code{\link{accuracy}()},
\code{\link{cross_entropy}()},
\code{\link{dice}()},
\code{\link{entropy}()},
\code{\link{erfcinv}()},
\code{\link{erfc}()},
\code{\link{erfinv}()},
\code{\link{gini_impurity}()},
\code{\link{huber_loss}()},
\code{\link{iou}()},
\code{\link{log_cosh_loss}()},
\code{\link{mae}()},
\code{\link{mape}()},
\code{\link{mse}()},
\code{\link{msle}()},
\code{\link{quantile_loss}()},
\code{\link{rmse}()},
\code{\link{rmsle}()},
\code{\link{rmspe}()},
\code{\link{sse}()},
\code{\link{stderror}()},
\code{\link{vc}()},
\code{\link{wape}()},
\code{\link{wmape}()}
}
\concept{Metrics}
