% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepML.r
\name{erf}
\alias{erf}
\title{Error function (from MATLAB)}
\usage{
erf(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
Error function as the integral of the Gaussian distribution with 0 mean and variance 1/2.
}
\description{
Error function (from MATLAB)
}
