% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepML.r
\name{erfc}
\alias{erfc}
\title{Complementary error function (from MATLAB)}
\usage{
erfc(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
Complementary error function, defined as 1 - \code{erf}.
}
\description{
Complementary error function (from MATLAB)
}
\seealso{
Other Machine Learning: 
\code{\link{cross_validation_split}()},
\code{\link{distance}()},
\code{\link{erfcinv}()},
\code{\link{erfinv}()},
\code{\link{erf}()},
\code{\link{k_nearest_neighbors}()},
\code{\link{moving_average}()},
\code{\link{naive_forecast}()},
\code{\link{predict.kmeans}()},
\code{\link{similarity}()}
}
\concept{Machine Learning}
