% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepQuality.r
\name{erfc}
\alias{erfc}
\title{Complementary error function (from MATLAB)}
\usage{
erfc(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
Complementary error function, defined as 1 - \code{erf}.
}
\description{

}
\seealso{
Other Quality: 
\code{\link{accuracy}()},
\code{\link{cross_entropy}()},
\code{\link{entropy}()},
\code{\link{erfcinv}()},
\code{\link{erfinv}()},
\code{\link{erf}()},
\code{\link{gini_impurity}()},
\code{\link{huber_loss}()},
\code{\link{log_cosh_loss}()},
\code{\link{mae}()},
\code{\link{mape}()},
\code{\link{mse}()},
\code{\link{msle}()},
\code{\link{quantile_loss}()},
\code{\link{rmse}()},
\code{\link{rmsle}()},
\code{\link{rmspe}()},
\code{\link{vc}()}
}
\concept{Quality}
