% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepML.r
\name{erfcinv}
\alias{erfcinv}
\title{Inverse complementary error function (from MATLAB)}
\usage{
erfcinv(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
Inverse complementary error function.
}
\description{
Inverse complementary error function (from MATLAB)
}
\seealso{
Other Machine Learning: 
\code{\link{cross_validation_split}()},
\code{\link{distance}()},
\code{\link{erfc}()},
\code{\link{erfinv}()},
\code{\link{erf}()},
\code{\link{k_nearest_neighbors}()},
\code{\link{moving_average}()},
\code{\link{naive_forecast}()},
\code{\link{predict.kmeans}()},
\code{\link{similarity}()}
}
\concept{Machine Learning}
