% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepQuality.r
\name{erfcinv}
\alias{erfcinv}
\title{Inverse complementary error function (from MATLAB)}
\usage{
erfcinv(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
Inverse complementary error function.
}
\description{

}
\seealso{
Other Quality: 
\code{\link{accuracy}()},
\code{\link{cross_entropy}()},
\code{\link{entropy}()},
\code{\link{erfc}()},
\code{\link{erfinv}()},
\code{\link{erf}()},
\code{\link{gini_impurity}()},
\code{\link{huber_loss}()},
\code{\link{log_cosh_loss}()},
\code{\link{mae}()},
\code{\link{mape}()},
\code{\link{mse}()},
\code{\link{msle}()},
\code{\link{quantile_loss}()},
\code{\link{rmse}()},
\code{\link{rmsle}()},
\code{\link{vc}()}
}
\concept{Quality}
