% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepML.r
\name{erfinv}
\alias{erfinv}
\title{Inverse error function (from MATLAB)}
\usage{
erfinv(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
Inverse error function.
}
\description{
Inverse error function (from MATLAB)
}
\seealso{
Other Machine Learning: 
\code{\link{cross_validation_split}()},
\code{\link{erfcinv}()},
\code{\link{erfc}()},
\code{\link{erf}()},
\code{\link{euclidean_distance}()},
\code{\link{k_nearest_neighbors}()},
\code{\link{moving_average}()},
\code{\link{naive_forecast}()}
}
\concept{Machine Learning}
