% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepML.r
\name{euclidean_distance}
\alias{euclidean_distance}
\title{Euclidean distance}
\usage{
euclidean_distance(x1, x2)
}
\arguments{
\item{x1}{A numerical vector.}

\item{x2}{A numerical vector.}
}
\value{
The euclidean distance between these two vectors.
}
\description{
\code{euclidean_distance} calculates the euclidean distance between two vectors (or scalars).
}
\seealso{
Other Machine Learning: 
\code{\link{cross_validation_split}()},
\code{\link{k_nearest_neighbors}()},
\code{\link{naive_forecast}()}
}
\concept{Machine Learning}
