% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepArray.r
\name{expand_dims}
\alias{expand_dims}
\title{Expand the shape of an array}
\usage{
expand_dims(a, axis = -1L)
}
\arguments{
\item{a}{An array.}

\item{axis}{Index position of the new axis in the expanded array. Negative numbers count from the back.}
}
\value{
The expanded array \code{a} with new shape.
}
\description{
Insert a new axis that will appear at the axis position in the expanded array shape.
}
\details{
This function corresponds to \code{expand_dims()} from NumPy.
}
\references{
Implementation credits go to \url{https://github.com/t-kalinowski/listarrays}.
}
\seealso{
Other Array: 
\code{\link{column_stack}()},
\code{\link{dimC<-}()},
\code{\link{empty}()},
\code{\link{eye}()},
\code{\link{flatten}()},
\code{\link{flip}()},
\code{\link{full}()},
\code{\link{insert}()},
\code{\link{mabind}()},
\code{\link{mamatrix}()},
\code{\link{marray}()},
\code{\link{ndim}()},
\code{\link{nsize}()},
\code{\link{ones}()},
\code{\link{reshape.array}()},
\code{\link{rot90}()},
\code{\link{row_stack}()},
\code{\link{slice}()},
\code{\link{squeeze}()},
\code{\link{transpose}()},
\code{\link{vander}()},
\code{\link{zeros}()}
}
\concept{Array}
