% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepArray.r
\name{eye}
\alias{eye}
\title{Create 2D identity matrix}
\usage{
eye(n, m = NULL)
}
\arguments{
\item{n}{The number of rows.}

\item{m}{The number of columns, default \code{NULL}.}
}
\value{
An identity matrix with n rows and n or rather m columns.
}
\description{
Create 2D identity matrix
}
\details{
This function corresponds to \code{eye()} from NumPy.
}
\seealso{
Other Array: 
\code{\link{column_stack}()},
\code{\link{dimC<-}()},
\code{\link{empty}()},
\code{\link{expand_dims}()},
\code{\link{flatten}()},
\code{\link{flip}()},
\code{\link{full}()},
\code{\link{insert}()},
\code{\link{mabind}()},
\code{\link{mamatrix}()},
\code{\link{marray}()},
\code{\link{ndim}()},
\code{\link{nsize}()},
\code{\link{ones}()},
\code{\link{reshape.array}()},
\code{\link{rot90}()},
\code{\link{row_stack}()},
\code{\link{slice}()},
\code{\link{squeeze}()},
\code{\link{transpose}()},
\code{\link{vander}()},
\code{\link{zeros}()}
}
\concept{Array}
