% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepUtils.r
\name{eye}
\alias{eye}
\title{Multidimensional array creation: 2D identity matrix}
\usage{
eye(n, m = NULL)
}
\arguments{
\item{n}{The number of rows.}

\item{m}{The number of columns, default \code{NULL}.}
}
\value{
An identity matrix with n rows and n or rather m columns.
}
\description{
Multidimensional array creation: 2D identity matrix
}
\details{
The Python package NumPy offers an analog array creation function \code{np.eye(n, m)}.
}
\seealso{
Other Utils: 
\code{\link{as_ANN_matrix}()},
\code{\link{as_tensor_1d}()},
\code{\link{as_tensor_2d}()},
\code{\link{as_tensor_3d}()},
\code{\link{degree}()},
\code{\link{dimC<-}()},
\code{\link{distance}()},
\code{\link{expand_dims}()},
\code{\link{flatten}()},
\code{\link{insert}()},
\code{\link{list_as_numeric}()},
\code{\link{mabind}()},
\code{\link{mbind}()},
\code{\link{ndim}()},
\code{\link{nsamples}()},
\code{\link{nsubsequences}()},
\code{\link{ntimesteps}()},
\code{\link{nunits}()},
\code{\link{probability}()},
\code{\link{radian}()},
\code{\link{random_seed}()},
\code{\link{re.factor}()},
\code{\link{reshape.array}()},
\code{\link{sd_pop}()},
\code{\link{similarity}()},
\code{\link{slice}()},
\code{\link{squeeze}()},
\code{\link{ta}()},
\code{\link{tensor}()},
\code{\link{vander}()},
\code{\link{var_pop}()},
\code{\link{vector_as_ANN_matrix}()},
\code{\link{vector_as_numeric}()}
}
\concept{Utils}
