% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepRNN.r
\name{fit.LSTM}
\alias{fit.LSTM}
\title{Fit LSTM model}
\usage{
fit.LSTM(
  X,
  Y,
  timesteps = 1,
  epochs = 100,
  batch_size = c(1, FALSE),
  validation_split = 0.2,
  k.fold = NULL,
  k.optimizer = NULL,
  hidden = NULL,
  dropout = NULL,
  output.activation = "linear",
  stateful = FALSE,
  return_sequences = FALSE,
  loss = "mean_squared_error",
  optimizer = "adam",
  metrics = c("mean_absolute_error")
)
}
\arguments{
\item{X}{A feature data set, usually a matrix or data.frame, returned by \code{get.LSTM.XY}.}

\item{Y}{n outcome data set, usually a vector, matrix or data.frame, returned by \code{get.LSTM.XY}.}

\item{timesteps}{The number of timesteps; stands for the number of different periods within one sample (record) of the resampled feature matrix, returned by \code{as.LSTM.X}.}

\item{epochs}{The number of epochs.}

\item{batch_size}{A vector with two elements. The first element holds the batch size, the number of samples used per gradient update.
The second element is boolean to indicate whether the batch size is used for input layer too (\code{TRUE}).}

\item{validation_split}{Fraction of the training data used as validation data.}

\item{k.fold}{Number of folds within k-fold cross validation or \code{NULL} if no grid search is desired.}

\item{k.optimizer}{Either \code{min} or \code{max} to indicate which type of quality measuring is used; if \code{NULL} no quality measure is extracted.}

\item{hidden}{A data.frame with two columns whereby the first column contains the number of hidden units 
and the second column the activation function. The number of rows determines the number of hidden layers.}

\item{dropout}{A numerical vector with dropout rates, the fractions of input units to drop or \code{NULL} if no dropout is desired.}

\item{output.activation}{A name of the output activation function.}

\item{stateful}{A boolean that indicates whether the last cell state of a LSTM unit at t-1 is used as initial cell state of the unit at period t (\code{TRUE}).}

\item{return_sequences}{A boolean that indicates whether an outcome unit produces one value (\code{FALSE}) or values per each timestep (\code{TRUE}).}

\item{loss}{Name of objective function or objective function. If the model has multiple outputs, 
different loss on each output can be used by passing a dictionary or a list of objectives.
The loss value that will be minimized by the model will then be the sum of all individual losses.}

\item{optimizer}{Name of optimizer or optimizer instance.}

\item{metrics}{Vector or list of metrics to be evaluated by the model during training and testing.}
}
\value{
A list with named elements
  \code{hyperparamter}: A list with named elements \code{features} and \code{output.units}.
  \code{model}: A trained model object with stacked layers.
  \code{avg_qual}: Only if k-fold cross validation is used. A data.frame with two columns whereby the 
                   first columns stores the epoch number and the second column the mean of the underpinned quality metric.
}
\description{
\code{fit.LSTM} is a wrapper function for building and fitting a LSTM model.
}
\seealso{
\code{\link{build.LSTM}}, \code{\link{get.LSTM.XY}},
  \code{\link[keras]{compile.keras.engine.training.Model}}, \code{\link[keras]{fit.keras.engine.training.Model}}.

Other Recurrent Neural Network (RNN), Long Short-Term Memory (LSTM): 
\code{\link{as.LSTM.X}()},
\code{\link{as.LSTM.Y}()},
\code{\link{as.LSTM.data.frame}()},
\code{\link{as.LSTM.period_outcome}()},
\code{\link{as.lag}()},
\code{\link{as.timesteps}()},
\code{\link{build.LSTM}()},
\code{\link{get.LSTM.X.samples}()},
\code{\link{get.LSTM.X.timesteps}()},
\code{\link{get.LSTM.X.units}()},
\code{\link{get.LSTM.XY}()},
\code{\link{get.LSTM.Y.samples}()},
\code{\link{get.LSTM.Y.timesteps}()},
\code{\link{get.LSTM.Y.units}()},
\code{\link{get.LSTM.period_shift}()},
\code{\link{predict.LSTM}()},
\code{\link{start.LSTM.invert_differencing}()}
}
\concept{Recurrent Neural Network (RNN), Long Short-Term Memory (LSTM)}
