% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepMLP.r
\name{fit.MLP}
\alias{fit.MLP}
\title{Fit SLP/MLP model}
\usage{
fit.MLP(
  model,
  X,
  Y,
  batch_size = 1,
  epochs = 10,
  verbose = 1,
  validation_split = 0,
  cross_validation = NULL
)
}
\arguments{
\item{model}{A model object to train, e.g. returned by \code{build.MLP}.}

\item{X}{A feature data set, usually a matrix or data frame.}

\item{Y}{An outcome data set, usually a vector, matrix or data frame.}

\item{batch_size}{Batch size, the number of samples per gradient update.}

\item{epochs}{Number of epochs to train the model.}

\item{verbose}{Verbosity mode (0 = silent, 1 = progress bar, 2 = one line per epoch) determines how the training progress is visualized.}

\item{validation_split}{Float between 0 and 1. Fraction of the training data used as validation data.}

\item{cross_validation}{List or \code{NULL} (default). The list contains two elements whereby the first element stands for the number of folds (k)
and the second element indicates the type \code{min} or \code{max} for quality measuring.}
}
\value{
A trained model object.
}
\description{
\code{fit.MLP} is a wrapper function for fitting a feedforward SLP or MLP.
}
\seealso{
\code{\link{build.MLP}},
  \code{\link[keras]{fit.keras.engine.training.Model}}, \code{\link[keras]{evaluate.keras.engine.training.Model}}.

Other Single & Multi Layer Perceptron (SLP, MLP): 
\code{\link{as.ANN.matrix}()},
\code{\link{as.MLP.X}()},
\code{\link{as.MLP.Y}()},
\code{\link{as.tensor.1D}()},
\code{\link{as.tensor.2D}()},
\code{\link{as.tensor.3D}()},
\code{\link{as.tensor}()},
\code{\link{build.MLP}()},
\code{\link{get.MLP.X.samples}()},
\code{\link{get.MLP.X.units}()},
\code{\link{get.MLP.Y.samples}()},
\code{\link{get.MLP.Y.units}()},
\code{\link{load_weights.ANN}()},
\code{\link{predict.ANN}()},
\code{\link{save_weights.ANN}()},
\code{\link{vector.as.ANN.matrix}()},
\code{\link{vector.as.numeric}()}
}
\concept{Single & Multi Layer Perceptron (SLP, MLP)}
