% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepSOM.r
\name{fit_SOM}
\alias{fit_SOM}
\title{Process a complete Self-Organizing Map analysis}
\usage{
fit_SOM(
  x,
  xdim = NULL,
  ydim = NULL,
  topo = c("rectangular", "hexagonal"),
  neighbourhood.fct = c("bubble", "gaussian"),
  alpha = c(0.05, 0.01),
  radius = 1,
  distance_method = c("euclidean", "maximum", "manhattan", "canberra", "binary",
    "minkowski"),
  agglomeration_method = c("ward.D", "ward.D2", "single", "complete", "average",
    "mcquitty", "median", "centroid"),
  cluster = 3L,
  colour_palette = c("#FF0000", "#00FF00", "#0000FF")
)
}
\arguments{
\item{x}{A feature data set.}

\item{xdim}{x-dimension of the grid.}

\item{ydim}{y-dimension of the grid.}

\item{topo}{Topology of the grid: \code{rectangular} or \code{hexagonal}.}

\item{neighbourhood.fct}{The type of neighbourhood function: \code{bubble} or \code{gaussian}.}

\item{alpha}{Learning rate, a vector of two numbers indicating the amount of change.}

\item{radius}{The radius of the neighborhood.}

\item{distance_method}{The distance measure to be used.}

\item{agglomeration_method}{The agglomeration method to be used.}

\item{cluster}{The number of clusters.}

\item{colour_palette}{A palette with colour codes for the different clusters.}
}
\value{
The fitted SOM model.
}
\description{

}
\seealso{
\code{\link[kohonen]{unit.distances}}, \code{\link[kohonen]{supersom}}, \code{\link[kohonen]{plot.kohonen}},
  \code{\link[stats]{dist}}, \code{\link[stats]{hclust}}.
}
\concept{Self-Organizing Map (SOM)}
