% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepArray.r
\name{flatten}
\alias{flatten}
\title{Data flattening}
\usage{
flatten(data, axis = NULL, order = c("C", "F"))
}
\arguments{
\item{data}{Data to be flatten.}

\item{axis}{The axes to be fixed while iterating over the remaining axes of \code{data}.
By default (\code{NULL}), the structure of \code{data} is interpret as a stack (of stack...) of matrices,
with either the first axis (\code{C}-order) or the second axis (\code{F}-order) and all remaining axes are fixed.}

\item{order}{The order in which elements of \code{data} should be read during flattening.
By default, the order is equivalent to the \code{C}-style ordering and means elements should be read in row-major order.
In opposite, the \code{Fortran}-style ordering means elements should be read in column-major order.}
}
\value{
The flatten data in form of a one-dimensional array.
}
\description{
Flatten data into a one-dimensional array.
}
\details{
This function corresponds to \code{ndarray.flatten()} from NumPy.
}
\examples{
v <- (1:24); dim(v); ndim(v)
l <- list(x1 = 1:10, x2 = seq(10, 100, 10), x3 = list(a = 11, b = c(2, 23)))
m <- matrix(1:24, nrow = 6); dim(m); length(m);
a3 <- array(v, dim = c(4, 3, 2)); dim(a3); length(a3)
a4 <- array(1:48, dim = c(4, 3, 2, 2))
data <- a4; data
flatten(data, order = "F"); flatten(data, order = "C")
}
\seealso{
Other Array: 
\code{\link{DIM}()},
\code{\link{column_stack}()},
\code{\link{delete}()},
\code{\link{dimC<-}()},
\code{\link{empty}()},
\code{\link{expand_dims}()},
\code{\link{eye}()},
\code{\link{flip}()},
\code{\link{full}()},
\code{\link{insert}()},
\code{\link{mabind}()},
\code{\link{mamatrix}()},
\code{\link{marray}()},
\code{\link{ndim}()},
\code{\link{nsize}()},
\code{\link{ones}()},
\code{\link{reshape.array}()},
\code{\link{rot90}()},
\code{\link{row_stack}()},
\code{\link{slice}()},
\code{\link{squeeze}()},
\code{\link{transpose}()},
\code{\link{vander}()},
\code{\link{zeros}()}
}
\concept{Array}
