% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepArray.r
\name{flip}
\alias{flip}
\title{Reverse the order of elements in an array along the given axes}
\usage{
flip(a, axis = -1L)
}
\arguments{
\item{a}{An array.}

\item{axis}{Axis or axes along which to flip over.}
}
\value{
The reversed array \code{a} along axes.
}
\description{
Reverse the order of elements in an array along the given axes
}
\details{
This function corresponds to \code{flip()} from NumPy.
}
\seealso{
Other Array: 
\code{\link{column_stack}()},
\code{\link{dimC<-}()},
\code{\link{empty}()},
\code{\link{expand_dims}()},
\code{\link{eye}()},
\code{\link{flatten}()},
\code{\link{full}()},
\code{\link{insert}()},
\code{\link{mabind}()},
\code{\link{mamatrix}()},
\code{\link{marray}()},
\code{\link{ndim}()},
\code{\link{nsize}()},
\code{\link{ones}()},
\code{\link{reshape.array}()},
\code{\link{rot90}()},
\code{\link{row_stack}()},
\code{\link{slice}()},
\code{\link{squeeze}()},
\code{\link{transpose}()},
\code{\link{vander}()},
\code{\link{zeros}()}
}
\concept{Array}
