% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepArray.r
\name{full}
\alias{full}
\title{Array creation}
\usage{
full(dim = NULL, fill_value = NA, dimnames = NULL, order = c("C", "F"))
}
\arguments{
\item{dim}{Shape of the new array.}

\item{fill_value}{Value to fill the array.}

\item{dimnames}{Either \code{NULL} or the names of the dimensions. This must be a list with one component for each dimension, either \code{NULL} or a character vector of the length given by \code{dim} for that dimension.}

\item{order}{Whether to store multidimensional array in C- or Fortran-contiguous (row- or column-wise) order.}
}
\value{
An array of \code{fill_value} with the given shape.
}
\description{
Create array filled with value.
}
\details{
This function corresponds to \code{full()} from NumPy.
}
\seealso{
Other Array: 
\code{\link{DIM}()},
\code{\link{column_stack}()},
\code{\link{delete}()},
\code{\link{dimC<-}()},
\code{\link{embedseries}()},
\code{\link{empty}()},
\code{\link{expand_dims}()},
\code{\link{eye}()},
\code{\link{flatten}()},
\code{\link{flip}()},
\code{\link{insert}()},
\code{\link{mabind}()},
\code{\link{mamatrix}()},
\code{\link{marray}()},
\code{\link{ndim}()},
\code{\link{nsize}()},
\code{\link{ones}()},
\code{\link{rearrange}()},
\code{\link{reshape.array}()},
\code{\link{rot90}()},
\code{\link{row_stack}()},
\code{\link{slice}()},
\code{\link{squeeze}()},
\code{\link{transpose}()},
\code{\link{vander}()},
\code{\link{zeros}()}
}
\concept{Array}
