% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepCNN.r
\name{images_load}
\alias{images_load}
\title{Load images from different sources like from files or web}
\usage{
images_load(images, FUN, ...)
}
\arguments{
\item{images}{A vector or list of files, urls etc. containing images.}

\item{FUN}{The function to be applied for loading \code{images}. If no function is specified \code{image_load()} from keras is called.}

\item{...}{Optional arguments to \code{FUN}.}
}
\value{
A list of images.
}
\description{
Load images from different sources like from files or web
}
\examples{
  For an example see \code{\link{as_images_tensor}}.
}
\seealso{
\code{\link[base]{list.files}}, \code{\link[keras]{image_load}}.

Other Convolutional Neural Network (CNN): 
\code{\link{alexnet}()},
\code{\link{as_CNN_image_X}()},
\code{\link{as_CNN_image_Y}()},
\code{\link{as_CNN_temp_X}()},
\code{\link{as_CNN_temp_Y}()},
\code{\link{as_images_array}()},
\code{\link{as_images_tensor}()},
\code{\link{images_resize}()},
\code{\link{inception_resnet_v2}()},
\code{\link{inception_v3}()},
\code{\link{lenet5}()},
\code{\link{mobilenet_v2}()},
\code{\link{mobilenet_v3}()},
\code{\link{mobilenet}()},
\code{\link{nasnet}()},
\code{\link{resnet}},
\code{\link{unet}()},
\code{\link{vgg}},
\code{\link{xception}()},
\code{\link{zfnet}()}
}
\concept{Convolutional Neural Network (CNN)}
