% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepCNN.r
\name{images_resize}
\alias{images_resize}
\title{Resize loaded images}
\usage{
images_resize(imagelist, FUN, ...)
}
\arguments{
\item{imagelist}{A list of loaded images returned by \code{images_load()}.}

\item{FUN}{The function to be applied for resizing images within \code{imagelist}. If no function is specified the images within \code{imagelist} aren't resized.}

\item{...}{Optional arguments to \code{FUN}.}
}
\value{
A list of (resized) images.
}
\description{

}
\examples{
  For an example see \code{\link{as_images_tensor}}.
}
\seealso{
Other Convolutional Neural Network (CNN): 
\code{\link{as_CNN_image_X}()},
\code{\link{as_CNN_image_Y}()},
\code{\link{as_CNN_temp_X}()},
\code{\link{as_CNN_temp_Y}()},
\code{\link{as_images_array}()},
\code{\link{as_images_tensor}()},
\code{\link{build_CNN_alexnet}()},
\code{\link{build_CNN_inception_resnet_v2}()},
\code{\link{build_CNN_inception_v3}()},
\code{\link{build_CNN_lenet5}()},
\code{\link{build_CNN_mobilenet_v2}()},
\code{\link{build_CNN_mobilenet_v3}()},
\code{\link{build_CNN_mobilenet}()},
\code{\link{build_CNN_nasnet}()},
\code{\link{build_CNN_resnet50}()},
\code{\link{build_CNN_vgg16}()},
\code{\link{build_CNN_vgg19}()},
\code{\link{build_CNN_xception}()},
\code{\link{build_CNN_zfnet}()},
\code{\link{images_load}()}
}
\concept{Convolutional Neural Network (CNN)}
