% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepArray.r
\name{insert}
\alias{insert}
\alias{insert.array}
\alias{insert.marray}
\title{Multidimensional array insertion}
\usage{
insert(a, ...)

\method{insert}{array}(a, x, index = dim(a)[deepANN::ndim(a)] + 1L, order = c("C", "F"))

\method{insert}{marray}(a, x, index = dim(a)[deepANN::ndim(a)] + 1L, order = c("C", "F"))
}
\arguments{
\item{a}{An array from type \code{marray}.}

\item{x}{An R object which is inserted into \code{a}.}

\item{index}{The position \code{x} should be inserted.}

\item{order}{The order in which elements of \code{x} should be read during insertion.
By default, the order is equivalent to the \code{C}-style ordering and means elements should be read in row-major order.
In opposite, the \code{Fortran}-style ordering means elements should be read in column-major order.}
}
\value{
The marray \code{a} with inserted \code{x}.
}
\description{
Multidimensional array insertion
}
\details{
For insertion, the last dimension of \code{a} is fixed and \code{x} is coerced into the remaining dimensions. If \code{a} is a
  one-dimensional array, \code{x} is flattened.
}
\examples{
  ma1 <- marray(1:24)
  ma2 <- marray(1:12, dim = c(4, 3))
  ma3 <- marray(1:24, dim = c(4, 3, 2), order = "F")
  x <- c(1:12)
  insert(ma3, x, order = "F")
}
\seealso{
\code{\link{marray}}.

Other Array: 
\code{\link{column_stack}()},
\code{\link{dimC<-}()},
\code{\link{empty}()},
\code{\link{eye}()},
\code{\link{flatten}()},
\code{\link{full}()},
\code{\link{mabind}()},
\code{\link{mamatrix}()},
\code{\link{marray}()},
\code{\link{ndim}()},
\code{\link{nsize}()},
\code{\link{ones}()},
\code{\link{reshape.array}()},
\code{\link{row_stack}()},
\code{\link{slice}()},
\code{\link{squeeze}()},
\code{\link{transpose}()},
\code{\link{vander}()},
\code{\link{zeros}()}
}
\concept{Array}
