% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepUtils.r
\name{insert}
\alias{insert}
\alias{insert.marray}
\title{Multidimensional Array (marray)}
\usage{
insert(data, ...)

\method{insert}{marray}(a, x, index = dim(a)[deepANN::ndim(a)] + 1L, order = c("C", "F"))
}
\arguments{
\item{a}{An array from type \code{marray}.}

\item{x}{An R object which is inserted into \code{a}.}

\item{index}{The position \code{x} should be inserted.}

\item{order}{The order in which elements of \code{x} should be read during insertion.
By default, the order is equivalent to the \code{C}-style ordering and means elements should be read in row-major order.
In opposite, the \code{Fortran}-style ordering means elements should be read in column-major order.}
}
\value{
The marray \code{a} with inserted \code{x}.
}
\description{
Multidimensional Array (marray)
}
\details{
For insertion, the last dimension of \code{a} is fixed and \code{x} is coerced into the remaining dimensions. If \code{a} is a
  one-dimensional array, \code{x} is flattened.
}
\examples{
  ma1 <- marray(1:24)
  ma2 <- marray(1:12, dim = c(4, 3))
  ma3 <- marray(1:24, dim = c(4, 3, 2), order = "F")
  x <- c(1:12)
  insert(ma3, x, order = "F")
}
\seealso{
\code{\link{marray}}.

Other Utils: 
\code{\link{as_ANN_matrix}()},
\code{\link{as_tensor_1D}()},
\code{\link{as_tensor_2D}()},
\code{\link{as_tensor_3D}()},
\code{\link{degree}()},
\code{\link{distance}()},
\code{\link{flatten}()},
\code{\link{list_as_numeric}()},
\code{\link{marray}()},
\code{\link{mbind}()},
\code{\link{ndim}()},
\code{\link{nsamples}()},
\code{\link{nsubsequences}()},
\code{\link{ntimesteps}()},
\code{\link{nunits}()},
\code{\link{probability}()},
\code{\link{radian}()},
\code{\link{random_seed}()},
\code{\link{re.factor}()},
\code{\link{sd_pop}()},
\code{\link{similarity}()},
\code{\link{ta}()},
\code{\link{tensor}()},
\code{\link{var_pop}()},
\code{\link{vector_as_ANN_matrix}()},
\code{\link{vector_as_numeric}()}
}
\concept{Utils}
