% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepTimeSeries.r
\name{invert_differencing}
\alias{invert_differencing}
\title{Invert a differenced data series}
\usage{
invert_differencing(delta, origin)
}
\arguments{
\item{delta}{A stationary or differenced data series.}

\item{origin}{A scalar or vector with original value(s) to invert the stationary series.}
}
\value{
A vector whose elements are the cumulative sums of \code{delta} and \code{origin}.
}
\description{
\code{invert_differencing} inverts a differenced data series.
}
\seealso{
\code{\link{build.stationary}}, \code{\link[base]{cumsum}}.

Other TimeSeries: 
\code{\link{build.stationary}()},
\code{\link{get.season}()},
\code{\link{lags}()},
\code{\link{period}()}
}
\concept{TimeSeries}
