% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepTimeSeries.r
\name{invert_differencing}
\alias{invert_differencing}
\title{Invert a differenced data series}
\usage{
invert_differencing(delta, origin, type = c("simple", "log", "percentage"))
}
\arguments{
\item{delta}{A differenced numeric vector.}

\item{origin}{A scalar or numeric vector with original value(s) to invert \code{delta}.}

\item{type}{The type of differencing to be used. Available types are \code{simple}, \code{log} and \code{percentage}.}
}
\value{
The inverted \code{delta}.
}
\description{
Invert a differenced data series
}
\seealso{
\code{\link{stationary}}.

Other Time Series: 
\code{\link{diff_log}()},
\code{\link{diff_percentage}()},
\code{\link{diffinv_log}()},
\code{\link{diffinv_percentage}()},
\code{\link{diffinv_simple}()},
\code{\link{get_season}()},
\code{\link{lags}()},
\code{\link{partition}()},
\code{\link{period}()},
\code{\link{stationary}()}
}
\concept{Time Series}
