% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepML.r
\name{k_nearest_neighbors}
\alias{k_nearest_neighbors}
\title{k-nearest neighbors}
\usage{
k_nearest_neighbors(y, X, test, k = 1L)
}
\arguments{
\item{y}{A column vector of the categories within a feature matrix or data frame.}

\item{X}{Matrix or data frame with feature values.}

\item{test}{The query or test instance.}

\item{k}{Number of samples resp. categories used to determine the majority category; default \code{1}.}
}
\value{
The majority category for a query or test instance.
}
\description{
\code{k_nearest_neighbors} calculates the majority category (class label) of k nearest neighbors of a query or test instance.
}
\examples{
  df <- data.frame(height = c(158, 158, 158, 160, 160, 163, 163, 160, 163, 165, 165, 165, 168, 168, 168, 170, 170),
                   weight = c(58, 59, 63, 59, 60, 60, 61, 64, 64, 61, 62, 65, 62, 63, 66, 63, 64),
                   size = c(rep("M", 6), rep("L", 11)))
  df$size <- as.factor(df$size)
  x <- setNames(c(161, 61), c("height", "weight")) # query instance, e.g. a new customer
  k_nearest_neighbors(df$size, df[, 1L:2L], x, k = 5L)
}
\seealso{
\code{\link{euclidean_distance}}.

Other Machine Learning: 
\code{\link{cross_validation_split}()},
\code{\link{erfcinv}()},
\code{\link{erfc}()},
\code{\link{erfinv}()},
\code{\link{erf}()},
\code{\link{euclidean_distance}()},
\code{\link{moving_average}()},
\code{\link{naive_forecast}()},
\code{\link{predict.kmeans}()}
}
\concept{Machine Learning}
