% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepML.r
\name{k_nearest_neighbors}
\alias{k_nearest_neighbors}
\title{k-nearest neighbors}
\usage{
k_nearest_neighbors(y, X, test, k = 1L)
}
\arguments{
\item{y}{A column vector of the categories within a feature matrix or data frame.}

\item{X}{Matrix or data frame with feature values.}

\item{test}{The query or test instance.}

\item{k}{Number of samples resp. categories used to determine the majority category; default \code{1}.}
}
\value{
The majority category for a query or test instance.
}
\description{
\code{k_nearest_neighbors} calculates the majority category (class label) of k nearest neighbors of a query or test instance.
}
\seealso{
\code{\link{euclidean_distance}}.

Other Machine Learning: 
\code{\link{cross_validation_split}()},
\code{\link{euclidean_distance}()},
\code{\link{naive_forecast}()}
}
\concept{Machine Learning}
