% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepTimeSeries.r
\name{lags}
\alias{lags}
\title{Build a lagged data set or series}
\usage{
lags(x, k = 1, between = FALSE, na = NA)
}
\arguments{
\item{x}{A vector for building a lagged data series or an entire data set.}

\item{k}{The amount of period shifts (lags).}

\item{between}{Controls whether only the lag for \code{k} is calculated (\code{FALSE}) or 
all lags between the periods \code{1:k} are also considered (\code{TRUE}).}

\item{na}{A sign which is used to indicate a missing value in a prior period.}
}
\value{
A lagged data series or a lagged data set.
}
\description{
\code{lags} builds a lagged data series or entire data set.
}
\seealso{
Other TimeSeries: 
\code{\link{build.stationary}()},
\code{\link{get.season}()},
\code{\link{invert_differencing}()},
\code{\link{period}()}
}
\concept{TimeSeries}
