% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepQuality.r
\name{log_cosh_loss}
\alias{log_cosh_loss}
\title{Log-Cosh loss}
\usage{
log_cosh_loss(actuals, preds, na.rm = FALSE)
}
\arguments{
\item{actuals}{A numeric vector of actual values.}

\item{preds}{A numeric vector of prediction values.}

\item{na.rm}{A logical value indicating whether actual and prediction pairs with at least one NA value should be ignored.}
}
\value{
Log-Cosh loss.
}
\description{
Log-Cosh loss
}
\seealso{
Other Quality: 
\code{\link{accuracy}()},
\code{\link{gini_impurity}()},
\code{\link{huber_loss}()},
\code{\link{mae}()},
\code{\link{mape}()},
\code{\link{mse}()},
\code{\link{msle}()},
\code{\link{quantile_loss}()},
\code{\link{rmse}()},
\code{\link{rmsle}()},
\code{\link{vc}()}
}
\concept{Quality}
