% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepUtils.r
\name{mabind}
\alias{mabind}
\title{Combine multidimensional arrays along a specified dimension}
\usage{
mabind(..., input_shape = NULL, axis = -1, order = c("C", "F"))
}
\arguments{
\item{...}{Any numbers of objects they are coerced to arrays. The objects can be packed into a \code{list}. The dimensions of these objects must be equal if they are not to be coerced into a certain \code{input_shape}.}

\item{input_shape}{The dimension the input objects are to be coerced. By default \code{NULL}, the original dimensions are used.}

\item{axis}{The dimension along the objects are combined. By default (\code{-1}), the last dimension is used for binding the arrays.}

\item{order}{The order in which elements of the objects should be read during coercing to arrays.
By default, the order is equivalent to the \code{C}-style ordering and means elements should be read in row-major order.
In opposite, the \code{Fortran}-style ordering means elements should be read in column-major order.}
}
\value{
An array as a combination of all input arrays along a specified dimension.
}
\description{
Combine multidimensional arrays along a specified dimension
}
\examples{
a1 <- marray(1:24, dim = c(4, 3, 2), order = "F"); a1
a2 <- marray(-1:-24, dim = c(4, 3, 2), order = "F"); a2
a3 <- marray(sample(24), dim = c(4, 3, 2), order = "F"); a3
mabind(a1, a2, a3) # output is an 4x3x6 array
mabind(a1, a2, a3, axis = 1) # output is an 12x3x2 array
mabind(a1, a2, a3, axis = 2) # output is an 4x9x2 array
}
\seealso{
Other Utils: 
\code{\link{as_ANN_matrix}()},
\code{\link{as_tensor_1d}()},
\code{\link{as_tensor_2d}()},
\code{\link{as_tensor_3d}()},
\code{\link{degree}()},
\code{\link{dimC<-}()},
\code{\link{distance}()},
\code{\link{expand_dims}()},
\code{\link{eye}()},
\code{\link{flatten}()},
\code{\link{insert}()},
\code{\link{list_as_numeric}()},
\code{\link{mbind}()},
\code{\link{ndim}()},
\code{\link{nsamples}()},
\code{\link{nsubsequences}()},
\code{\link{ntimesteps}()},
\code{\link{nunits}()},
\code{\link{probability}()},
\code{\link{radian}()},
\code{\link{random_seed}()},
\code{\link{re.factor}()},
\code{\link{reshape.array}()},
\code{\link{sd_pop}()},
\code{\link{similarity}()},
\code{\link{slice}()},
\code{\link{squeeze}()},
\code{\link{ta}()},
\code{\link{tensor}()},
\code{\link{vander}()},
\code{\link{var_pop}()},
\code{\link{vector_as_ANN_matrix}()},
\code{\link{vector_as_numeric}()}
}
\concept{Utils}
