% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepQuality.r
\name{mae}
\alias{mae}
\title{Mean absolute error}
\usage{
mae(y, yhat)
}
\arguments{
\item{y}{A numeric vector with actual values (to-be values).}

\item{yhat}{A numeric vector with estimated values (as-is values).}
}
\value{
Mean absolute error (MAE).
}
\description{
Mean absolute error
}
\seealso{
Other Quality: 
\code{\link{accuracy}()},
\code{\link{huber}()},
\code{\link{log_cosh}()},
\code{\link{mape}()},
\code{\link{mse}()},
\code{\link{quantile_loss}()},
\code{\link{rmse}()},
\code{\link{vc}()}
}
\concept{Quality}
