% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepQuality.r
\name{mae}
\alias{mae}
\title{Mean absolute error}
\usage{
mae(actual, predicted)
}
\arguments{
\item{actual}{A vector with to-be values.}

\item{predicted}{A vector with as-is values.}
}
\value{
Mean absolute error (MAE).
}
\description{
\code{mae} calculates the mean absolute error.
}
\seealso{
\code{\link{mape}} for mean absolute percentage error, \code{\link{mse}} for mean squared error,
  \code{\link{rmse}} for root mean square error, \code{\link{vc}} for variance coefficient.

Other Quality: 
\code{\link{mape}()},
\code{\link{rmse}()},
\code{\link{vc}()}
}
\concept{Quality}
