% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepArray.r
\name{mamatrix}
\alias{mamatrix}
\title{Array to matrix}
\usage{
mamatrix(a, order = c("C", "F"))
}
\arguments{
\item{a}{An array.}

\item{order}{The order in which elements of data should be read during combination.
By default, the order is equivalent to the \code{C}-style ordering and means elements should be read in row-major order.
In opposite, the \code{Fortran}-style ordering means elements should be read in column-major order.}

\item{...}{Additional arguments to be passed to the method.}
}
\value{
A two-dimensional array with combined matrices of the first two dimensions of \code{x}.
}
\description{
Shrinks an array by combining the respective first two dimensions of the array by columns or rows.
}
\seealso{
Other Array: 
\code{\link{column_stack}()},
\code{\link{dimC<-}()},
\code{\link{empty}()},
\code{\link{expand_dims}()},
\code{\link{eye}()},
\code{\link{flatten}()},
\code{\link{flip}()},
\code{\link{full}()},
\code{\link{insert}()},
\code{\link{mabind}()},
\code{\link{marray}()},
\code{\link{ndim}()},
\code{\link{nsize}()},
\code{\link{ones}()},
\code{\link{reshape.array}()},
\code{\link{rot90}()},
\code{\link{row_stack}()},
\code{\link{slice}()},
\code{\link{squeeze}()},
\code{\link{transpose}()},
\code{\link{vander}()},
\code{\link{zeros}()}
}
\concept{Array}
