% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepQuality.r
\name{mape}
\alias{mape}
\title{Mean absolute percentage error (MAPE)}
\usage{
mape(y, yhat)
}
\arguments{
\item{y}{A numeric vector with actual values (to-be values).}

\item{yhat}{A numeric vector with estimated values (as-is values).}
}
\value{
Mean absolute percentage error.
}
\description{
Mean absolute percentage error (MAPE)
}
\seealso{
Other Quality: 
\code{\link{accuracy}()},
\code{\link{huber_loss}()},
\code{\link{log_cosh_loss}()},
\code{\link{mae}()},
\code{\link{mse}()},
\code{\link{quantile_loss}()},
\code{\link{rmse}()},
\code{\link{vc}()}
}
\concept{Quality}
