% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepQuality.r
\name{mape}
\alias{mape}
\title{Mean absolute percentage error}
\usage{
mape(actual, predicted)
}
\arguments{
\item{actual}{A numeric vector with to-be values.}

\item{predicted}{A numeric vector with as-is values.}
}
\value{
Mean absolute percentage error (MAPE).
}
\description{
\code{mape} calculates the mean absolute percentage error.
}
\seealso{
\code{\link{mae}} for mean absolute error, \code{\link{mse}} for mean squared error,
  \code{\link{rmse}} for root mean square error, \code{\link{vc}} for variance coefficient.

Other Quality: 
\code{\link{mae}()},
\code{\link{rmse}()},
\code{\link{vc}()}
}
\concept{Quality}
