% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepQuality.r
\name{mape}
\alias{mape}
\title{Mean absolute percentage error (MAPE)}
\usage{
mape(actuals, preds, na.rm = FALSE)
}
\arguments{
\item{actuals}{A numeric vector of actual values.}

\item{preds}{A numeric vector of prediction values.}

\item{na.rm}{A logical value indicating whether actual and prediction pairs with at least one NA value should be ignored.}
}
\value{
Mean absolute percentage error.
}
\description{
Mean absolute percentage error (MAPE)
}
\seealso{
Other Quality: 
\code{\link{accuracy}()},
\code{\link{entropy}()},
\code{\link{gini_impurity}()},
\code{\link{huber_loss}()},
\code{\link{log_cosh_loss}()},
\code{\link{mae}()},
\code{\link{mse}()},
\code{\link{msle}()},
\code{\link{quantile_loss}()},
\code{\link{rmse}()},
\code{\link{rmsle}()},
\code{\link{vc}()}
}
\concept{Quality}
