% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepQuality.r
\name{mse}
\alias{mse}
\title{Mean squared error (MSE)}
\usage{
mse(y, yhat)
}
\arguments{
\item{y}{A numeric vector with actual values (to-be values).}

\item{yhat}{A numeric vector with estimated values (as-is values).}
}
\value{
Mean squared error.
}
\description{
Mean squared error (MSE)
}
\details{
In Machine and Deep Learning, MSE is also known as L2 loss function.
}
\seealso{
Other Quality: 
\code{\link{accuracy}()},
\code{\link{huber_loss}()},
\code{\link{log_cosh_loss}()},
\code{\link{mae}()},
\code{\link{mape}()},
\code{\link{quantile_loss}()},
\code{\link{rmse}()},
\code{\link{vc}()}
}
\concept{Quality}
