% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepQuality.r
\name{mse}
\alias{mse}
\title{Mean squared error}
\usage{
mse(y, yhat)
}
\arguments{
\item{y}{A numeric vector with actual values (to-be values).}

\item{yhat}{A numeric vector with estimated values (as-is values).}
}
\value{
Mean squared error (MSE).
}
\description{
Mean squared error
}
\seealso{
Other Quality: 
\code{\link{accuracy}()},
\code{\link{huber}()},
\code{\link{log_cosh}()},
\code{\link{mae}()},
\code{\link{mape}()},
\code{\link{quantile_loss}()},
\code{\link{rmse}()},
\code{\link{vc}()}
}
\concept{Quality}
