% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepML.r
\name{naive_forecast}
\alias{naive_forecast}
\title{Naive forecasting}
\usage{
naive_forecast(x, drift = 0, na = NA)
}
\arguments{
\item{x}{A vector, usually a numeric vector.}

\item{drift}{The number of periods used to calculate the change over time (it's called a drift).
If drift is more than 1 the mean value of the changes over time is used; default \code{0}.}

\item{na}{The value, default \code{NA}, used for gaps caused by the drift in the resulting vector.}
}
\value{
A series of naive predicted values based upon \code{x}.
}
\description{

}
\details{
The following naive forecast approaches are implemented:
\itemize{
\item Random Walk: \eqn{y(t+1) = y(t)}
\item One drift: \eqn{y(t+1) = y(t) + [y(t)-y(t-1)]}
\item Many drifts: \eqn{y(t+1) = y(t) + [(1/drifts) * \sum ([y(t)-y(t-1)])]}
}
}
\seealso{
Other Machine Learning: 
\code{\link{cross_validation_split}()},
\code{\link{decision_tree}()},
\code{\link{k_nearest_neighbors}()},
\code{\link{moving_average}()},
\code{\link{naive_bayes}()},
\code{\link{predict.decisiontree}()},
\code{\link{predict.kmeans}()},
\code{\link{predict.naivebayes}()}
}
\concept{Machine Learning}
