% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepUtils.r
\name{nsamples}
\alias{nsamples}
\alias{nsamples.array}
\alias{nsamples.tensor}
\alias{nsamples.numpy}
\alias{nsamples.matrix}
\title{Number of samples within a data structure}
\usage{
nsamples(data, ...)

\method{nsamples}{array}(x)

\method{nsamples}{tensor}(x)

\method{nsamples}{numpy}(x)

\method{nsamples}{matrix}(x)
}
\arguments{
\item{data}{A data structure, the number of samples is extracted.}

\item{x}{A multidimensional data structure like array, tensor, numpy or matrix.}
}
\value{
Number of samples.
}
\description{

}
\details{
The number of samples is stored in the first dimension of \code{x}.
}
\seealso{
Other Utils: 
\code{\link{as_ANN_matrix}()},
\code{\link{as_tensor_1D}()},
\code{\link{as_tensor_2D}()},
\code{\link{as_tensor_3D}()},
\code{\link{degree}()},
\code{\link{distance}()},
\code{\link{flatten}()},
\code{\link{nsubsequences}()},
\code{\link{ntimesteps}()},
\code{\link{numpy}()},
\code{\link{nunits}()},
\code{\link{probability}()},
\code{\link{radian}()},
\code{\link{sd_pop}()},
\code{\link{similarity}()},
\code{\link{tensor}()},
\code{\link{var_pop}()},
\code{\link{vector_as_ANN_matrix}()},
\code{\link{vector_as_numeric}()}
}
\concept{Utils}
