% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepArray.r
\name{nsize}
\alias{nsize}
\title{Number of dimensions}
\usage{
nsize(x)
}
\arguments{
\item{x}{A multidimensional data structure like array, marray, tensor, matrix or data.frame.}
}
\value{
Number of elements.
}
\description{
Number of dimensions
}
\details{
This function corresponds to \code{ndarray.size} from NumPy.
}
\seealso{
Other Array: 
\code{\link{column_stack}()},
\code{\link{dimC<-}()},
\code{\link{empty}()},
\code{\link{eye}()},
\code{\link{flatten}()},
\code{\link{full}()},
\code{\link{insert}()},
\code{\link{mabind}()},
\code{\link{mamatrix}()},
\code{\link{marray}()},
\code{\link{ndim}()},
\code{\link{ones}()},
\code{\link{reshape.array}()},
\code{\link{row_stack}()},
\code{\link{slice}()},
\code{\link{squeeze}()},
\code{\link{transpose}()},
\code{\link{vander}()},
\code{\link{zeros}()}
}
\concept{Array}
