% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepUtils.r
\name{nsubsequences}
\alias{nsubsequences}
\alias{nsubsequences.array}
\alias{nsubsequences.tensor}
\alias{nsubsequences.numpy}
\title{Number of subsequences within a data structure}
\usage{
nsubsequences(data, ...)

\method{nsubsequences}{array}(x, default = 0L)

\method{nsubsequences}{tensor}(x, default = 0L)

\method{nsubsequences}{numpy}(x, default = 0L)
}
\arguments{
\item{data}{A data structure, the number of subsequences is extracted.}

\item{x}{A multidimensional data structure like array, tensor or numpy.}
}
\value{
Number of subsequences.
}
\description{

}
\details{
The number of subsequences is stored in the second dimension of a four-dimensional \code{x}, usually used for a temporal CNN.
}
\seealso{
Other Utils: 
\code{\link{as_ANN_matrix}()},
\code{\link{as_tensor_1D}()},
\code{\link{as_tensor_2D}()},
\code{\link{as_tensor_3D}()},
\code{\link{degree}()},
\code{\link{distance}()},
\code{\link{flatten}()},
\code{\link{nsamples}()},
\code{\link{ntimesteps}()},
\code{\link{numpy}()},
\code{\link{nunits}()},
\code{\link{probability}()},
\code{\link{radian}()},
\code{\link{sd_pop}()},
\code{\link{similarity}()},
\code{\link{tensor}()},
\code{\link{var_pop}()},
\code{\link{vector_as_ANN_matrix}()},
\code{\link{vector_as_numeric}()}
}
\concept{Utils}
