% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepUtils.r
\name{ntimesteps}
\alias{ntimesteps}
\alias{ntimesteps.array}
\alias{ntimesteps.tensor}
\alias{ntimesteps.numpy}
\title{Number of timesteps within a data structure}
\usage{
ntimesteps(data, ...)

\method{ntimesteps}{array}(x, default = 1L)

\method{ntimesteps}{tensor}(x, default = 1L)

\method{ntimesteps}{numpy}(x, default = 1L)
}
\arguments{
\item{data}{A data structure, the number of timesteps is extracted.}

\item{x}{A multidimensional data structure like array, tensor or numpy.}
}
\value{
Number of timesteps.
}
\description{

}
\details{
The number of timesteps is stored in the second dimension of a three-dimensional \code{x}, usually used for a LSTM,
  or in the third dimension of a four-dimensional \code{x}, usually used for a temporal CNN.
}
\seealso{
Other Utils: 
\code{\link{as_ANN_matrix}()},
\code{\link{as_tensor_1D}()},
\code{\link{as_tensor_2D}()},
\code{\link{as_tensor_3D}()},
\code{\link{degree}()},
\code{\link{distance}()},
\code{\link{flatten}()},
\code{\link{nsamples}()},
\code{\link{nsubsequences}()},
\code{\link{numpy}()},
\code{\link{nunits}()},
\code{\link{probability}()},
\code{\link{radian}()},
\code{\link{sd_pop}()},
\code{\link{similarity}()},
\code{\link{tensor}()},
\code{\link{var_pop}()},
\code{\link{vector_as_ANN_matrix}()},
\code{\link{vector_as_numeric}()}
}
\concept{Utils}
