% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepUtils.r
\name{nunits}
\alias{nunits}
\alias{nunits.array}
\alias{nunits.tensor}
\alias{nunits.marray}
\alias{nunits.matrix}
\title{Number of units within a data structure}
\usage{
nunits(data, ...)

\method{nunits}{array}(x)

\method{nunits}{tensor}(x)

\method{nunits}{marray}(x)

\method{nunits}{matrix}(x)
}
\arguments{
\item{data}{A data structure, the number of units is extracted.}

\item{x}{A multidimensional data structure like array, tensor, marray or matrix.}
}
\value{
Number of units.
}
\description{

}
\details{
The number of units is stored in the last dimension of \code{x}.
  What a unit is or what it stands for is determined by the context. Usually, a unit is an attribute (feature or outcome).
  In the context of image processing, a unit on feature side represents a color channel.
}
\seealso{
Other Utils: 
\code{\link{as_ANN_matrix}()},
\code{\link{as_tensor_1d}()},
\code{\link{as_tensor_2d}()},
\code{\link{as_tensor_3d}()},
\code{\link{degree}()},
\code{\link{dimC<-}()},
\code{\link{distance}()},
\code{\link{eye}()},
\code{\link{flatten}()},
\code{\link{insert}()},
\code{\link{list_as_numeric}()},
\code{\link{mbind}()},
\code{\link{ndim}()},
\code{\link{nsamples}()},
\code{\link{nsubsequences}()},
\code{\link{ntimesteps}()},
\code{\link{probability}()},
\code{\link{radian}()},
\code{\link{random_seed}()},
\code{\link{re.factor}()},
\code{\link{reshape.array}()},
\code{\link{sd_pop}()},
\code{\link{similarity}()},
\code{\link{slice}()},
\code{\link{squeeze}()},
\code{\link{ta}()},
\code{\link{tensor}()},
\code{\link{vander}()},
\code{\link{var_pop}()},
\code{\link{vector_as_ANN_matrix}()},
\code{\link{vector_as_numeric}()}
}
\concept{Utils}
