% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepDummy.r
\name{one_hot_decode}
\alias{one_hot_decode}
\title{One-hot decoding}
\usage{
one_hot_decode(x_encoded)
}
\arguments{
\item{x_encoded}{An already one-hot encoded variable; the outcome from \code{one_hot_encode}.}
}
\value{
A vector with the original levels (categories) of a non-metric variable.
}
\description{
\code{one_hot_decode} builds back an already one-hot encoded variable into its original value form.
}
\seealso{
\code{\link{one_hot_encode}}.

Other Dummyfication: 
\code{\link{dummify}()},
\code{\link{effectcoding}()},
\code{\link{one_hot_encode}()},
\code{\link{resample.imbalanced}()}
}
\concept{Dummyfication}
