% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepDummy.r
\name{one_hot_encode}
\alias{one_hot_encode}
\title{One-hot encoding}
\usage{
one_hot_encode(x, ordered = FALSE)
}
\arguments{
\item{x}{A vector with values (levels) of a categorical variable.}

\item{ordered}{A logical value indicating whether the factor levels of \code{x} should be encoded in an ordered way or not (default).}
}
\value{
A matrix with all levels as columns with either \code{0} or \code{1} values.
}
\description{
\code{one_hot_encode} rebuilds a categorical variable to a so-called 'one-hot vector'.
  Within a sample (row) of a one-hot vector (matrix) each level of the variable is rebuild in the binary form \code{(0|1,0|1,0|1,...)}.
}
\details{
An unordered encoding creates an indicator matrix with the value \code{1} only for the given level of \code{x} and \code{0} for all other levels.
  In opposite, an ordered encoding assumes an intrinsic order of all levels whereby a given level automatically means reaching and exceeding all previous levels.
  In that case, all previous levels are also encoded with value \code{1}.
}
\seealso{
\code{\link{one_hot_decode}}.

Other Dummyfication: 
\code{\link{append_rows}()},
\code{\link{dummify.multilabel}()},
\code{\link{dummify}()},
\code{\link{effectcoding}()},
\code{\link{one_hot_decode}()},
\code{\link{remove_columns}()},
\code{\link{resample.imbalanced}()}
}
\concept{Dummyfication}
