% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepArray.r
\name{ones}
\alias{ones}
\title{Create array filled with ones}
\usage{
ones(dim = NULL)
}
\arguments{
\item{dim}{Shape of the new array.}
}
\value{
An array of ones with the given shape.
}
\description{
Create array filled with ones
}
\details{
This function corresponds to \code{ones()} from NumPy.
}
\seealso{
Other Array: 
\code{\link{column_stack}()},
\code{\link{dimC<-}()},
\code{\link{empty}()},
\code{\link{expand_dims}()},
\code{\link{eye}()},
\code{\link{flatten}()},
\code{\link{flip}()},
\code{\link{full}()},
\code{\link{insert}()},
\code{\link{mabind}()},
\code{\link{mamatrix}()},
\code{\link{marray}()},
\code{\link{ndim}()},
\code{\link{nsize}()},
\code{\link{reshape.array}()},
\code{\link{row_stack}()},
\code{\link{slice}()},
\code{\link{squeeze}()},
\code{\link{transpose}()},
\code{\link{vander}()},
\code{\link{zeros}()}
}
\concept{Array}
