% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepOutlier.r
\name{outlier}
\alias{outlier}
\title{Definition and detection of outliers}
\usage{
outlier(x, type = c("tukey"), ...)
}
\arguments{
\item{x}{A numeric vector.}

\item{type}{The type of outlier definition and detection. \code{tukey} refers to the method of Tukey (1977).}

\item{...}{Depends on the type value. For \code{tukey} the constant \code{k} can be specified, otherwise it's value is \code{1.5}.}
}
\value{
A named list of lower and upper boundaries and values.
}
\description{
Definition and detection of outliers
}
\references{
Tukey, John W. (1977): Exploratory Data Analysis. 1977. Reading: Addison-Wesley.
}
\seealso{
\code{\link[stats]{quantile}}, \code{\link[stats]{IQR}}, \code{\link{winsorize}}.

Other Outlier: 
\code{\link{winsorize}()}
}
\concept{Outlier}
