% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepOutlier.r
\name{outlier}
\alias{outlier}
\title{Definition and detection of outliers}
\usage{
outlier(x, type = c("tukey", "ml"), na.replace = FALSE, ...)
}
\arguments{
\item{x}{A numeric vector.}

\item{type}{The type of outlier definition and detection. 
\code{tukey} refers to the method of Tukey (1977).
\code{ml} denotes maximum likelihood estimation.}

\item{na.replace}{Boolean that indicates whether outliers shell be replaced by \code{NA}.}

\item{...}{Dependent on \code{type}. 
For \code{tukey} the constant \code{k} can be specified, otherwise it's value is \code{1.5}. 
For \code{ml} the constant \code{k} can be specified, otherwise it's value is \code{2}.}
}
\value{
Dependent on \code{na.replace}
  By default (\code{FALSE}), a named list of lower and upper boundaries and values.
  Otherwise, the vector \code{x} with replaced outliers.
}
\description{
Definition and detection of outliers
}
\references{
Tukey, John W. (1977): Exploratory Data Analysis. 1977. Reading: Addison-Wesley.
}
\seealso{
\code{\link[stats]{quantile}}, \code{\link[stats]{IQR}}, \code{\link{outlier.dataset}}, \code{\link{winsorize}}.

Other Outlier: 
\code{\link{outlier.dataset}()},
\code{\link{winsorize}()}
}
\concept{Outlier}
