% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepTimeSeries.r
\name{period}
\alias{period}
\title{Subset data set/time series to specific periodically data}
\usage{
period(
  dataset,
  p = 1,
  type = c("seq", "weekday", "monthday", "week", "month", "quarter", "season"),
  ...
)
}
\arguments{
\item{dataset}{A data set or time series, usually a data.frame.}

\item{p}{The index of the periodic column.}

\item{type}{Different subsetting types
\code{seq} A sequence specified thru start index and increment passed in \code{...} as numbers.
\code{weekday} Certain days of week are extracted from the data set whereby the days are passed in \code{...} as strings.
\code{monthday} Certain days of month are extracted from the data set whereby the days are passed in \code{...} as numbers.
\code{week} Certain weeks are extracted from the data set whereby the weeks are passed in \code{...} as numbers.
\code{month} Certain months are extracted from the data set whereby the months are passed in \code{...} as strings.
\code{quarter} Certain quarters are extracted from the data set whereby the quarters are passed in \code{...} as numbers.
\code{season} Certain seasons are extracted from the data set whereby the seasons are passed in \code{...} as strings.}

\item{...}{Arguments dependent from the \code{type}.}
}
\value{
A subset of dataset with periodically specified data.
}
\description{
Subset data set/time series to specific periodically data
}
\seealso{
Other TimeSeries: 
\code{\link{build.stationary}()},
\code{\link{get.season}()},
\code{\link{invert_differencing}()},
\code{\link{lags}()}
}
\concept{TimeSeries}
