% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepML.r
\name{predict.kmeans}
\alias{predict.kmeans}
\title{Prediction for kmeans}
\usage{
\method{predict}{kmeans}(object, newdata)
}
\arguments{
\item{object}{An object from type result of kmeans.}

\item{newdata}{A vector or matrix with new data to predict labels for.}
}
\value{
A vector of predicted labels.
}
\description{
Prediction for kmeans
}
\seealso{
\code{\link[stats]{kmeans}}.

Other Machine Learning: 
\code{\link{as.degree}()},
\code{\link{as.radian}()},
\code{\link{cross_validation_split}()},
\code{\link{distance}()},
\code{\link{erfcinv}()},
\code{\link{erfc}()},
\code{\link{erfinv}()},
\code{\link{erf}()},
\code{\link{k_nearest_neighbors}()},
\code{\link{moving_average}()},
\code{\link{naive_forecast}()},
\code{\link{sd.pop}()},
\code{\link{similarity}()},
\code{\link{var.pop}()}
}
\concept{Machine Learning}
