% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepMLP.r
\name{predict.MLP}
\alias{predict.MLP}
\title{Predict with SLP/MLP model}
\usage{
\method{predict}{MLP}(mlp, X, batch_size = 1, scale_type = NULL, scaler = NULL)
}
\arguments{
\item{mlp}{A model object, returned by \code{fit.MLP} in the list element \code{model}.}

\item{X}{A feature data set, usually a matrix or data.frame.}

\item{batch_size}{Batch size, the number of samples used per gradient update.}

\item{scale_type}{Type of scaling with supported techniques min-max scaling (\code{minmax}), z-score scaling (\code{zscore}) and log transformation (\code{log}).
Per default (\code{NULL}) no inverted scaling is done.}

\item{scaler}{Scaling factors for the different scaling types. The type min-max scaling needs a list with vectors of min and max values for each outcome,
z-score scaling needs a list with vectors of mean and sd values for each outcome, log transformation needs no scaler.}
}
\value{
A matrix with predicted outcome values per column.
}
\description{
Predict with SLP/MLP model
}
\seealso{
\code{\link{fit.MLP}}, \code{\link[stats]{predict}}, \code{\link{scale.datasets}}.

Other Single & Multi Layer Perceptron (SLP, MLP): 
\code{\link{as.ANN.matrix}()},
\code{\link{as.MLP.X}()},
\code{\link{as.MLP.Y}()},
\code{\link{as.tensor.1D}()},
\code{\link{as.tensor.2D}()},
\code{\link{as.tensor.3D}()},
\code{\link{as.tensor}()},
\code{\link{build.MLP}()},
\code{\link{fit.MLP}()},
\code{\link{get.MLP.X.units}()},
\code{\link{get.MLP.Y.units}()},
\code{\link{vector.as.ANN.matrix}()}
}
\concept{Single & Multi Layer Perceptron (SLP, MLP)}
