% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepUtils.r
\name{radian}
\alias{radian}
\title{Degree to radian}
\usage{
radian(degree)
}
\arguments{
\item{degree}{Degree.}
}
\value{
The radian of \code{degree}.
}
\description{
Degree to radian
}
\examples{
  radian(180) # pi
}
\seealso{
Other Utils: 
\code{\link{as_ANN_matrix}()},
\code{\link{degree}()},
\code{\link{distance}()},
\code{\link{list_as_numeric}()},
\code{\link{probability}()},
\code{\link{random_seed}()},
\code{\link{re.factor}()},
\code{\link{sd_pop}()},
\code{\link{similarity}()},
\code{\link{var_pop}()},
\code{\link{vector_as_ANN_matrix}()},
\code{\link{vector_as_numeric}()}
}
\concept{Utils}
