% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepUtils.r
\name{re.factor}
\alias{re.factor}
\title{Renew a factor object}
\usage{
re.factor(x)
}
\arguments{
\item{x}{A factor object.}
}
\value{
The renewed factor object \code{x}.
}
\description{

}
\seealso{
\code{\link{factor}}.

Other Utils: 
\code{\link{as_ANN_matrix}()},
\code{\link{as_tensor_1D}()},
\code{\link{as_tensor_2D}()},
\code{\link{as_tensor_3D}()},
\code{\link{degree}()},
\code{\link{distance}()},
\code{\link{flatten}()},
\code{\link{marray}()},
\code{\link{nsamples}()},
\code{\link{nsubsequences}()},
\code{\link{ntimesteps}()},
\code{\link{nunits}()},
\code{\link{probability}()},
\code{\link{radian}()},
\code{\link{sd_pop}()},
\code{\link{similarity}()},
\code{\link{tensor}()},
\code{\link{var_pop}()},
\code{\link{vector_as_ANN_matrix}()},
\code{\link{vector_as_numeric}()}
}
\concept{Utils}
