% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepUtils.r
\name{re.factor}
\alias{re.factor}
\title{Renew an (ordered) factor object}
\usage{
re.factor(x)
}
\arguments{
\item{x}{An (ordered) factor object.}
}
\value{
The renewed (ordered) factor object \code{x}.
}
\description{

}
\seealso{
\code{\link{factor}}.

Other Utils: 
\code{\link{as_ANN_matrix}()},
\code{\link{as_tensor_1d}()},
\code{\link{as_tensor_2d}()},
\code{\link{as_tensor_3d}()},
\code{\link{degree}()},
\code{\link{dimC<-}()},
\code{\link{distance}()},
\code{\link{eye}()},
\code{\link{flatten}()},
\code{\link{insert}()},
\code{\link{list_as_numeric}()},
\code{\link{mbind}()},
\code{\link{ndim}()},
\code{\link{nsamples}()},
\code{\link{nsubsequences}()},
\code{\link{ntimesteps}()},
\code{\link{nunits}()},
\code{\link{probability}()},
\code{\link{radian}()},
\code{\link{random_seed}()},
\code{\link{reshape.array}()},
\code{\link{sd_pop}()},
\code{\link{similarity}()},
\code{\link{slice}()},
\code{\link{squeeze}()},
\code{\link{ta}()},
\code{\link{tensor}()},
\code{\link{vander}()},
\code{\link{var_pop}()},
\code{\link{vector_as_ANN_matrix}()},
\code{\link{vector_as_numeric}()}
}
\concept{Utils}
