% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepQuality.r
\name{rmse}
\alias{rmse}
\title{Root mean square error (RMSE)}
\usage{
rmse(y, yhat)
}
\arguments{
\item{y}{A numeric vector with actual values (to-be values).}

\item{yhat}{A numeric vector with estimated values (as-is values).}
}
\value{
Root mean square error.
}
\description{
Root mean square error (RMSE)
}
\seealso{
Other Quality: 
\code{\link{accuracy}()},
\code{\link{huber}()},
\code{\link{log_cosh}()},
\code{\link{mae}()},
\code{\link{mape}()},
\code{\link{mse}()},
\code{\link{quantile_loss}()},
\code{\link{vc}()}
}
\concept{Quality}
