% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepQuality.r
\name{rmse}
\alias{rmse}
\title{Root mean square error}
\usage{
rmse(actual, predicted)
}
\arguments{
\item{actual}{A numeric vector with to-be values.}

\item{predicted}{A numeric vector with as-is values.}
}
\value{
Root mean square error (RMSE).
}
\description{
\code{rmse} calculates the root mean square error.
}
\seealso{
\code{\link{mae}} for mean absolute error, \code{\link{mape}} for mean absolute percentage error,
  \code{\link{mse}} for mean squared error, \code{\link{vc}} for variance coefficient.

Other Quality: 
\code{\link{mae}()},
\code{\link{mape}()},
\code{\link{vc}()}
}
\concept{Quality}
