% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepQuality.r
\name{rmse}
\alias{rmse}
\title{Root mean square error (RMSE)}
\usage{
rmse(actuals, preds, na.rm = FALSE)
}
\arguments{
\item{actuals}{A numeric vector of actual values.}

\item{preds}{A numeric vector of prediction values.}

\item{na.rm}{A logical value indicating whether actual and prediction pairs with at least one NA value should be ignored.}
}
\value{
Root mean square error.
}
\description{

}
\seealso{
Other Quality: 
\code{\link{accuracy}()},
\code{\link{cross_entropy}()},
\code{\link{entropy}()},
\code{\link{erfcinv}()},
\code{\link{erfc}()},
\code{\link{erfinv}()},
\code{\link{erf}()},
\code{\link{gini_impurity}()},
\code{\link{huber_loss}()},
\code{\link{log_cosh_loss}()},
\code{\link{mae}()},
\code{\link{mape}()},
\code{\link{mse}()},
\code{\link{msle}()},
\code{\link{quantile_loss}()},
\code{\link{rmsle}()},
\code{\link{vc}()}
}
\concept{Quality}
