% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepScaling.r
\name{scale.dataset}
\alias{scale.dataset}
\title{Scaling of a data set}
\usage{
\method{scale}{dataset}(
  dataset,
  columns = NULL,
  type = c("minmax", "zscore", "log"),
  invert = FALSE,
  ...
)
}
\arguments{
\item{dataset}{A data set, usually a matrix or data frame.}

\item{columns}{The names or indices of the columns to be scaled.}

\item{type}{Type of scaling with supported techniques min-max scaling (\code{minmax}), z-score scaling (\code{zscore}) and log transformation (\code{log}).}

\item{invert}{A logical value indicating the direction of scaling. If set to \code{TRUE}, \code{columns} are already scaled vectors and scaling will be inverted.}

\item{...}{Further arguments depend on the type of scaling. Min-max scaling and z-score scaling need two further arguments if necessary.}
}
\value{
The scaled \code{dataset}.
}
\description{
Scaling of a data set
}
\seealso{
\code{\link{scaling}}

Other Scaling: 
\code{\link{scale.datasets}()},
\code{\link{scale.log}()},
\code{\link{scale.minmax}()},
\code{\link{scale.zscore}()},
\code{\link{scaling}()}
}
\concept{Scaling}
