% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepScaling.r
\name{scale.datasets}
\alias{scale.datasets}
\title{Scaling of a train and test data set}
\usage{
\method{scale}{datasets}(trainset, testset, type = c("minmax", "zscore", "log"))
}
\arguments{
\item{trainset}{A train dataset.}

\item{testset}{A test dataset.}

\item{type}{Type of scaling with supported techniques min-max scaling (\code{minmax}), z-score scaling (\code{zscore}) and log transformation (\code{log}).}
}
\value{
A named list dependent on the \code{type}.
  \code{minmax}: The first element stores the min value, the second element the max value, the third element the scaled train dataset and the fourth element the scaled test data set.
  \code{zscore}: The first element stores the mean value, the second element the sd value, the third element the scaled train dataset and the fourth element the scaled test data set.
  \code{log}: The first element stores the scaled train dataset and the second element the scaled test data set.
}
\description{
Scaling of a train and test data set
}
\seealso{
\code{\link{scaling}}

Other Scaling: 
\code{\link{scale.log}()},
\code{\link{scale.minmax}()},
\code{\link{scale.zscore}()},
\code{\link{scaling}()}
}
\concept{Scaling}
