% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepScaling.r
\name{scaling}
\alias{scaling}
\title{Scaling of a numeric vector}
\usage{
scaling(
  x,
  type = c("minmax", "zscore", "log"),
  use.attr = TRUE,
  invert = FALSE,
  ...
)
}
\arguments{
\item{x}{A numeric vector whose elements shell be scaled.}

\item{type}{Type of scaling with supported techniques min-max scaling (\code{minmax}), z-score scaling (\code{zscore}) and log transformation (\code{log}).}

\item{use.attr}{A boolean that controls whether scaling factors like \code{min}, \code{max}, \code{mean} and \code{sd} are stored as named attributes of the scaled vector \code{x}.}

\item{invert}{A boolean that indicates the direction of scaling. If set to \code{TRUE}, \code{x} is an already scaled vector and scaling will be inverted.}

\item{...}{Further arguments depend on the type of scaling. Min-max scaling and z-score scaling need two further arguments if necessary.}
}
\value{
The scaled or re-scaled (inverted) numeric vector \code{x}.
}
\description{
Scaling of a numeric vector
}
\seealso{
\code{\link{scale.minmax}}, \code{\link{scale.zscore}}, \code{\link{scale.log}}.

Other Scaling: 
\code{\link{scale.datasets}()},
\code{\link{scale.log}()},
\code{\link{scale.minmax}()},
\code{\link{scale.zscore}()}
}
\concept{Scaling}
