% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepScaling.r
\name{scaling}
\alias{scaling}
\title{Scaling of a numeric object}
\usage{
scaling(
  x,
  type = c("minmax", "zscore", "log", "center"),
  use_attr = TRUE,
  invert = FALSE,
  ...
)
}
\arguments{
\item{x}{A numeric object whose elements should be scaled.}

\item{type}{Type of scaling with supported techniques min-max scaling (\code{minmax}), z-score scaling (\code{zscore}), log transformation (\code{log}) and centering (\code{center}).}

\item{use_attr}{A logical value indicating whether scaling factors like \code{min}, \code{max}, \code{mean} and \code{sd} are stored as named attributes of the scaled \code{x}.}

\item{invert}{A logical value indicating the direction of scaling. If set to \code{TRUE}, \code{x} is already scaled and scaling will be inverted.}

\item{...}{Further arguments depend on the type of scaling. Min-max scaling, z-score scaling and centering need two or one further arguments if necessary.}
}
\value{
The scaled or re-scaled (inverted) \code{x}.
}
\description{

}
\seealso{
\code{\link{scale_minmax}}, \code{\link{scale_zscore}}, \code{\link{scale_log}}, \code{\link{scale_center}}.

Other Scaling: 
\code{\link{scale_center}()},
\code{\link{scale_dataset}()},
\code{\link{scale_log}()},
\code{\link{scale_minmax}()},
\code{\link{scale_train_test}()},
\code{\link{scale_zscore}()}
}
\concept{Scaling}
