% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepUtils.r
\name{slice}
\alias{slice}
\alias{slice.array}
\alias{slice.marray}
\alias{slice.matrix}
\title{Multidimensional array slicing}
\usage{
slice(data, ...)

\method{slice}{array}(a, ..., drop = TRUE)

\method{slice}{marray}(a, ..., drop = TRUE)

\method{slice}{matrix}(a, ..., drop = TRUE)
}
\arguments{
\item{...}{Indexing instructions in form of \code{name = value} pairs. The names of the arguments specify the dimensions and the values its values.}

\item{a}{A vector, matrix, or array.}

\item{drop}{For matrices and arrays. If \code{TRUE} the result is coerced to the lowest possible dimension. This only works for extracting elements, not for the replacement. See \code{\link[base]{drop}} for further details.}
}
\value{
An extracted part of \code{a}.
}
\description{
Multidimensional array slicing
}
\details{
\code{slice} is an alternative way to handle indexing array objects, usually done with \code{\link[base]{[}}. The dimensions must be indexed by names,
  i for the first dimension, j for the second and so on. The assigned values are the values (elements) of the corresponding dimension. The indexing expressions are the same as for \code{\link[base]{[}}.
}
\examples{
  a <- array(1:48, dim = c(4, 3, 2, 2))
  slice(a) # complete four-dimensional array
  slice(a, l = 2) # the values of the array of the second element of the last dimension (4th dimension)
  slice(a, i = 1, j = 3) the values of the array of the first element of the first dimension (1st row) and the third element of the second dimension (3rd column) across all bunches of the remaining dimensions 3 and 4.
}
\references{
\code{slice} is inspired by the function with the same name from package \code{arrayhelpers}. Implementation credits go the the author of this package.
}
\seealso{
Other Utils: 
\code{\link{as_ANN_matrix}()},
\code{\link{as_tensor_1d}()},
\code{\link{as_tensor_2d}()},
\code{\link{as_tensor_3d}()},
\code{\link{degree}()},
\code{\link{distance}()},
\code{\link{eye}()},
\code{\link{flatten}()},
\code{\link{insert}()},
\code{\link{list_as_numeric}()},
\code{\link{marray}()},
\code{\link{mbind}()},
\code{\link{ndim}()},
\code{\link{nsamples}()},
\code{\link{nsubsequences}()},
\code{\link{ntimesteps}()},
\code{\link{nunits}()},
\code{\link{probability}()},
\code{\link{radian}()},
\code{\link{random_seed}()},
\code{\link{re.factor}()},
\code{\link{sd_pop}()},
\code{\link{similarity}()},
\code{\link{squeeze}()},
\code{\link{ta}()},
\code{\link{tensor}()},
\code{\link{vander}()},
\code{\link{var_pop}()},
\code{\link{vector_as_ANN_matrix}()},
\code{\link{vector_as_numeric}()}
}
\concept{Utils}
