% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepArray.r
\name{squeeze}
\alias{squeeze}
\title{Compress the shape of an array}
\usage{
squeeze(a, axis = NULL, order = c("C", "F"))
}
\arguments{
\item{a}{A multidimensional array.}

\item{axis}{The dimensions which should be removed. If \code{NULL} (default), all dimensions of length one are removed.}

\item{order}{The order in which elements of data should be read during rearrangement after removing of corresponding dimensions.
By default, the order is equivalent to the \code{C}-style ordering and means elements should be read in row-major order.
In opposite, the \code{Fortran}-style ordering means elements should be read in column-major order.}
}
\value{
The array \code{a} usually without dimensions of length one.
}
\description{
Remove dimensions of length one from a multidimensional array.
}
\details{
This function corresponds to \code{squeeze()} from NumPy.
}
\seealso{
Other Array: 
\code{\link{column_stack}()},
\code{\link{dimC<-}()},
\code{\link{empty}()},
\code{\link{expand_dims}()},
\code{\link{eye}()},
\code{\link{flatten}()},
\code{\link{flip}()},
\code{\link{full}()},
\code{\link{insert}()},
\code{\link{mabind}()},
\code{\link{mamatrix}()},
\code{\link{marray}()},
\code{\link{ndim}()},
\code{\link{nsize}()},
\code{\link{ones}()},
\code{\link{reshape.array}()},
\code{\link{row_stack}()},
\code{\link{slice}()},
\code{\link{transpose}()},
\code{\link{vander}()},
\code{\link{zeros}()}
}
\concept{Array}
