% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepArray.r
\name{squeeze}
\alias{squeeze}
\title{Array compressing}
\usage{
squeeze(a, axis = NULL, order = c("C", "F"))
}
\arguments{
\item{a}{An array.}

\item{axis}{The dimensions which should be removed. If \code{NULL} (default), all dimensions of length one are removed.}

\item{order}{The order in which elements of data should be read during rearrangement after removing of corresponding dimensions.
By default, the order is equivalent to the \code{C}-style ordering and means elements should be read in row-major order.
In opposite, the \code{Fortran}-style ordering means elements should be read in column-major order.}
}
\value{
The array \code{a} usually without dimensions of length one.
}
\description{
Compress the shape of an array by removing dimensions of length one from the array.
}
\details{
This function corresponds to \code{squeeze()} from NumPy.
  The base R function \code{\link{drop}} does the same as this function. In opposite to \code{drop} this function
  allows reordering the elements of the newly created array as well as specifying only certain axes.
}
\seealso{
\code{\link{drop}}.

Other Array: 
\code{\link{column_stack}()},
\code{\link{dimC<-}()},
\code{\link{empty}()},
\code{\link{expand_dims}()},
\code{\link{eye}()},
\code{\link{flatten}()},
\code{\link{flip}()},
\code{\link{full}()},
\code{\link{insert}()},
\code{\link{mabind}()},
\code{\link{mamatrix}()},
\code{\link{marray}()},
\code{\link{ndim}()},
\code{\link{nsize}()},
\code{\link{ones}()},
\code{\link{reshape.array}()},
\code{\link{rot90}()},
\code{\link{row_stack}()},
\code{\link{slice}()},
\code{\link{transpose}()},
\code{\link{vander}()},
\code{\link{zeros}()}
}
\concept{Array}
