% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepUtils.r
\name{ta}
\alias{ta}
\alias{ta.array}
\alias{ta.marray}
\alias{ta.matrix}
\title{Transpose multidimensional array}
\usage{
ta(x)

\method{ta}{array}(x)

\method{ta}{marray}(x)

\method{ta}{matrix}(x)
}
\arguments{
\item{x}{A multidimensional array.}
}
\value{
The array \code{x} with the first two dimensions swapped.
}
\description{
This function transposes a multidimensional array by swapping the first two dimensions.
}
\seealso{
\code{\link{t}}.

Other Utils: 
\code{\link{as_ANN_matrix}()},
\code{\link{as_tensor_1D}()},
\code{\link{as_tensor_2D}()},
\code{\link{as_tensor_3D}()},
\code{\link{degree}()},
\code{\link{distance}()},
\code{\link{flatten}()},
\code{\link{list_as_numeric}()},
\code{\link{marray}()},
\code{\link{mbind}()},
\code{\link{ndim}()},
\code{\link{nsamples}()},
\code{\link{nsubsequences}()},
\code{\link{ntimesteps}()},
\code{\link{nunits}()},
\code{\link{probability}()},
\code{\link{radian}()},
\code{\link{re.factor}()},
\code{\link{sd_pop}()},
\code{\link{similarity}()},
\code{\link{tensor}()},
\code{\link{var_pop}()},
\code{\link{vector_as_ANN_matrix}()},
\code{\link{vector_as_numeric}()}
}
\concept{Utils}
