% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepArray.r
\name{transpose}
\alias{transpose}
\title{Transpose multidimensional array}
\usage{
transpose(a, perm = NULL)
}
\arguments{
\item{a}{A multidimensional array.}

\item{perm}{The permutation vector of the dimensions. The default \code{NULL} indicates to reverse the order of the dimensions.}
}
\value{
The array \code{a} with swapped dimensions.
}
\description{
Transpose multidimensional array
}
\seealso{
\code{\link{t}}.

Other Array: 
\code{\link{column_stack}()},
\code{\link{dimC<-}()},
\code{\link{empty}()},
\code{\link{expand_dims}()},
\code{\link{eye}()},
\code{\link{flatten}()},
\code{\link{flip}()},
\code{\link{full}()},
\code{\link{insert}()},
\code{\link{mabind}()},
\code{\link{mamatrix}()},
\code{\link{marray}()},
\code{\link{ndim}()},
\code{\link{nsize}()},
\code{\link{ones}()},
\code{\link{reshape.array}()},
\code{\link{rot90}()},
\code{\link{row_stack}()},
\code{\link{slice}()},
\code{\link{squeeze}()},
\code{\link{vander}()},
\code{\link{zeros}()}
}
\concept{Array}
