% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepArray.r
\name{vander}
\alias{vander}
\title{Array creation}
\usage{
vander(data, n)
}
\arguments{
\item{data}{The data to be reshaped to a Vandermonde matrix.}

\item{n}{The number of columns of the resulting matrix.}
}
\value{
A Vandermonde matrix.
}
\description{
Create Vandermonde matrix.
}
\details{
This function corresponds to \code{vander()} from NumPy.
}
\seealso{
Other Array: 
\code{\link{DIM}()},
\code{\link{column_stack}()},
\code{\link{delete}()},
\code{\link{dimC<-}()},
\code{\link{empty}()},
\code{\link{expand_dims}()},
\code{\link{eye}()},
\code{\link{flatten}()},
\code{\link{flip}()},
\code{\link{full}()},
\code{\link{insert}()},
\code{\link{mabind}()},
\code{\link{mamatrix}()},
\code{\link{marray}()},
\code{\link{ndim}()},
\code{\link{nsize}()},
\code{\link{ones}()},
\code{\link{reshape.array}()},
\code{\link{rot90}()},
\code{\link{row_stack}()},
\code{\link{slice}()},
\code{\link{squeeze}()},
\code{\link{transpose}()},
\code{\link{zeros}()}
}
\concept{Array}
