% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepArray.r
\name{vander}
\alias{vander}
\title{Multidimensional array creation: Vandermonde matrix}
\usage{
vander(data, n)
}
\arguments{
\item{data}{The data to be reshaped to a Vandermonde matrix.}

\item{n}{The number of columns of the resulting matrix.}
}
\value{
A Vandermonde matrix.
}
\description{
Multidimensional array creation: Vandermonde matrix
}
\details{
The Python package NumPy offers an analog array creation function \code{np.vander(x, n)}.
}
\seealso{
Other Array: 
\code{\link{column_stack}()},
\code{\link{dimC<-}()},
\code{\link{eye}()},
\code{\link{flatten}()},
\code{\link{insert}()},
\code{\link{mabind}()},
\code{\link{mamatrix}()},
\code{\link{marray}()},
\code{\link{ndim}()},
\code{\link{nsize}()},
\code{\link{reshape.array}()},
\code{\link{row_stack}()},
\code{\link{slice}()},
\code{\link{squeeze}()},
\code{\link{transpose}()}
}
\concept{Array}
