% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepQuality.r
\name{vc}
\alias{vc}
\title{Variance coefficient}
\usage{
vc(actual, predicted)
}
\arguments{
\item{actual}{A vector with to-be values.}

\item{predicted}{A vector with as-is values.}
}
\value{
Variance coefficient (VC).
}
\description{
\code{vc} calculates the variance coefficient.
}
\seealso{
\code{\link{mae}} for mean absolute error, \code{\link{mape}} for mean absolute percentage error,
  \code{\link{mse}} for mean squared error, \code{\link{rmse}} for root mean square error.

Other Quality: 
\code{\link{mae}()},
\code{\link{mape}()},
\code{\link{rmse}()}
}
\concept{Quality}
