% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepQuality.r
\name{vc}
\alias{vc}
\title{Variance coefficient}
\usage{
vc(y, yhat)
}
\arguments{
\item{y}{A numeric vector with actual values (to-be values).}

\item{yhat}{A numeric vector with estimated values (as-is values).}
}
\value{
Variance coefficient (VC).
}
\description{
Variance coefficient
}
\seealso{
Other Quality: 
\code{\link{accuracy}()},
\code{\link{huber}()},
\code{\link{log_cosh}()},
\code{\link{mae}()},
\code{\link{mape}()},
\code{\link{mse}()},
\code{\link{quantile_loss}()},
\code{\link{rmse}()}
}
\concept{Quality}
