% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepMLP.r
\name{vector.as.numeric}
\alias{vector.as.numeric}
\title{Transform a vector to a numeric vector}
\usage{
vector.as.numeric(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
The vector \code{x} as numeric vector.
}
\description{
Transform a vector to a numeric vector
}
\seealso{
Other Single & Multi Layer Perceptron (SLP, MLP): 
\code{\link{as.ANN.matrix}()},
\code{\link{as.MLP.X}()},
\code{\link{as.MLP.Y}()},
\code{\link{as.tensor.1D}()},
\code{\link{as.tensor.2D}()},
\code{\link{as.tensor.3D}()},
\code{\link{as.tensor}()},
\code{\link{build.MLP}()},
\code{\link{fit.MLP}()},
\code{\link{get.MLP.X.samples}()},
\code{\link{get.MLP.X.units}()},
\code{\link{get.MLP.Y.samples}()},
\code{\link{get.MLP.Y.units}()},
\code{\link{predict.ANN}()},
\code{\link{vector.as.ANN.matrix}()}
}
\concept{Single & Multi Layer Perceptron (SLP, MLP)}
