% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepUtils.r
\name{vector_as_ANN_matrix}
\alias{vector_as_ANN_matrix}
\title{Transform a vector into a ANN compatible matrix}
\usage{
vector_as_ANN_matrix(
  x,
  ncol = 1,
  by = c("row", "col", "step"),
  reverse = FALSE
)
}
\arguments{
\item{x}{A numeric vector.}

\item{ncol}{The number of columns in the resulting matrix. If \code{by = step}, the number of columns is equal to the number of timesteps used for a LSTM.}

\item{by}{Controls the transformation process. The options \code{row} and \code{col} lead to a matrix whereby the values are structured row-wise or column-wise.
The option \code{step} stands for a stepwise order of the values row by row (e.g. 1 2 3, 2 3 4, 4 5 6 etc.).}

\item{reverse}{Controls the order of the values in the transformed vector \code{x}. By default, they are used in the given order, but they can also be used in reverse order.}
}
\value{
The transformed or resampled vector \code{x} into a matrix.
}
\description{

}
\seealso{
\code{\link{as_tensor_3D}}.

Other Utils: 
\code{\link{as_ANN_matrix}()},
\code{\link{as_tensor_1D}()},
\code{\link{as_tensor_2D}()},
\code{\link{as_tensor_3D}()},
\code{\link{degree}()},
\code{\link{distance}()},
\code{\link{flatten}()},
\code{\link{marray}()},
\code{\link{nsamples}()},
\code{\link{nsubsequences}()},
\code{\link{ntimesteps}()},
\code{\link{nunits}()},
\code{\link{probability}()},
\code{\link{radian}()},
\code{\link{re.factor}()},
\code{\link{sd_pop}()},
\code{\link{similarity}()},
\code{\link{tensor}()},
\code{\link{var_pop}()},
\code{\link{vector_as_numeric}()}
}
\concept{Utils}
