% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepOutlier.r
\name{winsorize}
\alias{winsorize}
\title{Winsorize outliers}
\usage{
winsorize(
  x,
  minx = NULL,
  maxx = NULL,
  probs = c(0.05, 0.95),
  na.rm = FALSE,
  type = 7
)
}
\arguments{
\item{x}{A numeric vector to be winsorized.}

\item{minx}{The low border value, all values within \code{x} being lower than this value will be replaced by this value. By default, this value is set to the 5\\% quantile of \code{x}.}

\item{maxx}{The high border value, all values within \code{x} being higher than this value will be replaced by this value. By default, this value is set to the 95\\% quantile of \code{x}.}

\item{probs}{A numeric vector of probabilities with values in \link{0,1} as used in \code{\link[stats]{quantile}}.}

\item{na.rm}{A logical value indicating whether missing values should be omitted or not (default) to calculate the quantiles.}

\item{type}{An integer between 1 and 9 selecting one of the nine quantile algorithms detailed in \code{\link[stats]{quantile}} to be used.}
}
\value{
A vector of the same length as \code{x} containing the winsorized values.
}
\description{
\code{winsorize} sets outliers to low and high border values.
}
\seealso{
\code{\link[stats]{quantile}}, \code{\link{outlier}}.

Other Outlier: 
\code{\link{outlier_dataset}()},
\code{\link{outlier}()}
}
\concept{Outlier}
